### 2024-05-14 #6751 @editorialbot
[REVIEW]: LiGuard: A GUI-powered Python Framework for Processing Point-Cloud Data
**Submitting author:** @m-shahbaz-kharal (<a href="http://orcid.org/0009-0003-6377-8274">Muhammad  Shahbaz </a>)
**Repository:** https://github.com/m-shahbaz-kharal/LiGuard-2.x/
**Branch with paper.md** (empty if default branch): dev_2.x
**Version:** v2.1.5.post3
