### 2024-06-19 #6751 @hugoledoux
@chenzhaiyu @tgoelles just a small reminder not to forget this review
