### 2024-06-19 #6751 @tgoelles
Mine can be found here: https://github.com/m-shahbaz-kharal/LiGuard-JOSS/issues/1
I can not checkoff all the points. See my reasoning in my review there.
