### 2024-06-24 #6751 @chenzhaiyu
The authors present a Python GUI for processing LiDAR data, optionally accompanying imagery. They claim the contribution user-friendly with five sub-modules for IO, configuration, inter-process data sharing, data processing, and visualization.

## General Concerns
1. I am not sure if the magnitude of _scholarly_ contribution is above the bar. This concerns both the commit history and lines of code. In my opinion the code appears to be largely boilerplate on top of existing libraries.
2.  Although the paper does reference `Open3D` and `OpenCV` (line 44). It would be more appropriate to state clearly the diference from existing libs (not limited to the two), highlighting the authors' unique contributions rather than making broad claims.

## Paper
3. The authors emphasize multiple times _GUI_ to contrast their contribution from other software. I support @tgoelles's point that it is important to discuss and conceptually compare other GUI-based software to provide context.
4. I find the claim of _extensible_ (line 28) not sufficiently substantiated as much as the other ones.
5. Currently, the paper includes only one conceptual figure. Given that the GUI is the primary focus of the software presentation, incorporating visual elements would probably help.

## Documentation

6. The README sections could be made more user-friendly by including self-contained examples. As it stands, one must read through the lengthy documentation to get started with the software. A concise example at the beginning would help. The current [example usage](https://github.com/m-shahbaz-kharal/LiGuard-JOSS/tree/main#an-example-pipeline) is too conceptual and less ideal.
7. Documentation for modules and functions is incomplete and almost broken in formatting. E.g., https://m-shahbaz-kharal.github.io/LiGuard-JOSS/algo.html.
8. While automated tests seem to be present, their usage is undocumented.
9. If example data can be provided, users would be able to quickly test the software. Without such data I could only follow the README until [opening-the-pipeline-config-file](https://github.com/m-shahbaz-kharal/LiGuard-JOSS/tree/main#opening-the-pipeline-config-file), and thus could not confirm the functional claims beyond that point.
