### 2024-06-24 #6751 @m-shahbaz-kharal
Hi, I believe all the reviews are submitted as the assigned reviewers (@chenzhaiyu and @tgoelles) have submitted their responses; please inform me if that is not correct. In the following the response to the reviews is submitted.

Dear Reviewers,

Thanks for a thorough review. The submitted comments/recommendations/decisions helped us a lot in understanding the process of publication in JOSS more clearly, and more importantly, it has helped us see the potential issues in both the paper and the code/repository. The issues discussed in the reviews primarily included incomprehensive statement-of-need, missing references and comparisons to some similar libraries, broader (than evident) claims, improper tests and documentation, and a lack of enough commit history. Based on the comment from @hugoledoux (["We aim at bringing the submission to a certain standard that we deem accepted, in this case when all the checklists can be checked."](https://github.com/openjournals/joss-reviews/issues/6751#issuecomment-2178896294)), it seems like the submission can be improved and eventually published if it checks all the requirements; please inform us if that is not the case for this submission.

While it may take us sometime before all the raised points/concerns are covered/fixed, some quick comments are provided in the following to answer some of those questions.

- **Commit History**: This repository is a trimmed-down version of our original project [LiGuard-GUI](https://github.com/m-shahbaz-kharal/LiGuard-GUI) that is kept private as it contains some under-research algorithms that are (implemented using `LiGuard` but) are not published yet. Since it is not confirmed, as of the moment, if and when those algorithms are going to be made available, we created this repo and added only the files required to run the software. However, since the commit history is required for the JOSS review process, the reviewers (@hugoledoux, @chenzhaiyu, @tgoelles) are being added as collaborators to that repo so the commit history can be reviewed.
- **Tests**: The test-cases written for this repo are only meant to test if the signatures and return types of custom functions are being followed correctly and not necessarily the definition/logic of those functions. Therefore, the software may still crash, even if the tests doesn't show an error; in that case, the logs may be checked to figure out the errors.
-  **Comparison to `ROS`**: The comparison to `ROS` is deliberately not added for this project as it seems to us that it is not an apple-to-apple comparison. `ROS` is a meta-operating system whereas `LiGuard` is focused only on processing point-cloud (and accompanying image) data. Therefore, a reference to it will be added for completeness purpose but since `ROS` and `LiGuard` serve a different purpose, a detailed comparison seems unnecessary.
- **Supported File-Types**: Please note that, for lidar, supported file-types include `.bin`, `npy`, `.ply`, and `.pcd` (most of these are supported by `Open3D` so those are supported by extension) and for images `imread` function of `OpenCV` is used so all supported image file-types supported by `OpenCV` are also supported in our software (by extension).
- **Segmentation Fault**: The software is tested only `Windows 11` using `Python 3.10.0` and although our docs say that it should run as far as the dependencies are installed but as @tgoelles pointed out that's not the case. So, for now, the only environment it should be tested on is `Windows 10` with `Python 3.10.0` (we'll reflect this statement in the docs). Also, we'll update docs as more operating systems are tested.

# Summary of Future Updates

1. **Updates in Paper**: Based on the reviews, a major re-writing of the paper is needed. So, although the title and basics of the submission may remain similar, an improved version of the paper.md will be submitted that'll include a comprehensive statement-of-need, missed references, revised claims, novelty statement(s), and comparison to related SOTA packages.

2. **Updates in Repository**: Following are the major planned updates to the repo:
    - **Quick-Start Guide**: A few example datasets and yaml configuration-files will be added to allow quick trying of various features of `LiGuard`.
    - **Documentation Breakdown and Improvements**: The documentation will be broken down to multiple sub-docs to allow easy navigation.
    - **Installation Package**: The software will be bundled as python package to allow easier installation.

**Note**: The updates will be posted and informed in this thread in the same order as given in **Summary of Future Updates** above. Please let us know if anything is still missing/unaccounted from our side regarding the review process.

Thanks.
