### 2024-07-17 #6751 @hugoledoux
Sorry for taking so long to getting back to this. A few answers and comments:

- yes that's the idea that you get the chance to improve your submission so that eventually when the reviewers and myself think it fulfils the JOSS criteria, it can be accepted.
- my invite for your private repo expired, sorry. Could you reinvite me?
- however, why not keeping the not-published algos in a develop branch and have the main part in the main branch? I would find that simpler to be honest, also for you.
- Comparison to ROS: perhaps a full comparison is not necessary, but since a knowledgeable reviewer thought that there is a link you could explain why it is different, and how? And what are the similarities?
- it seems both reviewers couldn't run your software, this is worrying for JOSS. You need to dive into this and try your software with other OSes than Windows
- Comparison to PDAL: will you add this?
- what is missing is simple demos to run, with data (as @chenzhaiyu points out)

So I suggest you improve the software, the docs, and the paper. And then you let us know here what changes you've made, and then we can continue the review.

How long do you reckon it will take you?
