### 2024-07-24 27890d5
feat: add rotation functionality to lidar.py
