### 2024-07-24 #6751 @m-shahbaz-kharal
Hi,

Thanks @hugoledoux  for clarifications.

* I am sending invites again.
* The reason behind not keeping private algos in private develop branch is that those algorithms though developed using `LiGuard`'s framework, are still private to our research and since that research is not published yet, we can't risk making it public (develop branch is still public and I couldn't find a way to keep certain branches private in a public github repo, I believe it is not possible at least with current version of github.com).
* Regarding ROS: We'll be adding a section in the paper to compare common aspects of ROS and our framework.
* Software Runtime Errors: Those are believed to be related to `open3d.visualization.gui` sub-package of `Open3D`. We are fixing those for at least one more OS and that'll be Ubuntu 22.04 LTS. We'll also create a `pip` package to ease the installation.
* Comparison to PDAL: Yes, we'll add the missing references and comparison to other related software/libraries (including PDAL).
* The demo in written text seems longer than it actually is because of explanations inside. We'll add a quick-start demo to cover this concern.
* It may take us up to 2 more months to get all things fixed. If it is okay I'll notify here upon every update/fix is pushed to keep you in the loop, otherwise I'll notify here once all the raised concerns are fixed.

Thanks.
