### 2024-08-01 #6751 @hugoledoux
okay, I'll pause this review while you update the code.

As for the private/public repository, I'll ask my colleagues what they think of this. But one thing you could do is have the real repository (without JOSS in the name) as the main one, and another one private where you pull changes from the main one, and when you are ready to put the code public (I guess this is the idea that this will happen, right?) you merge with the main repository.
