### 2024-08-01 #6751 @m-shahbaz-kharal
Thank you for the update. I will inform everyone here once all the prospective changes are made so that reviewers can go through all parts at once.If I set the main repository to public, it will make all the code public, correct? Even the part that is supposed to be published later; forks also have histories. I will search to see if there is a way I can remove a specific piece of code and its history somehow. I would greatly appreciate it if someone could assist me with this.On Aug 1, 2024 1:39 AM, Hugo Ledoux ***@***.***> wrote:
okay, I'll pause this review while you update the code.
As for the private/public repository, I'll ask my colleagues what they think of this. But one thing you could do is have the real repository (without JOSS in the name) as the main one, and another one private where you pull changes from the main one, and when you are ready to put the code public (I guess this is the idea that this will happen, right?) you merge with the main repository.
