### 2024-08-01 #6751 @hugoledoux
yes it's possible to remove commits from the history, eg : https://graphite.dev/guides/git-delete-commit-from-history
