### 2024-08-07 0195973
feat: remove NaN values from point cloud in pre-processing step
