### 2024-08-08 3d8e222
feat: add keyboard shortcut 0 to reset frame index
