### 2024-08-09 329ab0b
feat: Add focus check for key event handling
