+++ img/viz.py
@@ -4,6 +4,8 @@ import numpy as np
 
 from gui.logger_gui import Logger
 
+import pygetwindow as gw
+
 class ImageVisualizer:
     """
     A class for visualizing images and bounding boxes using Open3D.
@@ -32,6 +34,17 @@ class ImageVisualizer:
         self.viz.create_window("Image Feed", width=int(1440/4), height=int(1080/4), left=480 - int(1440/4), top=30)
         # init
         self.reset(cfg, True)
+
+    def __is_focused__(self):
+        """
+        Checks if the visualizer is focused.
+
+        Returns:
+            bool: True if the visualizer is focused, False otherwise.
+        """
+        window = gw.getWindowsWithTitle('Image Feed')
+        if window: return window[0].isActive
+        return False
         
     def reset(self, cfg, reset_bounding_box=False):
         """
