+++ lbl/file_io.py
@@ -19,6 +19,7 @@ class FileIO:
         cfg (dict): Configuration dictionary.
         lbl_dir (str): Directory path for label files.
         lbl_type (str): Type of label files.
+        lbl_start_idx (int): Index of the first label file to process.
         lbl_count (int): Number of label files to process.
         lbl_ext (str): Extension of label files.
         reader (class): Handler class for reading label files.
@@ -37,117 +38,11 @@ class FileIO:
 
     """
     def __init__(self, cfg: dict, calib_reader: callable):
-        # Initialize the configuration dictionary
-        self.cfg = cfg
-        # Set the directory path for label files
-        self.lbl_dir = os.path.join(cfg['data']['path'], cfg['data']['label_subdir'])
-        # Set the type of label files
-        self.lbl_type = cfg['data']['label']['lbl_type']
-        # Set the number of label files to process
-        self.lbl_count = cfg['data']['size']
-        
-        # Check if the label type is supported
-        if self.lbl_type not in supported_label_types: raise NotImplementedError("Label type not supported. Supported file types: " + ', '.join(supported_label_types) + ".")
-        # Import the handler for the label type
-        h = __import__('lbl.handler_'+self.lbl_type, fromlist=['label_file_extension', 'Handler'])
-        # Set the extension and reader for the label files
-        self.lbl_ext, self.reader = h.label_file_extension, h.Handler
-        # Set the callable object for reading calibration data
-        self.clb_reader = calib_reader
-        
-        # Get all label files in the directory
-        files = glob.glob(os.path.join(self.lbl_dir, '*' + self.lbl_ext))
-        # Get the basenames of the label files and sort them
-        file_basenames = [os.path.splitext(os.path.basename(file))[0] for file in files]
-        file_basenames.sort(key=lambda file_name: int(''.join(filter(str.isdigit, file_name))))
-        # Set the list of file basenames
-        self.files_basenames = file_basenames[:self.lbl_count]
-        
-        # Initialize a lock for thread safety
-        self.data_lock = threading.Lock()
-        # Initialize a list for storing label file paths and annotations
-        self.data = []
-        # Initialize an event for stopping the async read thread
-        self.stop = threading.Event()
-        # Start the async read thread
-        threading.Thread(target=self.__async_read_fn__).start()
-        
-    def get_abs_path(self, idx: int) -> str:
-        """
-        Returns the absolute path of the label file at the given index.
-
-        Args:
-            idx (int): Index of the label file.
-
-        Returns:
-            str: Absolute path of the label file.
-
-        """
-        # Get the absolute path of the label file
-        lbl_path = os.path.join(self.lbl_dir, self.files_basenames[idx] + self.lbl_ext)
-        return lbl_path
-        
-    def __async_read_fn__(self):
-        """
-        Asynchronously reads label files and annotations.
-
-        """
-        # Loop through all label files
-        for idx in range(len(self.files_basenames)):
-            # Break the loop if the stop event is set
-            if self.stop.is_set(): break
-            # Get the absolute path of the label file
-            lbl_abs_path = self.get_abs_path(idx)
-            # Read the annotation of the label file
-            annotation = self.reader(lbl_abs_path, self.clb_reader(idx)[1] if self.clb_reader else None)
-            # Append the label file path and annotation to the data list
-            with self.data_lock: self.data.append((lbl_abs_path, annotation))
-            # Sleep for a while
-            time.sleep(self.cfg['threads']['io_sleep'])
-        
-    def __len__(self) -> int:
-        """
-        Returns the number of label files.
-
-        Returns:
-            int: Number of label files.
-
-        """
[…]
