### 2024-08-16 d3fa03d
feat: added manual lidar-camera calibration
