### 2024-08-16 fad1bfa
feat: implemented ultralytics yolov5 object detection algorithm
