+++ configs/config_template.yml
@@ -143,6 +143,19 @@ proc: # liguard processing configurations
         project_point_cloud_points: # project point cloud points to camera image
             enabled: False # set True to project point cloud points to camera image
             priority: 1 # priority of process - lower is higher
+        ultralytics_yolov5:
+            enabled: False
+            activate_on_key_set: 'current_image_numpy'
+            priority: 2
+            model: 'yolov5s' # can be yolov5s, yolov5m, yolov5l, yolov5x https://pytorch.org/hub/ultralytics_yolov5
+            class_colors: # in RGB format 
+                person: [1, 0, 0]
+                bicycle: [0, 1, 0]
+                car: [0, 0, 1]
+                motorcycle: [0, 1, 0]
+                bus: [0, 1, 1]
+                truck: [1, 0, 1]
+            score_threshold: 0.5
     calib:
         dummy: # dummy calibration process
             enabled: False # set True to enable
