### 2024-08-16 bb1490e
refactor: improve point cloud pre-processing function
