### 2024-08-16 d5fceb0
feat: added support for loading and saving Discrete Histograms of Distances-Per-Point background filter
