### 2024-08-16 3693750
feat: added support for loading and saving Spatio-Temporal Density Filter background filter. homogenized: added filter_set keys to data_dict (just house keeping).
