### 2024-08-16 0ba8fa5
homogenized: set BGFilterSTDF flag in data_dict based on filter loading status (just house keeping).
