### 2024-08-19 349e824
feat: added gen_bbox2d function to generate 2D bounding boxes from 3D bounding boxes.
