+++ configs/config_template.yml
@@ -149,6 +149,9 @@ proc: # liguard processing configurations
             priority: 7
             path_to_github_repo: 'algo/nn/PointPillars' # clone https://github.com/zhulf0804/PointPillars to this path and install the requirements
             score_threshold: 0.5
+        gen_bbox2d:
+            enabled: False
+            priority: 8
     camera:
         project_point_cloud_points: # project point cloud points to camera image
             enabled: False # set True to project point cloud points to camera image
