### 2024-08-19 4fb3013
fix: bbox_2d should be and are being drawn now even if the corresponding bbox_3d is already drawn.
