### 2024-08-19 376cd5f
refactor: rename gen_bbox2d function to gen_bbox_2d for consistency
