### 2024-08-19 9841703
refactor: removed the text ['CRITICAL'] in logger print statement (it is evident by logger meta).
