+++ algo/calib.py
@@ -5,7 +5,7 @@ from gui.logger_gui import Logger
 def dummy(data_dict: dict, cfg_dict: dict):
     # get logger object from data_dict
     if 'logger' in data_dict: logger:Logger = data_dict['logger']
-    else: print('[algo->calib.py->dummy][CRITICAL]: No logger object in data_dict. It is abnormal behavior as logger object is created by default. Please check if some script is removing the logger key in data_dict.'); return
+    else: print('[algo->calib.py->dummy]: No logger object in data_dict. It is abnormal behavior as logger object is created by default. Please check if some script is removing the logger key in data_dict.'); return
 
     # do stuff
     # ...
