### 2024-08-19 5116998
refactor: improved consistency in gen_bbox_2d function.
