### 2024-08-20 66ba64c
refactor: capitalize object class names in ultralytics_yolov5 function for consistency.
