### 2024-08-20 d37d490
fix: don't create bbox_2d if one is already present.
