### 2024-08-20 bbea786
refactor: renamed ultralytics_yolov5 function to UltralyticsYOLOv5 for consistency.
