+++ algo/camera.py
@@ -75,7 +75,7 @@ def project_point_cloud_points(data_dict: dict, cfg_dict: dict):
     # Update the image with the projected lidar points
     data_dict['current_image_numpy'][pixel_coords_valid[:, 1], pixel_coords_valid[:, 0]] = np.column_stack((pixel_depths_valid, np.zeros_like(pixel_depths_valid), np.zeros_like(pixel_depths_valid)))
 
-def ultralytics_yolov5(data_dict: dict, cfg_dict: dict):
+def UltralyticsYOLOv5(data_dict: dict, cfg_dict: dict):
     """
     Runs the Ultralytics YOLOv5 object detection algorithm on the current image.
 
@@ -87,26 +87,26 @@ def ultralytics_yolov5(data_dict: dict, cfg_dict: dict):
         None
     """
     if 'logger' in data_dict: logger:Logger = data_dict['logger']
-    else: print('[algo->camera.py->ultralytics_yolov5]: No logger object in data_dict. It is abnormal behavior as logger object is created by default. Please check if some script is removing the logger key in data_dict.'); return
+    else: print('[algo->camera.py->UltralyticsYOLOv5]: No logger object in data_dict. It is abnormal behavior as logger object is created by default. Please check if some script is removing the logger key in data_dict.'); return
     
     if 'current_image_numpy' not in data_dict:
-        logger.log('[algo->camera.py->ultralytics_yolov5]: current_image_numpy not found in data_dict', Logger.ERROR)
+        logger.log('[algo->camera.py->UltralyticsYOLOv5]: current_image_numpy not found in data_dict', Logger.ERROR)
         return
     
     # imports
     import torch
 
     # algo name and keys used in algo
-    algo_name = 'ultralytics_yolov5'
+    algo_name = 'UltralyticsYOLOv5'
     model_key = f'{algo_name}_model'
     tgt_cls_key = f'{algo_name}_target_classes'
     
     # get params
-    params = cfg_dict['proc']['camera']['ultralytics_yolov5']
+    params = cfg_dict['proc']['camera']['UltralyticsYOLOv5']
 
     # check if model is already loaded
     if model_key not in data_dict:
-        logger.log(f'[algo->camera.py->ultralytics_yolov5]: Loading model', Logger.INFO)
+        logger.log(f'[algo->camera.py->UltralyticsYOLOv5]: Loading model', Logger.INFO)
         data_dict[model_key] = torch.hub.load('ultralytics/yolov5', params['model'], pretrained=True, _verbose=False)
         vk_dict = {v.capitalize():k for (k,v) in data_dict[model_key].names.items()}
         data_dict[tgt_cls_key] = [vk_dict[key] for key in params['class_colors']]
