### 2024-08-20 db013cf
refactor: added 'added_by' into the label_dict for bbox_2d and bbox_3d label.
