### 2024-08-20 e7f0e5b
feat: added information fusion to enhance labels from a source bbox_2d to target bbox_2d and bbox_3d.
