+++ configs/config_template.yml
@@ -182,6 +182,11 @@ proc: # liguard processing configurations
             priority: 2 # priority of process - lower is higher
             min_points: 6 # minimum number of points to consider a label valid
     post:
+        FusePredictedBBoxesFromSourceToTarget: # enhance 3D bounding boxes using 2D bounding boxes
+            enabled: False # set True to enhance 3D bounding boxes
+            priority: 1 # priority of process - lower is higher
+            bbox_target_algo: 'Cluster2Object_2d' # algorithm that generated 3D bounding boxes
+            bbox_source_algo: 'UltralyticsYOLOv5' # algorithm that generated 2D bounding boxes
         create_per_object_pcdet_dataset: # create per object dataset in pcdet format
             enabled: False # set True to enable
             priority: 1 # priority of process - lower is higher
