### 2024-08-20 33b06eb
refactor: change log level to DEBUG for cluster with undetermined class.
