### 2024-08-20 6711b79
improve: remove_out_of_bound_labels can now use image_size to remove out of bound bbox_2d.
