+++ algo/post.py
@@ -14,7 +14,7 @@ def FusePredictedBBoxesFromSourceToTarget(data_dict: dict, cfg_dict: dict):
     Operation:
     - It uses KD-Tree to find the nearest bbox_2d.
     - It assigns the class of bbox_2d to bbox_3d.
-    - TODO: figure out a method to enhance respective bbox_3d.
+    - TODO: figure out more methods to enhance bbox fusion.
 
     Args:
         data_dict (dict): A dictionary containing the required data.
@@ -31,6 +31,8 @@ def FusePredictedBBoxesFromSourceToTarget(data_dict: dict, cfg_dict: dict):
     if 'current_label_list' not in data_dict:
         logger.log('[algo->post.py->FusePredictedBBoxesFromSourceToTarget]: current_label_list not found in data_dict', Logger.ERROR)
         return
+    if 'current_calib_data' not in data_dict:
+        logger.log('[algo->post.py->FusePredictedBBoxesFromSourceToTarget]: current_calib_data not found in data_dict. Ignoring bbox_3d <--> bbox_2d fusion.', Logger.WARNING)
     
     # imports
     import numpy as np
@@ -39,6 +41,8 @@ def FusePredictedBBoxesFromSourceToTarget(data_dict: dict, cfg_dict: dict):
 
     # algo name and dict keys
     algo_name = 'FusePredictedBBoxesFromSourceToTarget'
+    tr_pcd_to_img_key = f'{algo_name}_tr_pcd_to_img'
+    inv_tr_pcd_to_img_key = f'{algo_name}_inv_tr_pcd_to_img'
 
     # get params
     params = cfg_dict['proc']['post'][algo_name]
@@ -60,8 +64,11 @@ def FusePredictedBBoxesFromSourceToTarget(data_dict: dict, cfg_dict: dict):
             logger.log(f'[algo->post.py->FusePredictedBBoxesFromSourceToTarget]: Unknown bbox source: {bbox_2d_dict["added_by"]}', Logger.WARNING)
             continue
 
-    if len(target_points_list) == 0 or len(source_points_list) == 0:
-        logger.log(f'[algo->post.py->FusePredictedBBoxesFromSourceToTarget]: No bbox_2d found for {params["bbox_target_algo"]} or {params["bbox_source_algo"]}', Logger.WARNING)
+    if len(target_points_list) == 0:
+        logger.log(f'[algo->post.py->FusePredictedBBoxesFromSourceToTarget]: No target bbox_2d found', Logger.DEBUG)
+        return
+    if len(source_points_list) == 0:
+        logger.log(f'[algo->post.py->FusePredictedBBoxesFromSourceToTarget]: No source bbox_2d found', Logger.DEBUG)
         return
 
     # Create a KDTree for points_list_2
@@ -81,13 +88,34 @@ def FusePredictedBBoxesFromSourceToTarget(data_dict: dict, cfg_dict: dict):
     matches_kd_tree = list(zip(row_indices, col_indices))
     # ---------------------- KDTree Matching ---------------------- #
 
+    if tr_pcd_to_img_key not in data_dict and 'current_calib_data' in data_dict:
+        Tr_velo_to_cam = data_dict['current_calib_data']['Tr_velo_to_cam']
+        R0_rect = data_dict['current_calib_data']['R0_rect']
+        P2 = data_dict['current_calib_data']['P2']
+        P2_4x4 = np.eye(4)
+        P2_4x4[:3, :] = P2
+        data_dict[tr_pcd_to_img_key] = P2_4x4 @ R0_rect @ Tr_velo_to_cam
[…]
