+++ configs/config_template.yml
@@ -189,6 +189,7 @@ proc: # liguard processing configurations
             priority: 1 # priority of process - lower is higher
             bbox_target_algo: 'Cluster2Object_2d' # algorithm that generated 3D bounding boxes
             bbox_source_algo: 'UltralyticsYOLOv5' # algorithm that generated 2D bounding boxes
+            max_match_distance: 100 # maximum euclidean distance for matching source and target bounding boxes
         create_per_object_pcdet_dataset: # create per object dataset in pcdet format
             enabled: False # set True to enable
             priority: 1 # priority of process - lower is higher
