### 2024-08-21 f5d9fc3
fix: bbox_3d fusion happens if bbox_3d is in target label dict.
