### 2024-08-21 f605028
feat: Generate future trajectory using cubic spline interpolation.
