### 2024-08-21 f965206
refactor: used numpy arrays for past_trajectory and future_trajectory in GenerateKDTreePastTrajectory and GenerateCubicSplineFutureTrajectory, and added per-dimension CubicSpline in GenerateCubicSplineFutureTrajectory.
