### 2024-08-21 283b378
feat: added trajectory visualization to PointCloudVisualizer.
