### 2024-08-21 690fe04
feat: added GeneratePolyFitFutureTrajectory function for polynomial fit future trajectory prediction.
