### 2024-08-21 00af401
refactor: update GenerateCubicSplineFutureTrajectory to use minimum past trajectory length from params
