### 2024-08-21 d21d1f6
feat: added trajectories to camera visualization.
