### 2024-08-21 f0b1501
feat: added GenerateVelocityFromTrajectory.
