### 2024-08-21 6c2aea5
refactor: enable/disable visualization options in config_template.yml feat: added text info on bbox_2d in image visualizer. homogenize: moved bbox_3d ops under __add_bbox__ function of visualizers.
