+++ img/viz.py
@@ -121,7 +121,6 @@ class ImageVisualizer:
 
         for lbl in data_dict['current_label_list']:
             self.__add_bbox__(lbl, clb)
-            self.__add_trajectories__(lbl, clb)
         
     def __add_bbox__(self, label_dict, calib_dict):
         """
@@ -131,7 +130,7 @@ class ImageVisualizer:
             label_dict (dict): A dictionary containing the label information.
             calib_dict (dict): A dictionary containing the calibration information.
         """
-        if 'bbox_3d' in label_dict and calib_dict != None:
+        if 'bbox_3d' in label_dict and calib_dict != None and self.cfg['visualization']['camera']['draw_bbox_3d']:
             # bbox parameters
             bbox_3d_dict = label_dict['bbox_3d']
             xyz_center = bbox_3d_dict['xyz_center']
@@ -201,12 +200,44 @@ class ImageVisualizer:
             cv2.line(img_np,tuple(pts[1]),tuple(pts[5]),rgb_color,self.cfg['visualization']['camera']['bbox_line_width'])
             cv2.line(img_np,tuple(pts[2]),tuple(pts[6]),rgb_color,self.cfg['visualization']['camera']['bbox_line_width'])
             cv2.line(img_np,tuple(pts[3]),tuple(pts[7]),rgb_color,self.cfg['visualization']['camera']['bbox_line_width'])
+
+            # add past bbox_3d trajectories if available
+            if 'past_trajectory' in bbox_3d_dict and self.cfg['visualization']['camera']['draw_trajectory']:
+                past_trajectory = bbox_3d_dict['past_trajectory']
+                past_trajectory = np.concatenate([past_trajectory, np.ones((past_trajectory.shape[0], 1))], axis=1)
+                past_trajectory = calib_dict['P2'] @ calib_dict['R0_rect'] @ calib_dict['Tr_velo_to_cam'] @ past_trajectory.T
+                past_trajectory = past_trajectory[:, past_trajectory[2] > 0]
+                past_trajectory = past_trajectory[:2] / past_trajectory[2]
+                past_trajectory = past_trajectory.T
+                past_trajectory = past_trajectory.astype(np.int32)
+                
+                for i in range(len(past_trajectory) - 1): cv2.line(img_np,
+                                                                   tuple(past_trajectory[i]),
+                                                                   tuple(past_trajectory[i + 1]),
+                                                                   (bbox_3d_dict['rgb_color'] * 127).tolist(),
+                                                                   self.cfg['visualization']['camera']['trajectory_line_width'])
             
+            # add future bbox_3d trajectories if available
+            if 'future_trajectory' in bbox_3d_dict and self.cfg['visualization']['camera']['draw_trajectory']:
+                future_trajectory = bbox_3d_dict['future_trajectory']
+                future_trajectory = np.concatenate([future_trajectory, np.ones((future_trajectory.shape[0], 1))], axis=1)
+                future_trajectory = calib_dict['P2'] @ calib_dict['R0_rect'] @ calib_dict['Tr_velo_to_cam'] @ future_trajectory.T
+                future_trajectory = future_trajectory[:, future_trajectory[2] > 0]
+                future_trajectory = future_trajectory[:2] / future_trajectory[2]
+                future_trajectory = future_trajectory.T
+                future_trajectory = future_trajectory.astype(np.int32)
+                
+                for i in range(len(future_trajectory) - 1): cv2.line(img_np,
+                                                                     tuple(future_trajectory[i]),
+                                                                     tuple(future_trajectory[i + 1]),
+                                                                     (bbox_3d_dict['rgb_color'] * 255).tolist(),
+                                                                     self.cfg['visualization']['camera']['trajectory_line_width'])
+
             # add to visualizer
[…]
