### 2024-08-22 22d8525
refactor: made changes to keep the config_file_path_textedit and button_container always on top (previously it was scrollable which makes button bar hard to use).
