### 2024-08-22 f546539
fix: if draw_bbox_3d under visualization->camera is true then text_info is shown on bbox_3d->2d if bbox_3d is available, otherwise it is shown on bbox_2d.
