### 2024-08-22 3d6d5b8
refactor: keep visualizers even if visualization is disabled in config to not run into errors. TODO: make the processing loop independent of visualizers completely.
