### 2024-08-22 6240495
multi: added visualize flag in gen_bbox_2d, added support for extras for img_visualizer, improved text_info in GenerateKDTreePastTrajectory, and some minor refactors.
