### 2024-08-23 4c32854
feat: added a simple profiler for time profiling calls, added it to main loop of LiGuard. feat: added support for saving current views of visualizers. refactor: changed the key to seek to starting frame index.
