+++ main.py
@@ -3,6 +3,7 @@ import open3d.visualization.gui as gui
 
 from gui.config_gui import BaseConfiguration as BaseConfigurationGUI
 from gui.logger_gui import Logger
+from profiler import Profiler
 
 from pcd.file_io import FileIO as PCD_File_IO
 from pcd.sensor_io import SensorIO as PCD_Sensor_IO
@@ -16,6 +17,8 @@ from calib.file_io import FileIO as CLB_File_IO
 from lbl.file_io import FileIO as LBL_File_IO
 
 import keyboard, threading, time
+
+profiler = Profiler('main')
 
 class LiGuard:
     def __init__(self):
@@ -68,7 +71,7 @@ class LiGuard:
         
     # handle the key events of right, left, and space keys
     def handle_key_event(self, event:keyboard.KeyboardEvent):
-        if event.event_type == keyboard.KEY_DOWN and self.__is_focused__():
+        if event.event_type == keyboard.KEY_DOWN:
             with self.lock:
                 if event.name == 'right':
                     self.is_playing = False
@@ -80,7 +83,7 @@ class LiGuard:
                         self.data_dict['current_frame_index'] -= 1
                 elif event.name == 'space':
                     self.is_playing = not self.is_playing
-                elif event.name == '0':
+                elif event.name == 'delete':
                     self.is_playing = False
                     self.data_dict['current_frame_index'] = 0
                 elif event.name == '[':
@@ -346,36 +349,45 @@ class LiGuard:
             
             # if the frame has changed, update the data dictionary with the new frame data
             if frame_changed:
+                profiler.add_target('Total Time / Step')
                 self.data_dict['previous_frame_index'] = self.data_dict['current_frame_index']
                 
                 if self.pcd_io:
+                    profiler.add_target('pcd_io')
                     current_point_cloud_path, current_point_cloud_numpy = self.pcd_io[self.data_dict['current_frame_index']]
                     self.data_dict['current_point_cloud_path'] = current_point_cloud_path
                     self.data_dict['current_point_cloud_numpy'] = current_point_cloud_numpy
+                    profiler.end_target('pcd_io')
                 elif 'current_point_cloud_numpy' in self.data_dict:
                     self.logger.log(f'[main.py->LiGuard->start]: current_point_cloud_numpy found in data_dict while pcd_io is None, removing ...', Logger.DEBUG)
                     self.data_dict.pop('current_point_cloud_numpy')
                 
                 if self.img_io:
+                    profiler.add_target('img_io')
                     current_image_path, current_image_numpy = self.img_io[self.data_dict['current_frame_index']]
                     self.data_dict['current_image_path'] = current_image_path
                     self.data_dict['current_image_numpy'] = current_image_numpy
+                    profiler.end_target('img_io')
                 elif 'current_image_numpy' in self.data_dict:
                     self.logger.log(f'[main.py->LiGuard->start]: current_image_numpy found in data_dict while img_io is None, removing ...', Logger.DEBUG)
                     self.data_dict.pop('current_image_numpy')
 
                 if self.clb_io:
+                    profiler.add_target('clb_io')
                     current_calib_path, current_calib_data = self.clb_io[self.data_dict['current_frame_index']]
                     self.data_dict['current_calib_path'] = current_calib_path
                     self.data_dict['current_calib_data'] = current_calib_data
+                    profiler.end_target('clb_io')
                 elif 'current_calib_data' in self.data_dict:
                     self.logger.log(f'[main.py->LiGuard->start]: current_calib_data found in data_dict while clb_io is None, removing ...', Logger.DEBUG)
                     self.data_dict.pop('current_calib_data')
                 
                 if self.lbl_io:
+                    profiler.add_target('lbl_io')
[…]
