### 2024-08-23 d5388a6
refactor: improved bbox regression in Cluster2Object.
