### 2024-08-23 5160c8e
refactor: remove unused 'oriented' flag in Cluster2Object.
