+++ configs/config_template.yml
@@ -104,7 +104,6 @@ proc: # liguard processing configurations
             enabled: False
             activate_on_key_set: 'current_point_cloud_numpy'
             priority: 5
-            oriented: False
             classes_require_orientation: ['Car', 'Van', 'Truck', 'Bus']
             size_constraints: # in increasing order of base lengths, units in meters, left inclusive, right exclusive
                 Pedestrian:
