### 2024-08-23 88dcda0
refactor: highlight text in ImageVisualizer using a filled rectangle.
