### 2024-08-30 4af7f5b
improve: added history_size parameter to GenerateKDTreePastTrajectory.
