+++ algo/post.py
@@ -162,7 +162,7 @@ def GenerateKDTreePastTrajectory(data_dict: dict, cfg_dict: dict):
     # algo name and dict keys
     algo_name = 'GenerateKDTreePastTrajectory'
     processed_frames_key = f'{algo_name}_processed_frames'
-    last_bbox_3d_labels_key = f'{algo_name}_last_bbox_3d_labels'
+    bbox_history_window_key = f'{algo_name}_last_bbox_3d_labels'
 
     # get params
     params = cfg_dict['proc']['post'][algo_name]
@@ -171,60 +171,73 @@ def GenerateKDTreePastTrajectory(data_dict: dict, cfg_dict: dict):
     if processed_frames_key in data_dict and data_dict['current_frame_index'] in data_dict[processed_frames_key]:
         logger.log(f'[algo->post.py->GenerateKDTreePastTrajectory]: Frame {data_dict["current_frame_index"]} already processed, skipping ...', Logger.WARNING)
         return
-
-    # ---------------------- KDTree Tracking ---------------------- #
+    
     current_bbox_3d_labels = []
     for i, label_dict in enumerate(data_dict['current_label_list']):
         if 'bbox_3d' not in label_dict: continue
         bbox_3d_dict = label_dict['bbox_3d']
         current_bbox_3d_labels.append([i, bbox_3d_dict])
-
-    if last_bbox_3d_labels_key not in data_dict:
-        data_dict[last_bbox_3d_labels_key] = current_bbox_3d_labels
-        return
-    last_bbox_3d_labels = data_dict[last_bbox_3d_labels_key]
     
+    # if no bbox_3d found in current frame, skip
     if len(current_bbox_3d_labels) == 0:
-        logger.log(f'[algo->post.py->GenerateKDTreePastTrajectory]: No bbox_3d found in current frame, skipping ...', Logger.WARNING)
+        logger.log(f'[algo->post.py->GenerateKDTreePastTrajectory]: No bbox_3d found in current frame, skipping ...', Logger.DEBUG)
+        return
+    
+    # create history window if it does not exist
+    if bbox_history_window_key not in data_dict:
+        data_dict[bbox_history_window_key] = [current_bbox_3d_labels]
         return
 
     # Create a KDTree for points_list
-    last_bbox_3d_xyz_center = np.array([bbox_3d_dict['xyz_center'] for _, bbox_3d_dict in last_bbox_3d_labels])
+    all_history_bbox_3d_xyz_centers = []
+    for bbox_3d_labels in data_dict[bbox_history_window_key]:
+        xyz_centers = np.array([bbox_3d_dict['xyz_center'] for _, bbox_3d_dict in bbox_3d_labels])
+        all_history_bbox_3d_xyz_centers.append(xyz_centers)
     current_bbox_3d_xyz_center = np.array([bbox_3d_dict['xyz_center'] for _, bbox_3d_dict in current_bbox_3d_labels])
     
+    # ---------------------- KDTree Tracking ---------------------- #
     kd_tree = KDTree(current_bbox_3d_xyz_center)
-    cost_matrix = np.zeros((len(last_bbox_3d_xyz_center), len(current_bbox_3d_xyz_center)))
+    already_matched = []
 
-    for i, point in enumerate(last_bbox_3d_xyz_center):
-        distances, indices = kd_tree.query(point, k=len(current_bbox_3d_xyz_center))
-        cost_matrix[i, indices] = distances
+    for history_idx, xyz_centers in enumerate(all_history_bbox_3d_xyz_centers):
+        cost_matrix = np.zeros((len(xyz_centers), len(current_bbox_3d_xyz_center)))
 
[…]
