### 2024-09-02 d4ffec0
feat: added support for saving and loading pcd visualizer last view.
