### 2024-09-02 71a3b3a
fix: skip GUI dictionary keys not present in global dictionary.
