### 2024-09-02 2e8147a
feat: added status strip to logger GUI.
