### 2024-09-02 488bb44
fix: update BaseConfiguration to handle missing GUI dictionary keys.
