+++ gui/config_gui.py
@@ -186,28 +186,28 @@ class BaseConfiguration:
             for key in item.keys():
                 # Create a global key by joining the parent keys and the current key
                 global_key = ".".join(parent_keys + [key])
-                if global_key not in G: continue # Skip if the key is not in the GUI dictionary
                 # If the value of the current key is a dictionary, recursively call the function
                 if type(item[key]) == dict:
                     self.__update_cfg_from_gui__(item[key], parent_keys + [key])
                 # If the value of the current key is a string, integer or float
-                elif type(item[key]) in [str, int, float]:
-                    # Update the value of the key in the item dictionary with the value from the GUI
-                    item[key] = G[global_key]['type'](G[global_key]['view'].text_value)
-                # If the value of the current key is a list
-                elif type(item[key]) == list:
-                    # Parse the list from the string value in the GUI
-                    item[key] = ast.literal_eval(G[global_key]['view'].text_value)
-                    # Convert each item in the list to the appropriate type
-                    item[key] = [G[global_key]['type'](i) for i in item[key]]    
-                # If the value of the current key is a boolean
-                elif type(item[key]) == bool:
-                    # Update the value of the key in the item dictionary with the checked state from the GUI
-                    item[key] = G[global_key]['type'](G[global_key]['view'].checked)
-                # If the value of the current key is of an unsupported type
-                else:
-                    # Raise an exception
-                    raise Exception("Unsupported type: {}".format(type(item[key])))
+                elif global_key in G:
+                    if type(item[key]) in [str, int, float]:
+                        # Update the value of the key in the item dictionary with the value from the GUI
+                        item[key] = G[global_key]['type'](G[global_key]['view'].text_value)
+                    # If the value of the current key is a list
+                    elif type(item[key]) == list:
+                        # Parse the list from the string value in the GUI
+                        item[key] = ast.literal_eval(G[global_key]['view'].text_value)
+                        # Convert each item in the list to the appropriate type
+                        item[key] = [G[global_key]['type'](i) for i in item[key]]    
+                    # If the value of the current key is a boolean
+                    elif type(item[key]) == bool:
+                        # Update the value of the key in the item dictionary with the checked state from the GUI
+                        item[key] = G[global_key]['type'](G[global_key]['view'].checked)
+                    # If the value of the current key is of an unsupported type
+                    else:
+                        # Raise an exception
+                        raise Exception("Unsupported type: {}".format(type(item[key])))
                 
     def __show_issue_dialog__(self):
         # create a dialog
