### 2024-09-02 cc49268
feat: update file_io.py to use global_zero for skipping fixed frames at start globally.
