+++ calib/file_io.py
@@ -27,6 +27,7 @@ class FileIO:
         self.clb_dir = os.path.join(cfg['data']['path'], cfg['data']['calib_subdir'])
         self.clb_type = cfg['data']['calib']['clb_type']
         self.clb_start_idx = cfg['data']['start']['calib']
+        self.global_zero - cfg['data']['start']['global_zero']
         self.clb_end_idx = self.clb_start_idx + cfg['data']['count']
         
         # Check if the calibration type is supported
@@ -39,7 +40,7 @@ class FileIO:
         files = glob.glob(os.path.join(self.clb_dir, '*' + self.clb_ext))
         file_basenames = [os.path.splitext(os.path.basename(file))[0] for file in files]
         file_basenames.sort(key=lambda file_name: int(''.join(filter(str.isdigit, file_name))))
-        self.files_basenames = file_basenames[self.clb_start_idx:self.clb_end_idx]
+        self.files_basenames = file_basenames[self.clb_start_idx:self.clb_end_idx][self.global_zero:]
         
         # read the calibration files in async mode
         self.data_lock = threading.Lock()
