+++ img/file_io.py
@@ -39,12 +39,13 @@ class FileIO:
         self.img_dir = os.path.join(cfg['data']['path'], cfg['data']['camera_subdir'])
         self.img_type = cfg['data']['camera']['img_type']
         self.img_start_idx = cfg['data']['start']['camera']
+        self.global_zero = cfg['data']['start']['global_zero']
         self.img_end_idx = self.img_start_idx + cfg['data']['count']
         files = glob.glob(os.path.join(self.img_dir, '*' + self.img_type))
         file_basenames = [os.path.splitext(os.path.basename(file))[0] for file in files]
         # Sort the file basenames based on the numerical part
         file_basenames.sort(key=lambda file_name: int(''.join(filter(str.isdigit, file_name))))
-        self.files_basenames = file_basenames[self.img_start_idx:self.img_end_idx]
+        self.files_basenames = file_basenames[self.img_start_idx:self.img_end_idx][self.global_zero:]
         self.reader = self.__read_img__
 
         self.data_lock = threading.Lock()
