+++ pcd/viz.py
@@ -32,7 +32,7 @@ class PointCloudVisualizer:
         # init
         # create necessary paths
         if cfg['visualization']['lidar']['save_images']:
-            self.lidar_save_path = cfg['visualization']['lidar']['save_path']
+            self.lidar_save_path = os.path.join(cfg['data']['outputs_dir'], 'pcd_viz')
             os.makedirs(self.lidar_save_path, exist_ok=True)
         # reset
         self.reset(cfg, True)
@@ -278,19 +278,21 @@ class PointCloudVisualizer:
         file_path = os.path.join(self.lidar_save_path, f'{frame_idx:08d}.png')
         self.viz.capture_screen_image(file_path)
 
-    def set_view_status(self):
+    def save_view_status(self):
         """
-        Updates the view status (parameters of looking camera) of the visualizer to the configuration.
+        Saves the view status (parameters of looking camera) of the visualizer.
         """
-        self.cfg['visualization']['lidar']['open3d_view_status_string'] = self.viz.get_view_status()
+        save_path = os.path.join(self.cfg['data']['outputs_dir'], 'view_status.txt')
+        with open(save_path, 'w') as file: file.write(str(self.viz.get_view_status()))
     
-    def reset_view_status(self):
+    def load_view_status(self):
         """
-        Resets the view status (parameters of looking camera) of the visualizer from the configuration.
+        Loads the view status (parameters of looking camera) of the visualizer from the configuration.
         """
-        if 'open3d_view_status_string' in self.cfg['visualization']['lidar']:
-            try: self.viz.set_view_status(self.cfg['visualization']['lidar']['open3d_view_status_string'])
-            except: pass
+        try:
+            load_path = os.path.join(self.cfg['data']['outputs_dir'], 'view_status.txt')
+            with open(load_path, 'r') as file: self.viz.set_view_status(file.read())
+        except: pass
 
     def quit(self):
         """ 
