### 2024-09-04 ed01f5d
chore: added ground_level, max_foot_height, and min_height in Cluster2Obj for better removal of noisy bboxes.
