+++ configs/config_template.yml
@@ -106,6 +106,9 @@ proc: # liguard processing configurations
             enabled: False
             activate_on_key_set: 'current_point_cloud_numpy'
             priority: 5
+            ground_level: -3.2
+            max_foot_level: 1.4
+            min_height: 0.6
             classes_require_orientation: ['Car', 'Van', 'Truck', 'Bus']
             size_constraints: # in increasing order of base lengths, units in meters, left inclusive, right exclusive
                 Pedestrian:
