### 2024-09-04 0855d2b
chore: safe call to self.pcd_visualizer.save_view_status(), added os.makedirs for cfg['data']['outputs_dir'] to fix missing dirs error later.
