### 2024-09-04 461c878
multi: refactored FusePredictedBBoxesFromSourceToTarget to Fuse2DPredictedBBoxes for better clearity, removed bbox_3d processing in it; fixed safe call to pcd_visualizer.save_view_status function; create cfg['data']['outputs_dir'] at start and reset
