+++ algo/post.py
@@ -2,18 +2,18 @@
 
 from gui.logger_gui import Logger
 
-def FusePredictedBBoxesFromSourceToTarget(data_dict: dict, cfg_dict: dict):
+def Fuse2DPredictedBBoxes(data_dict: dict, cfg_dict: dict):
     """
-    Fuse source bbox information to target bbox information.
+    Fuse bbox_2d information among ModalityA and ModalityB.
 
     Requirements:
-    - bbox_3d in current_label_list, predicted by an arbitrary <bbox_target_algo>.
-    - Enable proc->lidar->gen_2d_bbox to convert bbox_3d to bbox_2d.
-    - bbox_2d in current_2d_bbox_list, predicted by an arbitrary <bbox_source_algo>.
+    - bbox_2d in current_label_list, predicted by an arbitrary ModalityA's <algo_src>.
+    - bbox_2d in current_label_list, predicted by an arbitrary ModalityB's <algo_src>.
 
     Operation:
-    - It uses KD-Tree to find the nearest bbox_2d.
-    - It assigns the class of bbox_2d to bbox_3d.
+    - It uses KD-Tree to find the nearest bbox_2d from ModalityA's <algo_src> and ModalityB's <algo_src>.
+    - It assigns the class based on the highest confidence score.
+    - It assigns the depth based on the highest confidence score.
     - TODO: figure out more methods to enhance bbox fusion.
 
     Args:
@@ -25,107 +25,135 @@ def FusePredictedBBoxesFromSourceToTarget(data_dict: dict, cfg_dict: dict):
     """
     # Get logger object from data_dict
     if 'logger' in data_dict: logger:Logger = data_dict['logger']
-    else: print('[algo->post.py->FusePredictedBBoxesFromSourceToTarget]: No logger object in data_dict. It is abnormal behavior as logger object is created by default. Please check if some script is removing the logger key in data_dict.'); return
+    else: print('[algo->post.py->Fuse2DPredictedBBoxes]: No logger object in data_dict. It is abnormal behavior as logger object is created by default. Please check if some script is removing the logger key in data_dict.'); return
 
     # Check if required data is present in data_dict
     if 'current_label_list' not in data_dict:
-        logger.log('[algo->post.py->FusePredictedBBoxesFromSourceToTarget]: current_label_list not found in data_dict', Logger.ERROR)
+        logger.log('[algo->post.py->Fuse2DPredictedBBoxes]: current_label_list not found in data_dict', Logger.ERROR)
         return
     if 'current_calib_data' not in data_dict:
-        logger.log('[algo->post.py->FusePredictedBBoxesFromSourceToTarget]: current_calib_data not found in data_dict. Ignoring bbox_3d <--> bbox_2d fusion.', Logger.WARNING)
+        logger.log('[algo->post.py->Fuse2DPredictedBBoxes]: current_calib_data not found in data_dict', Logger.ERROR)
     
     # imports
     import numpy as np
     from scipy.spatial import KDTree
-    from scipy.optimize import linear_sum_assignment
 
     # algo name and dict keys
-    algo_name = 'FusePredictedBBoxesFromSourceToTarget'
-    tr_pcd_to_img_key = f'{algo_name}_tr_pcd_to_img'
-    inv_tr_pcd_to_img_key = f'{algo_name}_inv_tr_pcd_to_img'
+    algo_name = 'Fuse2DPredictedBBoxes'
 
     # get params
     params = cfg_dict['proc']['post'][algo_name]
 
     # ---------------------- KDTree Matching ---------------------- #
[…]
