### 2024-09-04 6a567ee
refactor: fix typo in file_io.py global_zero assignment and default value of algo in config_template.yml
