### 2024-10-29 012967d
refactor: removed NNCluster2Object function from lidar.py and its configuration from config_template.yml.
