+++ algo/lidar.py
@@ -514,127 +514,6 @@ def Cluster2Object(data_dict: dict, cfg_dict: dict):
         label['bbox_3d'] = {'xyz_center': xyz_center, 'xyz_extent': xyz_extent, 'xyz_euler_angles': xyz_euler_angles, 'rgb_color': rgb_color, 'predicted': True, 'added_by': algo_name}
         data_dict['current_label_list'].append(label)
 
-def NNCluster2Object(data_dict: dict, cfg_dict: dict):
-    """
-    Converts clustered lidar data into object labels using a neural network model.
-
-    Args:
-        data_dict (dict): A dictionary containing the required data for processing.
-        cfg_dict (dict): A dictionary containing the configuration parameters.
-
-    Returns:
-        None
-    """
-    # get logger object from data_dict
-    if 'logger' in data_dict: logger:Logger = data_dict['logger']
-    else: print('[algo->lidar.py->NNCluster2Object]: No logger object in data_dict. It is abnormal behavior as logger object is created by default. Please check if some script is removing the logger key in data_dict.'); return
-
-    # algo name and keys used in algo
-    algo_name = 'NNCluster2Object'
-    model_key = f'{algo_name}_model'
-    class_ids_key = f'{algo_name}_class_ids'
-    
-    # check if required data is present in data_dict
-    if "current_point_cloud_numpy" not in data_dict:
-        logger.log('[algo->lidar.py->NNCluster2Object]: current_point_cloud_numpy not found in data_dict', Logger.ERROR)
-        return
-    if 'current_label_list' not in data_dict:
-        logger.log('[algo->lidar.py->NNCluster2Object]: current_label_list not found in data_dict', Logger.ERROR)
-        return
-    if cfg_dict['proc']['lidar']['NNCluster2Object']['activate_on_key_set'] not in data_dict: return
-    
-    # get params
-    import os
-    params = cfg_dict['proc']['lidar'][algo_name]
-    checkpoint_path = os.path.join(data_dict['root_path'], params['point_nn_bbox_est_checkpoint_path'])
-    if os.path.exists(checkpoint_path) == False:
-        logger.log(f'[algo->lidar.py->NNCluster2Object]: checkpoint path {checkpoint_path} does not exist', Logger.ERROR)
-        return
-    num_points = params['point_nn_bbox_est_num_points']
-    
-    # load model
-    try: torch = __import__('torch')
-    except:
-        logger.log('[algo->lidar.py->NNCluster2Object]: torch package not found, please install the `torch` package using `pip install torch`.', Logger.ERROR)
-        return
-    device = 'cuda:0' if torch.cuda.is_available() else 'cpu'
-    import open3d as o3d
-    
-    if model_key not in data_dict:
-        try:
-            from algo.nn.PointNN.models.bbox_est import BBoxEst
-            model = BBoxEst(num_points, 512).to(device)
-            model.load_state_dict(torch.load(checkpoint_path))
-            model.eval()
-            data_dict[model_key] = model
-            from algo.nn.PointNN.ds_loaders.pcdet import PerObjectDataLoader
-            data_dict[class_ids_key] = PerObjectDataLoader.class_id_to_key
-        except:
-            logger.log('[algo->lidar.py->NNCluster2Object]: failed to load model', Logger.ERROR)
-            return
-    
-    # process clusters
-    with torch.no_grad():
-        for label_dict in data_dict['current_label_list']:
-            if 'lidar_cluster' not in label_dict: continue
-            
-            # get cluster
-            lidar_cluster_dict = label_dict['lidar_cluster']
-            point_indices = lidar_cluster_dict['point_indices']
-            cluster = data_dict['current_point_cloud_numpy'][point_indices][:, :3]
-
-            # this bbox is used to get the orientation of the object
-            orienter_box = None
[…]
