### 2024-10-29 a9b7f8b
refactor: improved file naming and handling in STDF.py
