+++ algo/lidar.py
@@ -300,6 +300,7 @@ def BGFilterSTDF(data_dict: dict, cfg_dict: dict):
         data_dict[filter_key] = lambda pcd, threshold: make_STDF_filter(pcd, threshold, **filter_params)
         # save filter
         filter_saved_path = save_STDF_params(filter_params, params['filter_path'])
+        cfg_dict['proc']['lidar']['BGFilterSTDF']['filter_path'] = filter_saved_path
         logger.log(f'[algo->lidar.py->BGFilterSTDF]: Filter generated and saved in {filter_saved_path}.', Logger.INFO)
         data_dict['BGFilterSTDF_set'] = True
     else:
