+++ algo/non_nn/STDF.py
@@ -2,6 +2,7 @@ import os
 import pickle
 
 import numpy as np
+import time
 
 """
 Author: Muhammad Shahbaz (m.shahbaz.kharal@outlook.com)
@@ -9,14 +10,13 @@ Github: github.com/m-shahbaz-kharal
 
 Spatio-Temporal Density Filter (STDF)
 
-It is a background filter for roadside LIDAR "unstructured" point cloud data.
-This (that is the term unstructured) essentially means that the point cloud cannot be represented by a fixed size HXV (horizontal resolution x vertical resolution) matrix.
+A simple implementation of a spatio-temporal density-based background filter that can operate on "unstructured" point cloud data (as well as structured). This term "unstructured", here, means that the point cloud cannot be represented by a fixed size HXV (horizontal resolution x vertical resolution) matrix.
 
 For example:
     - If a LiDAR generate arbitrary number of points in each frame,
     - and each point doesn't necessarily belong to a fixed azimuth and elevation ray.
     - then the point cloud is unstructured.
-The algorithm is designed to work with such unstructured point clouds.
+The algorithm can be used for such unstructured point clouds.
 
 The algorithm is summarized in the following steps:
 Since the number of points in each frame is not fixed, all the point clouds (gathered temporally) are stacked together to form a single point cloud.
@@ -27,7 +27,6 @@ Since the number of points in each frame is not fixed, all the point clouds (gat
     b. The density of the bin is then compared with a threshold:
         - If the density is less than the threshold, the point is considered as foreground,
         - otherwise background.
-
 """
 
 def calc_STDF_params(point_cloud_set: list, # a list of point clouds, points in each frame must be equal
@@ -63,42 +62,16 @@ def make_STDF_filter(point_cloud: np.ndarray, # Nx3 or Nx4,
 
 def save_STDF_params(filter_params, filename):
     if os.path.exists(filename):
-        if '.' in filename: filename = filename.split('.')[0]
-        if '_' in filename: possible_num = filename.split('_')[-1]
-        if possible_num.isdigit():
-            num = int(possible_num)
-            filename = filename.split('_')[:-1] + str(num+1).zfill(2)
-        else:
-            filename += '_01'
-        filename += '.pkl'
+        filename = filename.split('.')[0]
+        filename = filename + '_' + time.strftime("%Y%m%d-%H%M%S") + '.pkl'
     else:
-        if '.' in filename: filename = filename.split('.')[0]
-        if '_' in filename: possible_num = filename.split('_')[-1]
-        if possible_num.isdigit(): filename += '.pkl'
-        else: filename += '_01.pkl'
-    
+        filename = filename.split('.')[0]
+        filename = filename + '.pkl'
     with open(filename, 'wb') as f: pickle.dump(filter_params, f)
     return filename
 
 def load_STDF_params(filename):
-    if '.' in filename: filename = filename.split('.')[0]
-    if '_' in filename: possible_num = filename.split('_')[-1]
-    if possible_num.isdigit():
-        filename = filename + '.pkl'
-        if os.path.exists(filename):
-            with open(filename, 'rb') as f: return pickle.load(f)
-        else: return None
-    elif os.path.exists(filename + '.pkl'):
-        with open(filename + '.pkl', 'rb') as f: return pickle.load(f)
-    else:
-        try:
-            parent_path = os.path.dirname(filename)
-            if len(parent_path) > 0: files = os.listdir(parent_path)
-            else: files = os.listdir()
-            files = [f for f in files if f.startswith(filename)]
-            if len(files) == 0: return None
-            files.sort()
-            filename = files[-1]
-            filename = os.path.join(parent_path, filename)
[…]
