### 2024-10-29 b45b3fa
refactor: update file extension handling in DHistDPP.py and STDF.py
