+++ algo/non_nn/DHistDPP.py
@@ -79,6 +79,7 @@ def save_DHistDPP_params(filter_params, filename):
 
 def load_DHistDPP_params(filename):
     try:
+        filename = filename.split('.')[0] + '.pkl'
         with open(filename, 'rb') as f: return pickle.load(f)
     except:
         return None
