### 2024-10-29 08bdb85
refactor: simplified file naming and handling in DHistDPP.py and STDF.py
fix: remove bug in pointpillars implementation

- Fix a bug in the pointpillars implementation that was causing incorrect orientation of 3D bbox visualizations.

refactor: remove NNCluster2Object function from lidar.py and its configuration from config_template.yml

- Remove the NNCluster2Object function from lidar.py as it is no longer needed.
- Remove the corresponding configuration entry for NNCluster2Object from config_template.yml.

refactor: update logger message format in logger_gui.py

- Update the format of logge
[…]
