+++ algo/post.py
@@ -1,10 +1,13 @@
-# contains more generic post-processing algorithms for the data
+# contains post-processing algorithms for the data
 
 from gui.logger_gui import Logger
+from algo.utils import AlgoType, make_key, get_algo_params
 
-def Fuse2DPredictedBBoxes(data_dict: dict, cfg_dict: dict):
+algo_type = AlgoType.post
+
+def Fuse2DPredictedBBoxes(data_dict: dict, cfg_dict: dict, logger: Logger):
     """
-    Fuse bbox_2d information among ModalityA and ModalityB.
+    Fuses bbox_2d information among ModalityA and ModalityB.
 
     Requirements:
     - bbox_2d in current_label_list, predicted by an arbitrary ModalityA's <algo_src>.
@@ -14,35 +17,26 @@ def Fuse2DPredictedBBoxes(data_dict: dict, cfg_dict: dict):
     - It uses KD-Tree to find the nearest bbox_2d from ModalityA's <algo_src> and ModalityB's <algo_src>.
     - It assigns the class based on the highest confidence score.
     - It assigns the depth based on the highest confidence score.
-    - TODO: figure out more methods to enhance bbox fusion.
 
     Args:
         data_dict (dict): A dictionary containing the required data.
         cfg_dict (dict): A dictionary containing configuration parameters.
-
-    Returns:
-        None
+        logger (gui.logger_gui.Logger): A logger object for logging messages and errors in GUI.
     """
-    # Get logger object from data_dict
-    if 'logger' in data_dict: logger:Logger = data_dict['logger']
-    else: print('[algo->post.py->Fuse2DPredictedBBoxes]: No logger object in data_dict. It is abnormal behavior as logger object is created by default. Please check if some script is removing the logger key in data_dict.'); return
+    # get name and params
+    algo_name = 'Fuse2DPredictedBBoxes'
+    params = get_algo_params(cfg_dict, algo_type, algo_name, logger)
 
     # Check if required data is present in data_dict
     if 'current_label_list' not in data_dict:
-        logger.log('[algo->post.py->Fuse2DPredictedBBoxes]: current_label_list not found in data_dict', Logger.ERROR)
+        logger.log('current_label_list not found in data_dict', Logger.ERROR)
         return
     if 'current_calib_data' not in data_dict:
-        logger.log('[algo->post.py->Fuse2DPredictedBBoxes]: current_calib_data not found in data_dict', Logger.ERROR)
+        logger.log('current_calib_data not found in data_dict', Logger.ERROR)
     
     # imports
     import numpy as np
     from scipy.spatial import KDTree
-
-    # algo name and dict keys
-    algo_name = 'Fuse2DPredictedBBoxes'
-
-    # get params
-    params = cfg_dict['proc']['post'][algo_name]
 
     # ---------------------- KDTree Matching ---------------------- #
     # Extract info from bbox_2d from modality a and b
@@ -67,10 +61,10 @@ def Fuse2DPredictedBBoxes(data_dict: dict, cfg_dict: dict):
                 
 
     if len(mod_a_list) == 0:
-        logger.log(f"[algo->post.py->Fuse2DPredictedBBoxes]: No bbox_2d found from ModalityA's <algo_src>", Logger.DEBUG)
+        logger.log('No bbox_2d found from ModalityA\'s <algo_src>', Logger.DEBUG)
         return
     if len(mod_b_list) == 0:
-        logger.log(f"[algo->post.py->Fuse2DPredictedBBoxes]: No bbox_2d found from ModalityB's <algo_src>", Logger.DEBUG)
[…]
