### 2024-10-30 222d517
chore: use params instead of manually accessing cfg_dict
