### 2024-10-30 37e4e11
chore: added line for logger in docstring
