### 2024-10-30 fb7530a
chore: consistent use of logger.log function (no manually adding context as that is handled by logger class automatically via inspect library).
