### 2024-10-30 cc6bc13
chore: increased consistency by making convention how algo_name variable is set.
chore: increased consistency by making similiar imports in files under algo module

refactor: removed 'activate_on_key_set' functionality as it is redundent to 'enabled' in configuration.

chore: removed single line comment on top of every file under algo module, need to find better documentation strategy.
