+++ algo/lidar.py
@@ -1,14 +1,11 @@
-# contains point-cloud processing algorithms
+import inspect
+from gui.logger_gui import Logger
+from algo.utils import AlgoType, make_key, get_algo_params
+algo_type = AlgoType.lidar
 
 import os
 import sys
-
 import numpy as np
-
-from gui.logger_gui import Logger
-from algo.utils import AlgoType, make_key, get_algo_params
-
-algo_type = AlgoType.lidar
 
 def rotate(data_dict: dict, cfg_dict: dict, logger: Logger):
     """
@@ -20,7 +17,7 @@ def rotate(data_dict: dict, cfg_dict: dict, logger: Logger):
         logger (gui.logger_gui.Logger): A logger object for logging messages and errors in GUI.
     """
     # get name and params
-    algo_name = 'rotate'
+    algo_name = inspect.stack()[0].function
     params = get_algo_params(cfg_dict, algo_type, algo_name, logger)
     
     # check if required data is present in data_dict
@@ -51,7 +48,7 @@ def crop(data_dict: dict, cfg_dict: dict, logger: Logger):
         logger (gui.logger_gui.Logger): A logger object for logging messages and errors in GUI.
     """
     # get name and params
-    algo_name = 'crop'
+    algo_name = inspect.stack()[0].function
     params = get_algo_params(cfg_dict, algo_type, algo_name, logger)
 
     # check if required data is present in data_dict
@@ -83,7 +80,7 @@ def project_image_pixel_colors(data_dict: dict, cfg_dict: dict, logger: Logger):
         logger (gui.logger_gui.Logger): A logger object for logging messages and errors in GUI.
     """
     # get name and params
-    algo_name = 'project_image_pixel_colors'
+    algo_name = inspect.stack()[0].function
     params = get_algo_params(cfg_dict, algo_type, algo_name, logger)
     
     # check if required data is present in data_dict
@@ -130,7 +127,7 @@ def BGFilterDHistDPP(data_dict: dict, cfg_dict: dict, logger: Logger):
         logger (gui.logger_gui.Logger): A logger object for logging messages and errors in GUI.
     """
     # get name and params
-    algo_name = 'BGFilterDHistDPP'
+    algo_name = inspect.stack()[0].function
     params = get_algo_params(cfg_dict, algo_type, algo_name, logger).copy()
     live_editable_params = ['background_density_threshold'] # list of params that can be live edited and do not require re-computation of filter
     
@@ -160,11 +157,9 @@ def BGFilterDHistDPP(data_dict: dict, cfg_dict: dict, logger: Logger):
             data_dict[filter_key] = lambda pcd, threshold: make_DHistDPP_filter(pcd, threshold, **filter_params)
             data_dict[filter_loaded_key] = True
             logger.log(f'Filter loaded from {params["filter_path"]}', Logger.INFO)
-            data_dict[f'{algo_name}_set'] = True
         else:
             data_dict[filter_loaded_key] = False
             logger.log(f'Failed to load filter from {params["filter_path"]}. Calculating ...', Logger.WARNING)
-            data_dict[f'{algo_name}_set'] = False
     
     # generate filter if not exists
     if filter_key not in data_dict:
@@ -191,7 +186,6 @@ def BGFilterDHistDPP(data_dict: dict, cfg_dict: dict, logger: Logger):
         # save filter
         filter_saved_path = save_DHistDPP_params(filter_params, params['filter_path'])
         logger.log(f'Filter saved at {filter_saved_path}', Logger.INFO)
-        data_dict[f'{algo_name}_set'] = True
     else:
         # recompute filter if non-live-editable params are changed
         condition = False
[…]
