### 2024-10-30 ce4a412
added: traceback to logging
