+++ main.py
@@ -1,4 +1,6 @@
 import os
+import traceback
+
 import open3d.visualization.gui as gui
 
 from gui.config_gui import BaseConfiguration as BaseConfigurationGUI
@@ -148,16 +150,16 @@ class LiGuard:
             try:
                 self.pcd_io = PCD_File_IO(cfg)
                 self.logger.log('PCD_File_IO created', Logger.DEBUG)
-            except Exception as e:
-                self.logger.log('PCD_File_IO creation failed:\n{e}', Logger.CRITICAL)
+            except Exception:
+                self.logger.log(f'PCD_File_IO creation failed:\n{traceback.format_exc()}', Logger.CRITICAL)
                 self.pcd_io = None
         # if sensors are enabled
         elif cfg['sensors']['lidar']['enabled']:
             try:
                 self.pcd_io = PCD_Sensor_IO(cfg)
                 self.logger.log('PCD_Sensor_IO created', Logger.DEBUG)
-            except Exception as e:
-                self.logger.log('PCD_Sensor_IO creation failed:\n{e}', Logger.CRITICAL)
+            except Exception:
+                self.logger.log(f'PCD_Sensor_IO creation failed:\n{traceback.format_exc()}', Logger.CRITICAL)
                 self.pcd_io = None
         else: self.pcd_io = None
         # get the total number of pcd frames
@@ -172,8 +174,8 @@ class LiGuard:
                     self.pcd_visualizer = PointCloudVisualizer(self.app, cfg)
                     self.pcd_visualizer.load_view_status()
                     self.logger.log('PointCloudVisualizer created', Logger.DEBUG)
-                except Exception as e:
-                    self.logger.log('PointCloudVisualizer creation failed:\n{e}', Logger.CRITICAL)
+                except Exception:
+                    self.logger.log(f'PointCloudVisualizer creation failed:\n{traceback.format_exc()}', Logger.CRITICAL)
                     self.pcd_visualizer = None
         # manage image reading
         if self.img_io != None: self.img_io.close()
@@ -182,16 +184,16 @@ class LiGuard:
             try:
                 self.img_io = IMG_File_IO(cfg)
                 self.logger.log('IMG_File_IO created', Logger.DEBUG)
-            except Exception as e:
-                self.logger.log('IMG_File_IO creation failed:\n{e}', Logger.CRITICAL)
+            except Exception:
+                self.logger.log(f'IMG_File_IO creation failed:\n{traceback.format_exc()}', Logger.CRITICAL)
                 self.img_io = None
         # if sensors are enabled
         elif cfg['sensors']['camera']['enabled']:
             try:
                 self.img_io = IMG_Sensor_IO(cfg)
                 self.logger.log('IMG_Sensor_IO created', Logger.DEBUG)
-            except Exception as e:
-                self.logger.log('IMG_Sensor_IO creation failed:\n{e}', Logger.CRITICAL)
+            except Exception:
+                self.logger.log(f'IMG_Sensor_IO creation failed:\n{traceback.format_exc()}', Logger.CRITICAL)
                 self.img_io = None
         else: self.img_io = None
         # get the total number of image frames
@@ -204,8 +206,8 @@ class LiGuard:
                 if self.img_visualizer != None: self.img_visualizer.reset(cfg)
                 else: self.img_visualizer = ImageVisualizer(self.app, cfg)
                 self.logger.log('ImageVisualizer created', Logger.DEBUG)
-            except Exception as e:
-                self.logger.log('ImageVisualizer creation failed:\n{e}', Logger.CRITICAL)
+            except Exception:
+                self.logger.log(f'ImageVisualizer creation failed:\n{traceback.format_exc()}', Logger.CRITICAL)
[…]
