### 2024-10-30 d0f0f1d
chore: remove redundent calls to astype(np.float32)
