### 2024-10-30 4123a16
chore: made logging messages consistent with rest of the code
