### 2024-10-30 1f4c996
refactor: PointPillarDetection now allows adding github repo and checkpoint file from absolute paths instead of relative to LiGuard root_path.
chore: added better comments to config_gui.py

refactor: improved standardization of showing dialogs

added: browse button on configs that has 'dir_path|<value>' or 'file_path|<value>'

chore: checks on self.cfg
