+++ gui/config_gui.py
@@ -8,10 +8,10 @@ import ast
 class BaseConfiguration:
     def get_callbacks_dict():
         """
-        Returns a dictionary containing callback functions for different actions.
+        Generates a template dictionary containing keys for applicable callback function names to be used as `BaseConfiguration` class's `callbacks` argument.
 
         Returns:
-            dict: A dictionary with keys representing different actions and values as empty lists.
+            dict: A dictionary with keys representing applicable callback actions and values as empty lists.
         """
         return {
             'new_config': [],       # List to store callback functions for 'new_config' action
@@ -21,32 +21,26 @@ class BaseConfiguration:
             'apply_config': [],     # List to store callback functions for 'apply_config' action
             'quit_config': []       # List to store callback functions for 'quit_config' action
         }
-    def get_callbacks_dict():
-        return {'new_config': [],
-                'open_config': [],
-                'save_config': [],
-                'save_as_config': [],
-                'apply_config': [],
-                'quit_config': []}
         
     def __init__(self, app: gui.Application, callbacks = get_callbacks_dict()):
         """
         Initializes the ConfigGUI class.
 
-        Parameters:
-        - app (gui.Application): The application object.
-        - callbacks (dict): A dictionary of callbacks.
-
-        Returns:
-        - None
+        Args:
+            app (gui.Application): The application object.
+            callbacks (dict): A dictionary containing the callbacks for the GUI.
         """
         self.app = app
-        
-        self.mwin = app.create_window("Configuration", 480, 1080, x=0, y=30)
-        self.em = self.mwin.theme.font_size
         self.callbacks = callbacks
+
+        # Create a window for the configuration GUI
+        self.mwin = app.create_window("Configuration", 480, 1080, x=0, y=30)
+        # Define the em unit
+        self.em = self.mwin.theme.font_size
+        # Set the close callback
         self.mwin.set_on_close(self.__quit_config__)
         
+        # Initialize the layout
         self.__init__layout__()
         
     def __init__layout__(self):
@@ -142,22 +136,45 @@ class BaseConfiguration:
         with open(cfg_path, 'w') as f:
             yaml.safe_dump(cfg, f, sort_keys=False)
             
-    def __update_config_gui_from_cfg__(self, item, container, parent_keys=[], margin=0.2):
+    def __update_gui_from_cfg__(self, item, container, parent_keys=[], margin=0.2):
+        """
+        This function recursively generates the GUI from the configuration dictionary based on the item types.
+        """
         G = self.generated_config_gui_dict
         if type(item) == dict:
             for key in item.keys():
                 label_text = key
                 if type(item[key]) == dict:
                     collapsable_container = gui.CollapsableVert(label_text, self.em * 0.2, gui.Margins(self.em * margin, self.em * 0.2, self.em * 0.2, self.em * 0.2))
-                    self.__update_config_gui_from_cfg__(item[key], collapsable_container, parent_keys + [key], margin + 0.2)
+                    self.__update_gui_from_cfg__(item[key], collapsable_container, parent_keys + [key], margin + 0.2)
                     container.add_child(collapsable_container)
                 elif type(item[key]) in [str, int, float]:
                     sub_container = gui.Horiz()
                     sub_container.add_child(gui.Label(label_text + ":"))
                     global_key = ".".join(parent_keys + [key])
                     G[global_key] = {'view': gui.TextEdit(), 'type': type(item[key])}
[…]
