+++ configs/config_template.yml
@@ -1,7 +1,7 @@
 # This is a LiGuard pipeline configuration file.
 
 data: # dataset configurations
-    path: 'data' # root directory containing dataset
+    path: 'dir_path|data' # root directory containing dataset
     lidar_subdir: 'lidar' # subdirectory containing point clouds
     camera_subdir: 'camera' # subdirectory containing images
     label_subdir: 'label' # subdirectory containing labels
@@ -26,7 +26,7 @@ data: # dataset configurations
     label:
         enabled: False # set True to read labels from disk
         lbl_type: 'kitti' # can be kitti, openpcdet, or sustechpoints
-    outputs_dir: 'outputs' # directory to save outputs
+    outputs_dir: 'dir_path|outputs' # directory to save outputs
 
 sensors: # lidar and camera configurations
     lidar: # lidar sensor configurations, at this point only Ouster lidars are supported, support for other lidars is coming soon
@@ -67,7 +67,7 @@ proc: # liguard processing configurations
             lidar_range_in_unit_length: 100 # maximum range of lidar in lidar unit length
             bins_per_unit_length: 2 # number of bins per unit length
             background_density_threshold: 0.5 # threshold that tells if a bin is dense enough to be considered as background
-            filter_path: 'bg_filter_dhist_dpp' # path to save/load filter
+            filter_path: 'file_path|bg_filter_dhist_dpp' # path to save/load filter
             load_filter: False # set True to load filter
         rotate: # rotate point cloud
             enabled: False # set True to rotate point cloud
@@ -88,7 +88,7 @@ proc: # liguard processing configurations
             lidar_range_in_unit_length: 100 # maximum range of lidar in lidar unit length
             bins_per_unit_length: 2 # number of bins per unit length
             background_density_threshold: 0.5 # threshold that tells if a bin is dense enough to be considered as background
-            filter_path: 'bg_filter_stdf' # path to save/load filter
+            filter_path: 'file_path|bg_filter_stdf' # path to save/load filter
             load_filter: False # set True to load filter
         Clusterer_TEPP_DBSCAN: # Theoretically Efficient and Practical Parallel DBSCAN point clustering algorithm
             enabled: False # set True to cluster point cloud using TEPP DBSCAN
@@ -139,8 +139,8 @@ proc: # liguard processing configurations
         PointPillarDetection:
             enabled: False
             priority: 7
-            path_to_github_repo: 'algo/nn/PointPillars' # clone https://github.com/zhulf0804/PointPillars to this path and install the requirements
-            pretrained_ckpt_file: 'epoch_160.pth' # full name of checkpoint file under algo/nn/PointPillars/pretrained
+            path_to_github_repo: 'dir_path|' # clone https://github.com/zhulf0804/PointPillars to this path and install the requirements
+            pretrained_ckpt_file: 'file_path|' # full name of checkpoint file under algo/nn/PointPillars/pretrained
             score_threshold: 0.5
         gen_bbox_2d:
             enabled: False
