### 2024-10-31 696acdc
refactor: created get_abs_path function in config.gui to handle path consistently in entire app.
fix: no frame jump over max_frame_index.

chore: consistency of imports in files under algo.

fix: PointPillarDetection repository directory and checkpoint loading.

feat: now user can create yaml entry key ending with '_dir' or '_file' to automatically create a browse button in GUI.

chore: improved dialogs creation in config_gui.py.

chore: cleanup in Logger class.
