+++ algo/non_nn/DHistDPP.py
@@ -1,8 +1,5 @@
-import os
 import pickle
-
 import numpy as np
-import time
 
 """
 Author: Muhammad Shahbaz (m.shahbaz.kharal@outlook.com)
