+++ img/viz.py
@@ -1,4 +1,5 @@
 import os
+from gui.config_gui import get_abs_path
 
 import open3d as o3d
 import cv2
@@ -37,7 +38,7 @@ class ImageVisualizer:
         # init
         # create necessary paths
         if cfg['visualization']['camera']['save_images']:
-            self.image_save_path = os.path.join(cfg['data']['outputs_dir'], 'img_viz')
+            self.image_save_path = os.path.join(get_abs_path(cfg['data']['outputs_dir']), 'img_viz')
             os.makedirs(self.image_save_path, exist_ok=True)
         # reset
         self.reset(cfg, True)
