+++ pcd/viz.py
@@ -1,4 +1,5 @@
 import os
+from gui.config_gui import get_abs_path
 
 import open3d as o3d
 import numpy as np
@@ -32,7 +33,7 @@ class PointCloudVisualizer:
         # init
         # create necessary paths
         if cfg['visualization']['lidar']['save_images']:
-            self.lidar_save_path = os.path.join(cfg['data']['outputs_dir'], 'pcd_viz')
+            self.lidar_save_path = os.path.join(get_abs_path(cfg['data']['outputs_dir']), 'pcd_viz')
             os.makedirs(self.lidar_save_path, exist_ok=True)
         # reset
         self.reset(cfg, True)
@@ -283,7 +284,7 @@ class PointCloudVisualizer:
         Saves the view status (parameters of looking camera) of the visualizer.
         """
         if self.win_created:
-            save_path = os.path.join(self.cfg['data']['outputs_dir'], 'view_status.txt')
+            save_path = os.path.join(get_abs_path(self.cfg['data']['outputs_dir']), 'view_status.txt')
             with open(save_path, 'w') as file: file.write(str(self.viz.get_view_status()))
     
     def load_view_status(self):
@@ -291,7 +292,7 @@ class PointCloudVisualizer:
         Loads the view status (parameters of looking camera) of the visualizer from the configuration.
         """
         try:
-            load_path = os.path.join(self.cfg['data']['outputs_dir'], 'view_status.txt')
+            load_path = os.path.join(get_abs_path(self.cfg['data']['outputs_dir']), 'view_status.txt')
             with open(load_path, 'r') as file: self.viz.set_view_status(file.read())
         except: pass
 
