+++ calib/file_io.py
@@ -1,4 +1,5 @@
 import os
+from gui.config_gui import get_abs_path
 import glob
 import time
 import threading
@@ -23,7 +24,7 @@ class FileIO:
         """
         # get params from config
         self.cfg = cfg
-        self.clb_dir = os.path.join(cfg['data']['path'], cfg['data']['calib_subdir'])
+        self.clb_dir = os.path.join(get_abs_path(cfg['data']['main_dir']), cfg['data']['calib_subdir'])
         self.clb_type = cfg['data']['calib']['clb_type']
         self.clb_start_idx = cfg['data']['start']['calib']
         self.global_zero = cfg['data']['start']['global_zero']
