+++ algo/pre.py
@@ -1,4 +1,5 @@
 import inspect
+from gui.config_gui import get_abs_path
 from gui.logger_gui import Logger
 from algo.utils import AlgoType, make_key, get_algo_params
 algo_type = AlgoType.pre
