+++ gui/config_gui.py
@@ -1,9 +1,15 @@
 import open3d.visualization.gui as gui
 
 import os
+import sys
 import time
 import yaml
 import ast
+
+application_root_dir = os.path.abspath(os.path.dirname(sys.argv[0]))
+def get_abs_path(path:str) -> str:
+    if not os.path.isabs(path): path = os.path.join(application_root_dir, path)
+    return os.path.abspath(path)
 
 class BaseConfiguration:
     def get_callbacks_dict():
@@ -153,27 +159,20 @@ class BaseConfiguration:
                     sub_container.add_child(gui.Label(label_text + ":"))
                     global_key = ".".join(parent_keys + [key])
                     G[global_key] = {'view': gui.TextEdit(), 'type': type(item[key])}
+                    G[global_key]['view'].text_value = str(item[key])
                     sub_container.add_child(G[global_key]['view'])
-                    if str(item[key]).startswith('dir_path'):
-                        G[global_key]['view'].text_value = str(item[key]).split('|')[1]
-                        dir_path_view = G[global_key]['view']
+                    if key.endswith('_dir') or key.endswith('_file'):
                         browse_button = gui.Button("...")
-                        def set_text_value(view, path):
-                            view.text_value = path
-                            self.__close_dialog__()
-                        browse_button.set_on_clicked(lambda: self.__show_path_dialog__(f"Select Directory for {label_text.upper()}", gui.FileDialog.OPEN_DIR, dir_path_view.text_value, "", "", lambda path: set_text_value(dir_path_view, path)))
+                        def create_browse_dialog(title=label_text, path_edit=G[global_key]['view']):
+                            def on_done(path):
+                                path_edit.text_value = path
+                                self.__close_dialog__()
+                            dialog_title = f"Select {title.upper()}"
+                            mode = gui.FileDialog.OPEN_DIR if title.endswith('_dir') else gui.FileDialog.OPEN
+                            initial_path = get_abs_path(path_edit.text_value)
+                            self.__show_path_dialog__(dialog_title, mode, initial_path, "", on_done=on_done)
+                        browse_button.set_on_clicked(create_browse_dialog)
                         sub_container.add_child(browse_button)
-                    elif str(item[key]).startswith('file_path'):
-                        G[global_key]['view'].text_value = str(item[key]).split('|')[1]
-                        file_path_view = G[global_key]['view']
-                        browse_button = gui.Button("...")
-                        def set_text_value(view, path):
-                            view.text_value = path
-                            self.__close_dialog__()
-                        browse_button.set_on_clicked(lambda: self.__show_path_dialog__(f"Select File for {label_text.upper()}", gui.FileDialog.OPEN, file_path_view.text_value, "", "", lambda path: set_text_value(file_path_view, path)))
-                        sub_container.add_child(browse_button)
-                    else:
-                        G[global_key]['view'].text_value = str(item[key])
                     sub_container.add_stretch()
                     container.add_child(sub_container)
                 elif type(item[key]) == list:
@@ -225,6 +224,16 @@ class BaseConfiguration:
                     else:
                         # Raise an exception
                         raise Exception("Unsupported type: {}".format(type(item[key])))
+                    
+    def __show_dialog__(self, dialog):
+        if not hasattr(self, 'showing_dialog'): self.showing_dialog = True
+        elif self.showing_dialog: self.__close_dialog__()
[…]

[…]
