### 2024-10-31 803f111
added: argsparse to profiler.py
fix: logging in main.py
