+++ main.py
@@ -163,7 +163,7 @@ class LiGuard:
         else: self.pcd_io = None
         # get the total number of pcd frames
         self.data_dict['total_pcd_frames'] = len(self.pcd_io) if self.pcd_io else 0
-        self.logger.log('total_pcd_frames: {self.data_dict["total_pcd_frames"]}', Logger.DEBUG)
+        self.logger.log(f'total_pcd_frames: {self.data_dict["total_pcd_frames"]}', Logger.DEBUG)
         
         # manage pcd visualization
         if self.pcd_io:
@@ -197,7 +197,7 @@ class LiGuard:
         else: self.img_io = None
         # get the total number of image frames
         self.data_dict['total_img_frames'] = len(self.img_io) if self.img_io else 0
-        self.logger.log('total_img_frames: {self.data_dict["total_img_frames"]}', Logger.DEBUG)
+        self.logger.log(f'total_img_frames: {self.data_dict["total_img_frames"]}', Logger.DEBUG)
         
         # manage image visualization
         if self.img_io:
@@ -221,7 +221,7 @@ class LiGuard:
         else: self.clb_io = None
         # get the total number of calibration frames
         self.data_dict['total_clb_frames'] = len(self.clb_io) if self.clb_io else 0
-        self.logger.log('total_clb_frames: {self.data_dict["total_clb_frames"]}', Logger.DEBUG)
+        self.logger.log(f'total_clb_frames: {self.data_dict["total_clb_frames"]}', Logger.DEBUG)
         
         # manage label reading
         if self.lbl_io != None: self.lbl_io.close()
@@ -235,11 +235,11 @@ class LiGuard:
         else: self.lbl_io = None
         # get the total number of label frames
         self.data_dict['total_lbl_frames'] = len(self.lbl_io) if self.lbl_io else 0
-        self.logger.log('total_lbl_frames: {self.data_dict["total_lbl_frames"]}', Logger.DEBUG)
+        self.logger.log(f'total_lbl_frames: {self.data_dict["total_lbl_frames"]}', Logger.DEBUG)
         
         # get the maximum frame index
         self.data_dict['maximum_frame_index'] = max(self.data_dict['total_pcd_frames'], self.data_dict['total_img_frames'], self.data_dict['total_lbl_frames']) - 1
-        self.logger.log('maximum_frame_index: {self.data_dict["maximum_frame_index"]}', Logger.DEBUG)
+        self.logger.log(f'maximum_frame_index: {self.data_dict["maximum_frame_index"]}', Logger.DEBUG)
         ##########################################
         # processes
         ##########################################
@@ -254,9 +254,9 @@ class LiGuard:
                     process = __import__('algo.pre', fromlist=[proc]).__dict__[proc]
                     self.pre_processes[priority] = process
                 except Exception:
-                    self.logger.log(f'pre_processes creation failed for {proc}:\n{traceback.format_exc()}', Logger.CRITICAL)
+                    self.logger.log(f'pre_processes creation failed for {proc.__name__}:\n{traceback.format_exc()}', Logger.CRITICAL)
         self.pre_processes = [self.pre_processes[priority] for priority in sorted(self.pre_processes.keys())]
-        self.logger.log('enabled pre_processes: {self.pre_processes}', Logger.DEBUG)
+        self.logger.log(f'enabled pre_processes: {self.pre_processes}', Logger.DEBUG)
         
         # lidar processes
         self.lidar_processes = dict()
@@ -269,9 +269,9 @@ class LiGuard:
                     process = __import__('algo.lidar', fromlist=[proc]).__dict__[proc]
                     self.lidar_processes[priority] = process
                 except Exception:
[…]
