### 2024-10-31 b4475f6
chore: log model name based on user selection.
