### 2024-10-31 a05e6d2
refactor: log messages to log enabled processes by their function names
