+++ install_docs.md
@@ -9,3 +9,4 @@ The LiGuard Software Suite requires:
 - Spinnaker-SDK (`https://www.flir.com/support-center/iis/machine-vision/downloads/spinnaker-sdk-download/spinnaker-sdk--download-files/`)
 - DBSCAN (`pip install dbscan`)
 - PointPillars: Fast Encoders for Object Detection from Point Clouds (`https://github.com/zhulf0804/PointPillars`); Clone it Under `algo/nn` Directory; Tested on CUDA-11.7 and PyTorch 1.13.1 on Win11
+- Ultralytics (`pip install ultralytics`)
