+++ bulk_process.py
@@ -1,0 +1,234 @@
+import os
+from gui.config_gui import get_abs_path
+import yaml
+
+from pcd.file_io import FileIO as PCD_File_IO
+from img.file_io import FileIO as IMG_File_IO
+from calib.file_io import FileIO as CLB_File_IO
+from lbl.file_io import FileIO as LBL_File_IO
+
+from gui.logger_gui import Logger
+
+import threading
+from queue import Queue
+
+def reader2queue(reader, size, target_queue):
+    for idx in range(size): target_queue.put(reader[idx])
+    target_queue.put(None)
+
+def queue2dict2queue(source_queue, key1, key2, target_queue):
+    while True:
+        data = source_queue.get()
+        if data is None: break
+        data_dict = {key1: data[0], key2: data[1]}
+        target_queue.put(data_dict)
+        source_queue.task_done()
+    target_queue.put(None)
+
+def dicts2singledict(source_queues, target_queue):
+    while True:
+        data = [source_queues[i].get() for i in range(len(source_queues))]
+        if None in data: break
+        data_dict = dict()
+        for d in data: data_dict.update(d)
+        target_queue.put(data_dict)
+        for i in range(len(source_queues)): source_queues[i].task_done()
+    target_queue.put(None)
+
+def dict2proc2dict(source_queue, cfg, logger, processes, target_queue):
+    while True:
+        data = source_queue.get()
+        if data is None: break
+        for process in processes: process(data, cfg, logger)
+        if target_queue: target_queue.put(data)
+        source_queue.task_done()
+    target_queue.put(None)
+
+def main(args):
+    import yaml
+    with open(args.config) as f:cfg = yaml.safe_load(f)
+    logger = Logger()
+    if cfg['logging']['level'] < Logger.WARNING:
+        print('Logging level is too low. Setting to WARNING to prevent spam.')
+        cfg['logging']['level'] = Logger.WARNING
+    logger.reset(cfg)
+
+    # create dirs
+    if not os.path.exists(cfg['data']['outputs_dir']): os.makedirs(get_abs_path(cfg['data']['outputs_dir']), exist_ok=True)
+    
+    # reader queues
+    pcd_input_queue = Queue(maxsize=args.max_queue_size)
+    img_input_queue = Queue(maxsize=args.max_queue_size)
+    clb_input_queue = Queue(maxsize=args.max_queue_size)
+    lbl_input_queue = Queue(maxsize=args.max_queue_size)
+
+    # readers
+    pcd_reader = PCD_File_IO(cfg) if cfg['data']['lidar']['enabled'] else None
+    img_reader = IMG_File_IO(cfg) if cfg['data']['camera']['enabled'] else None
+    clb_reader = CLB_File_IO(cfg) if cfg['data']['calib']['enabled'] else None
+    lbl_reader = LBL_File_IO(cfg) if cfg['data']['label']['enabled'] else None
+
+    # reader threads
+    if pcd_reader:
+        pcd_io_thread = threading.Thread(target=reader2queue, args=(pcd_reader, len(pcd_reader), pcd_input_queue))
+        pcd_io_thread.start()
+    if img_reader:
+        img_io_thread = threading.Thread(target=reader2queue, args=(img_reader, len(img_reader), img_input_queue))
+        img_io_thread.start()
+    if clb_reader:
+        clb_io_thread = threading.Thread(target=reader2queue, args=(clb_reader, len(clb_reader), clb_input_queue))
+        clb_io_thread.start()
+    if lbl_reader:
+        lbl_io_thread = threading.Thread(target=reader2queue, args=(lbl_reader, len(lbl_reader), lbl_input_queue))
+        lbl_io_thread.start()
+
+    # data dict queues
+    pcd_data_dict_queue = Queue(maxsize=args.max_queue_size)
[…]
