### 2024-11-04 ca351d9
chore: cleanup for packaging
