+++ README.md
@@ -1,339 +1,0 @@
-# LiGuard
-`LiGuard` is a research-purposed, GUI-powered, Python-based framework that allows easy development of dynamic point-cloud (and accompanying image) data processing pipelines by decoupling processes algorithms from framework's source code. Please see [Usage](#usage) section below for more information.
-
-# Installation
-`LiGuard` is tested on Windows 11 with Python 3.10.0. However, it should work on any OS that support following dependencies:
-```
-keyboard==0.13.5
-pyyaml==6.0.1
-open3d==0.18.0
-opencv-python==4.9.0.80
-dbscan==0.0.12
-```
-
-We recommend using a virtual environment (like Conda) to install the `LiGuard`.
-
-First create a new virtual environment and activate it:
-```
-conda create -n liguard python=3.10
-conda activate liguard
-```
-
-Then, clone the repository:
-```
-git clone https://github.com/m-shahbaz-kharal/LiGuard-JOSS.git
-cd LiGuard-JOSS
-```
-
-Lastly, install the dependencies:
-```
-pip install -r requirements.txt
-```
-
-# Documentation
-The documentation for `LiGuard` is available in the [here](https://m-shahbaz-kharal.github.io/LiGuard-JOSS).
-
-# Usage
-`LiGuard` is designed to help researchers build dynamic point-cloud processing pipelines. It acts as an execution engine for I/O, processing, and visualization of both point-cloud and image data. A high-level architecture diagram is given [here](docs/extra/architecture.md). Let's see how to use `LiGuard` with a simple yet intuitive example.
-
-## An Example Pipeline
-In this example, we will implement a simple pipeline to read (a subset of) KITTI dataset, crop the point-clouds, remove any labels that are out of the crop-bound, also remove labels that have less than 1000 points, and save the processed dataset in OpenPCDet annotation standard. Let's break down the pipeline into steps:
-```
-# Pipeline
-1. Read: Read KITTI dataset.
-2. Process:
-- Crop: Crop the point-clouds.
-- Filter Operation 1: Remove out-of-bound labels.
-- Filter Operation 2: Remove labels with less than 1000 points (in the 3d bounding-box).
-3. Post-Process:
-- Save: Cropped point-clouds in npy format.
-- Save: Labels in OpenPCDet annotation standard.
-```
-## Implementing the Pipeline in LiGuard
-In `LiGuard`, a pipeline is implemented by creating a configuration YAML (.yml) file. A default template pipeline configuration (.yml) file `config_template.yml` is provided in the `configs` directory. Please do not delete or modify this file. Instead, duplicate it and rename to your pipeline name; for this example, let's name it `my_kitti_config.yml`.
-
-### Understanding the Pipeline Config File
-This is how it looks:
-```
-# This is a LiGuard pipeline configuration file.
-
-data: # dataset configurations
-    path: 'data' # root directory containing dataset
-    lidar_subdir: 'lidar' # subdirectory containing point clouds
-    camera_subdir: 'camera' # subdirectory containing images
-    label_subdir: 'label' # subdirectory containing labels
-    calib_subdir: 'calib' # subdirectory containing calibration files
-    size: 10 # number of frames to annotate
-
-    lidar:
-        enabled: True # set True to read point clouds from disk
-        pcd_type: '.bin' # can be .bin or .npy
-    camera:
-        enabled: False # set True to read images from disk
-        img_type: '.png' # most image types are supported
-    calib:
-        enabled: True # set True to read calibration files from disk
-        clb_type: 'kitti' # can be kitti or sustechpoints
-    label:
-        enabled: False # set True to read labels from disk
-        lbl_type: 'kitti' # can be kitti, openpcdet, or sustechpoints
-
-sensors: # lidar and camera configurations
[…]
