### 2024-11-04 460c185
added: README.md and LICENSE.txt.
refactor: removed unused dependency of pygetwindow

chore: config.gui now opens in collapsed form to allow easy navigation.

chore: config_template.yml now has default value True for lidar only.

packaging: added setup.cfg and pyproject.toml files.
