+++ pcd/viz.py
@@ -11,8 +11,6 @@ from gui.logger_gui import Logger
 import open3d as o3d
 import numpy as np
 
-import pygetwindow as gw
-
 class PointCloudVisualizer:
     """
     Class for visualizing point clouds and bounding boxes using Open3D.
@@ -37,17 +35,6 @@ class PointCloudVisualizer:
             os.makedirs(self.lidar_save_path, exist_ok=True)
         # reset
         self.reset(cfg, True)
-
-    def __is_focused__(self):
-        """
-        Checks if the visualizer is focused.
-
-        Returns:
-            bool: True if the visualizer is focused, False otherwise.
-        """
-        window = gw.getWindowsWithTitle('PointCloud Feed')
-        if window: return window[0].isActive
-        return False
         
     def reset(self, cfg, reset_bounding_box=False):
         """
