+++ setup.cfg
@@ -1,0 +1,59 @@
+[metadata]
+name = LiGuard
+version = 2.1.0
+author = Muhammad Shahbaz
+author_email = m.shahbaz.kharal@outlook.com
+description = A research-purposed, GUI-powered, Python-based framework that allows easy development of dynamic point-cloud (and accompanying image) data processing pipelines.
+long_description = file: README.md, LICENSE.txt
+long_description_content_type = text/markdown
+url = https://github.com/m-shahbaz-kharal/LiGuard-2.1
+classifiers =
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
+    License :: OSI Approved :: MIT License
+    Operating System :: Windows :: Windows 10
+    Operating System :: Windows :: Windows 11
+    Topic :: Scientific/Engineering :: Image Processing
+    Topic :: Scientific/Engineering :: Visualization
+kewords = point-cloud, image, processing, pipeline, dynamic, GUI, research, development, framework
+
+[options]
+packages = find:
+python_requires = >=3.8, <=3.11
+install_requires =
+    keyboard==0.13.5
+    numpy==1.26.4
+    open3d==0.18.0
+    opencv-python==4.9.0.80
+    pillow==10.3.0
+    pyyaml==6.0.1
+    scipy==1.12.0
+    setuptools==72.2.0
+
+    [options.extras_require]
+    sensor-sdks =
+        ouster-sdk==0.12.0
+        spinnaker-python==4.0.0.116
+    deep-detectors-cuda = 
+        torch==1.13.1+cu117
+        torchvision==0.14.1+cu117
+        ultralytics==8.2.78
+        [options.entry_points]
+        console_scripts =
+            clone_pointpillars = subprocess:run
+        [tool:subprocess]
+        run =
+            git clone https://github.com/zhulf0804/PointPillars algo/nn/PointPillars
+            cd algo/nn/PointPillars/ops
+            python setup.py develop
+    deep-detectors-cpu =
+        torch==1.13.1
+        torchvision==0.14.1
+        ultralytics==8.2.78
+    dev = 
+        pytest==8.1.1
+        sphinx==7.2.6
+        sphinx-rtd-theme==2.0.0
+        
