### 2024-11-04 20b1b58
fixed: setup.cfg to execute PointPillars cloning only if deep-detectors-cuda is set in extras for pip install.
