+++ setup.cfg
@@ -32,28 +32,24 @@ install_requires =
     scipy==1.12.0
     setuptools==72.2.0
 
-    [options.extras_require]
-    sensor-sdks =
-        ouster-sdk==0.12.0
-        spinnaker-python==4.0.0.116
-    deep-detectors-cuda = 
-        torch==1.13.1+cu117
-        torchvision==0.14.1+cu117
-        ultralytics==8.2.78
-        [options.entry_points]
-        console_scripts =
-            clone_pointpillars = subprocess:run
-        [tool:subprocess]
-        run =
-            git clone https://github.com/zhulf0804/PointPillars algo/nn/PointPillars
-            cd algo/nn/PointPillars/ops
-            python setup.py develop
-    deep-detectors-cpu =
-        torch==1.13.1
-        torchvision==0.14.1
-        ultralytics==8.2.78
-    dev = 
-        pytest==8.1.1
-        sphinx==7.2.6
-        sphinx-rtd-theme==2.0.0
+[options.extras_require]
+sensor-sdks =
+    ouster-sdk==0.12.0
+    spinnaker-python==4.0.0.116
+deep-detectors-cuda = 
+    torch==1.13.1+cu117
+    torchvision==0.14.1+cu117
+    ultralytics==8.2.78
+deep-detectors-cpu =
+    torch==1.13.1
+    torchvision==0.14.1
+    ultralytics==8.2.78
+dev = 
+    pytest==8.1.1
+    sphinx==7.2.6
+    sphinx-rtd-theme==2.0.0
+
+[options.entry_points]
+    console_scripts =
+        clone_pointpillars = subprocess:run_if_deep_detectors_cuda
         
