### 2024-11-04 ed9a872
feat: update entry points in setup.cfg and modify main.py to allow direct execution
