+++ setup.cfg
@@ -51,5 +51,6 @@ dev =
 
 [options.entry_points]
     console_scripts =
-        clone_pointpillars = subprocess:run_if_deep_detectors_cuda
+        clone_pointpillars = scripts:run_if_deep_detectors_cuda
+        liguard = main:main
         
