### 2024-11-04 3dbdb57
fix: update operating system classification in setup.cfg to standard given in https://pypi.org/classifiers.
