+++ setup.cfg
@@ -13,8 +13,8 @@ classifiers =
     Programming Language :: Python :: 3.10
     Programming Language :: Python :: 3.11
     License :: OSI Approved :: MIT License
-    Operating System :: Windows :: Windows 10
-    Operating System :: Windows :: Windows 11
+    Operating System :: Microsoft :: Windows :: Windows 10
+    Operating System :: Microsoft :: Windows :: Windows 11
     Topic :: Scientific/Engineering :: Image Processing
     Topic :: Scientific/Engineering :: Visualization
 kewords = point-cloud, image, processing, pipeline, dynamic, GUI, research, development, framework
