### 2024-11-04 457e307
feat: add __init__.py to expose main and bulk_process functions.
refactor: changes in bulk_process.py to have conventional function names.

refactor: converted setup.cfg for setup.py for conditional function calling at install time.
