### 2024-11-04 98c42c9
added: custom install command to handle PointPillars cloning based on deep-detectors-cuda flag.
