### 2024-11-04 3df7685
refactor: renamed main.py to liguard-gui.py and profiler.py to liguard-profiler.py for consistency.
removed: scripts.py as no longer needed.
