### 2024-11-04 43111cc
refactor: install_docs as the dependencies are being solved by setup.py.
refactor: moved setup.py and pyproject.toml outside the src code folder that mainly contains the github repository content close to the main branch.
