+++ install_docs.md
@@ -1,12 +1,0 @@
-# Installation
-The LiGuard Software Suite requires:
-- Keyboard (`pip install keybaord`)
-- pygetwindow (`pip install pygetwindow`)
-- Yaml (`pip install pyyaml`)
-- Open3D (`pip install open3d`)
-- OpenCV (`pip install opencv-python`)
-- Ouster-SDK (`pip install ouster-sdk`)
-- Spinnaker-SDK (`https://www.flir.com/support-center/iis/machine-vision/downloads/spinnaker-sdk-download/spinnaker-sdk--download-files/`)
-- DBSCAN (`pip install dbscan`)
-- PointPillars: Fast Encoders for Object Detection from Point Clouds (`https://github.com/zhulf0804/PointPillars`); Clone it Under `algo/nn` Directory; Tested on CUDA-11.7 and PyTorch 1.13.1 on Win11
-- Ultralytics (`pip install ultralytics`)
