+++ liguard_gui.py
@@ -5,7 +5,7 @@ import open3d.visualization.gui as gui
 
 from gui.config_gui import BaseConfiguration as BaseConfigurationGUI, get_abs_path
 from gui.logger_gui import Logger
-from profiler import Profiler
+from liguard_profiler import Profiler
 
 from pcd.file_io import FileIO as PCD_File_IO
 from pcd.sensor_io import SensorIO as PCD_Sensor_IO
