### 2024-11-04 380b144
fix: update config file path to use absolute path for loading configuration template.
