+++ algo/label.py
@@ -1,7 +1,7 @@
 import inspect
-from gui.config_gui import get_abs_path
-from gui.logger_gui import Logger
-from algo.utils import AlgoType, make_key, get_algo_params
+from src.gui.config_gui import get_abs_path
+from src.gui.logger_gui import Logger
+from src.algo.utils import AlgoType, make_key, get_algo_params
 algo_type = AlgoType.label
 
 import numpy as np
