### 2024-11-05 9076ec0
fix: added a check as last queue is None.
