### 2024-11-05 276d7d8
fix: fixed missing path issue if no ground truth label file paths are available (in case of auto labeling).
