### 2024-11-05 2386d69
docs: update README and add support documentation for sensors and deep object detectors.
