### 2024-11-05 23343c5
fix: update Ouster LIDARs dependency to use the latest version of ouster-sdk.
