### 2024-11-05 4c7152f
docs: update installation instructions and clarify extra dependencies for deep object detectors and sensors.
