### 2024-11-05 254e315
chore: added a sample pipeline for easy quick start.
