### 2024-11-05 3ab655c
fix: improve sensor configuration checks for lidar and camera.
