### 2024-11-05 16a43a2
chore: simplified docs for quick start; more detailed docs are seperated.
added: docs for supported sensors and deep object detectors.

chore: added docs for key bindings of visualizers.
