+++ supported_sensors.md
@@ -1,12 +1,0 @@
-# Supported Sensors
-`LiGuard` supports various LIDAR and RGB sensors. To use these sensors, you need to install official SDKs. The following table lists the supported sensors and their corresponding extra dependencies:
-
-## RGB Sensors
-| Detector | Extra Dependencies |
-|----------|--------------------|
-| Spinnaker Cameras   | `spinnaker-python` |
-
-## LIDAR Sensors
-| Detector | Extra Dependencies |
-|----------|--------------------|
-| Ouster LIDARs | `ouster-sdk` |
