+++ docs/visualizer_key_bindings.md
@@ -1,0 +1,15 @@
+## Visualizer key bindings:
+| Key | Description |
+| --- | --- |
+| `left arrow` | Previous frame |
+| `right arrow` | Next frame |
+| `space` | Play/Pause |
+| `+` | Increase Point Size |
+| `-` | Decrease Point Size |
+| `delete` | Reset Playback |
+| `[` | Jump to frame idx |
+| `mouse wheel` | Zoom in/out |
+| `left mouse button` | Scene Rotate |
+| `alt + left mouse button` | Camera Rotate |
+| `ctrl + left mouse button` | Camera Pan |
+| `shift + left mouse button` | Camera Roll |

[…]
