### 2024-11-05 b8f0e12
chore: removed unused pipeline entries from sample pipeline to keep it simple.
