+++ examples/pipelines/sample_pipeline.yml
@@ -28,23 +28,6 @@ data: # dataset configurations
         lbl_type: 'kitti' # can be kitti, openpcdet, or sustechpoints
     outputs_dir: 'examples/outputs/simple_crop_pipeline' # directory to save outputs
 
-sensors: # lidar and camera configurations
-    lidar: # lidar sensor configurations, at this point only Ouster lidars are supported, support for other lidars is coming soon
-        enabled: False # set True to stream point clouds from sensor, please set False if reading from disk
-        hostname: '192.168.1.2' # sensor ip address or hostname
-        manufacturer: 'Ouster' # sensor manufacturer
-        model: 'OS1-64' # sensor model
-        serial_number: '000000000000' # sensor serial number
-    camera: # camera sensor configurations, at this point only Flir cameras are supported, support for other cameras is coming soon
-        enabled: False # set True to stream point clouds from sensor, please set False if reading from disk
-        hostname: '192.168.1.3' # sensor ip address or hostname
-        manufacturer: 'Flir' # sensor manufacturer
-        model: 'BFS-PGE-16S2C-CS' # sensor model
-        serial_number: '00000000' # sensor serial number
-        camera_matrix: [1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0] # camera matrix (K)
-        distortion_coeffs: [0, 0, 0, 0, 0] # distortion coefficients (D)
-        T_lidar_camera: [[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 1, 0], [0, 0, 0, 1]] # 4x4 transformation matrix from camera to lidar
-
 proc: # liguard processing configurations
     pre:
         remove_nan_inf_allzero_from_pcd: # remove nan points from point cloud
