### 2024-11-06 526f0f1
refactor: replaced keyboard library with pynput for key event handling because keyboard requires sudo access in linux.
