### 2024-11-06 3e2582b
refactor: removed status state from logger GUI as it is working as expected.
