### 2024-11-06 f833c50
refactor: removed log after the application has quitted.
