### 2024-11-07 6e07f5e
major_refactor:
* moved code to liguard module.
* changed starting path to <user-home>/liguard-default-workspace.
* added pypi related files to create a package for distribution.
* added examples directory in liguard-default-workspace.
