### 2024-11-08 bf74889
feat: added "Add New Function" in Configuration Window that directly allows coding of custom functions using default editor in OS.
