+++ liguard/gui/config_gui.py
@@ -142,17 +142,59 @@ class BaseConfiguration:
         """
         with open(cfg_path, 'w') as f:
             yaml.safe_dump(cfg, f, sort_keys=False)
+
+    def open_file_with_default_program(self, file_path):
+        import subprocess, platform
+        system_platform = platform.system()
+        if system_platform == "Windows": os.startfile(file_path) # windows
+        elif system_platform == "Darwin": subprocess.run(['open', file_path]) # macOS
+        else: subprocess.run(['xdg-open', file_path]) # linux
             
     def __update_gui_from_cfg__(self, item, container, parent_keys=[], margin=0.2):
         """
         This function recursively generates the GUI from the configuration dictionary based on the item types.
         """
         G = self.generated_config_gui_dict
+        algo_types = ['pre', 'lidar', 'camera', 'calib', 'label', 'post']
         if type(item) == dict:
             for key in item.keys():
                 label_text = key
                 if type(item[key]) == dict:
                     collapsable_container = gui.CollapsableVert(label_text, self.em * 0.2, gui.Margins(self.em * margin, self.em * 0.2, self.em * 0.2, self.em * 0.2))
+                    if label_text in algo_types:
+                        def add_algo_button_callback(algo_type=label_text):
+                            import shutil
+                            algo_dir = os.path.join(self.last_workspace_dir, 'algo', algo_type)
+                            os.makedirs(algo_dir, exist_ok=True)
+                            def __input_dialog_callback__():
+                                func_name = getattr(self.input_dialog_widget, self.input_dialog_widget_value_variable)
+                                self.__close_dialog__()
+                                
+                                shutil.copy(os.path.join(application_root_dir, 'resources', 'algo_func_template.py'), os.path.join(algo_dir, f'{func_name}.py'))
+                                algo_txt = open(os.path.join(algo_dir, f'{func_name}.py')).read()
+                                algo_txt = algo_txt.replace('AGLO_TYPE', f'AlgoType.{algo_type.lower()}')
+                                algo_txt = algo_txt.replace('FUNCTION_NAME', func_name)
+                                with open(os.path.join(algo_dir, f'{func_name}.py'), 'w') as f: f.write(algo_txt)
+                                
+                                shutil.copy(os.path.join(application_root_dir, 'resources', 'algo_func_template.yml'), os.path.join(algo_dir, f'{func_name}.yml'))
+                                conf_txt = open(os.path.join(algo_dir, f'{func_name}.yml')).read()
+                                conf_txt = conf_txt.replace('FUNCTION_NAME', func_name)
+                                with open(os.path.join(algo_dir, f'{func_name}.yml'), 'w') as f: f.write(conf_txt)
+                                
+                                self.app.run_in_thread(lambda: self.open_file_with_default_program(os.path.join(algo_dir, f'{func_name}.py')))
+                                self.app.run_in_thread(lambda: self.open_file_with_default_program(os.path.join(algo_dir, f'{func_name}.yml')))
+                            self.show_input_dialog(title='Add New Function',
+                                                   query=f'name:',
+                                                   key=f'new_{algo_type}_function_name',
+                                                   ans_base_type=gui.TextEdit,
+                                                   ans_type=None,
+                                                   value_variable='text_value',
+                                                   default_value=f'my_custom_{algo_type}_function',
+                                                   custom_callback=__input_dialog_callback__)
+
+                        add_button = gui.Button(f"Add New {label_text} Function")
[…]
