### 2024-11-08 30447c7
refactor: made a decorator 'algo_func' to wrap all algo functions and add a standard code snippet to be used in all algo functions. refactor: made imports more standard. refactor: added resolve_for_application_root, resolve_for_default_workspace functions. some bug fixes and improvements.
