+++ liguard/img/viz.py
@@ -1,5 +1,5 @@
 import os
-from liguard.gui.config_gui import get_abs_path
+from liguard.gui.config_gui import resolve_for_application_root, resolve_for_default_workspace
 
 import open3d as o3d
 import cv2
@@ -36,7 +36,7 @@ class ImageVisualizer:
         # init
         # create necessary paths
         if cfg['visualization']['camera']['save_images']:
-            self.image_save_path = os.path.join(get_abs_path(cfg['data']['outputs_dir']), 'img_viz')
+            self.image_save_path = os.path.join(resolve_for_default_workspace(cfg['data']['outputs_dir']), 'img_viz')
             os.makedirs(self.image_save_path, exist_ok=True)
         # reset
         self.reset(cfg, True)
