### 2024-11-08 8129803
fix: updated default workspace directory resolution to use 'resolve_for_default_workspace' function.
