### 2024-11-12 b5e3f14
refactor: liguard_gui.py and config_gui.py to allow importing of custom functions from pipeline_dir.
refactor: pipelines are now a directory containing base_config.yml that contains the basic IO, build-in functions, Visualization, etc. configuration.

chore: browse button infront of file or dir path now looks better.

fix: show_input_dialog now shows dialog correctly.

chore: removed save_as button as pipelines are now dirs and not configuration files alone.

feat: custom functions adding works perfectly for algos now.

chore: changed to required_data=[] in decorator in alg_func_template.py.
