+++ liguard/liguard_gui.py
@@ -87,7 +87,7 @@ class LiGuard:
                 if self.pcd_visualizer: self.pcd_visualizer.load_view_status()
             elif key == pynput.keyboard.KeyCode(char='['):
                 self.is_playing = False
-                self.config.show_input_dialog('Enter the frame index:', f'0-{self.data_dict["maximum_frame_index"]}', 'jump_to_frame')
+                self.config.show_input_dialog('Enter the frame index:', f'Jump to Frame (0-{self.data_dict["maximum_frame_index"]})', 'jump_to_frame')
                     
     def reset(self, cfg):
         """
@@ -241,13 +241,14 @@ class LiGuard:
         ##########################################
         # pre processes
         self.pre_processes = dict()
+        built_in_pre_modules = __import__('liguard.algo.pre', fromlist=['*']).__dict__
         for proc in cfg['proc']['pre']:
-            if 'enabled' not in cfg['proc']['pre'][proc]: cfg['proc']['pre'][proc]['enabled'] = 'True'
             enabled = cfg['proc']['pre'][proc]['enabled']
             if enabled:
                 try:
                     priority = cfg['proc']['pre'][proc]['priority']
-                    process = __import__('liguard.algo.pre', fromlist=[proc]).__dict__[proc]
+                    if proc in built_in_pre_modules: process = built_in_pre_modules[proc]
+                    else: process = __import__(proc, fromlist=['*']).__dict__[proc]
                     self.pre_processes[priority] = process
                 except Exception:
                     self.logger.log(f'pre_processes creation failed for {proc.__name__}:\n{traceback.format_exc()}', Logger.CRITICAL)
@@ -256,13 +257,14 @@ class LiGuard:
         
         # lidar processes
         self.lidar_processes = dict()
+        built_in_lidar_modules = __import__('liguard.algo.lidar', fromlist=['*']).__dict__
         for proc in cfg['proc']['lidar']:
-            if 'enabled' not in cfg['proc']['lidar'][proc]: cfg['proc']['lidar'][proc]['enabled'] = 'True'
             enabled = cfg['proc']['lidar'][proc]['enabled']
             if enabled:
                 try:
                     priority = cfg['proc']['lidar'][proc]['priority']
-                    process = __import__('liguard.algo.lidar', fromlist=[proc]).__dict__[proc]
+                    if proc in built_in_lidar_modules: process = built_in_lidar_modules[proc]
+                    else: process = __import__(proc, fromlist=['*']).__dict__[proc]
                     self.lidar_processes[priority] = process
                 except Exception:
                     self.logger.log(f'lidar_processes creation failed for {proc.__name__}:\n{traceback.format_exc()}', Logger.CRITICAL)
@@ -271,13 +273,14 @@ class LiGuard:
         
         # camera processes
         self.camera_processes = dict()
+        built_in_camera_modules = __import__('liguard.algo.camera', fromlist=['*']).__dict__
         for proc in cfg['proc']['camera']:
-            if 'enabled' not in cfg['proc']['camera'][proc]: cfg['proc']['camera'][proc]['enabled'] = 'True'
             enabled = cfg['proc']['camera'][proc]['enabled']
             if enabled:
                 try:
                     priority = cfg['proc']['camera'][proc]['priority']
-                    process = __import__('liguard.algo.camera', fromlist=[proc]).__dict__[proc]
+                    if proc in built_in_camera_modules: process = built_in_camera_modules[proc]
+                    else: process = __import__(proc, fromlist=['*']).__dict__[proc]
                     self.camera_processes[priority] = process
                 except Exception:
                     self.logger.log(f'camera_processes creation failed for {proc.__name__}:\n{traceback.format_exc()}', Logger.CRITICAL)
[…]
