+++ liguard/gui/config_gui.py
@@ -70,21 +70,21 @@ class BaseConfiguration:
         button_container = gui.Horiz(self.em * 0.2, gui.Margins(self.em * 0.2, self.em * 0.2, self.em * 0.2, self.em * 0.2))
         self.new_config_button = gui.Button("New")
         self.open_config_button = gui.Button("Open")
+        self.reload_config_button = gui.Button("Reload")
         self.save_config_button = gui.Button("Save")
-        self.save_as_config_button = gui.Button("Save As")
         self.apply_config_button = gui.Button("Apply")
         # -- set callbacks
         self.new_config_button.set_on_clicked(self.__new_config__)
         self.open_config_button.set_on_clicked(self.__open_config__)
+        self.reload_config_button.set_on_clicked(self.__reload_config__)
         self.save_config_button.set_on_clicked(self.__save_config__)
-        self.save_as_config_button.set_on_clicked(self.__save_as_config__)
         self.apply_config_button.set_on_clicked(self.__apply_config__)
         # -- add to button container
         button_container.add_stretch()
         button_container.add_child(self.new_config_button)
         button_container.add_child(self.open_config_button)
+        button_container.add_child(self.reload_config_button)
         button_container.add_child(self.save_config_button)
-        button_container.add_child(self.save_as_config_button)
         button_container.add_child(self.apply_config_button)
         
         # add to top container
@@ -133,6 +133,21 @@ class BaseConfiguration:
             cfg = yaml.safe_load(f)
         return cfg
     
+    def load_pipeline_algos(self):
+        custom_algo_dir = os.path.join(self.last_pipeline_dir, 'algo')
+        self.custom_algos_cfg = dict()
+        if os.path.exists(custom_algo_dir):
+            for algo_type_path in os.listdir(custom_algo_dir):
+                algo_type_path = os.path.join(custom_algo_dir, algo_type_path)
+                if not os.path.isdir(algo_type_path): continue
+                if algo_type_path not in sys.path: sys.path.append(algo_type_path)
+                algo_type = os.path.basename(algo_type_path)
+                if algo_type not in self.custom_algos_cfg: self.custom_algos_cfg[algo_type] = dict()
+                for algo_file_name in os.listdir(algo_type_path):
+                    if not algo_file_name.endswith('.yml'): continue
+                    self.custom_algos_cfg[algo_type].update(self.load_config(os.path.join(algo_type_path, algo_file_name)))
+                self.cfg['proc'][algo_type].update(self.custom_algos_cfg[algo_type])
+    
     def save_config(self, cfg, cfg_path):
         """
         Save the configuration to a YAML file.
@@ -144,6 +159,15 @@ class BaseConfiguration:
         Returns:
             None
         """
+        # remove custom algo configs from the main config
+        if hasattr(self, 'custom_algos_cfg'):
+            for algo_type in self.custom_algos_cfg:
+                for algo_name in self.custom_algos_cfg[algo_type]:
+                    if algo_name in self.cfg['proc'][algo_type]: self.cfg['proc'][algo_type].pop(algo_name)
+                    algo_dir = os.path.join(self.last_pipeline_dir, 'algo', algo_type)
+                    with open(os.path.join(algo_dir, f'{algo_name}.yml'), 'w') as f:
+                        algo_config = {algo_name: self.custom_algos_cfg[algo_type][algo_name]}
+                        yaml.safe_dump(algo_config, f, sort_keys=False)
         with open(cfg_path, 'w') as f:
             yaml.safe_dump(cfg, f, sort_keys=False)
 
@@ -153,7 +177,7 @@ class BaseConfiguration:
         if system_platform == "Windows": os.startfile(file_path) # windows
         elif system_platform == "Darwin": subprocess.run(['open', file_path]) # macOS
[…]
