+++ liguard/img/file_io.py
@@ -38,7 +38,9 @@ class FileIO:
 
     def __init__(self, cfg: dict):
         self.cfg = cfg
-        self.img_dir = os.path.join(resolve_for_default_workspace(cfg['data']['main_dir']), cfg['data']['camera_subdir'])
+        main_dir = cfg['data']['main_dir']
+        if not os.path.isabs(main_dir): main_dir = os.path.join(self.cfg['data']['pipeline_dir'], main_dir)
+        self.img_dir = os.path.join(main_dir, cfg['data']['camera_subdir'])
         self.img_type = cfg['data']['camera']['img_type']
         self.img_start_idx = cfg['data']['start']['camera']
         self.global_zero = cfg['data']['start']['global_zero']
