### 2024-11-12 b5dbd49
chore: updated liguard_cmd.py to reflect changes in liguard_gui.py.
chore: added tqdm progress bar to liguard_cmd.

fix: initial path bug in browse buttons in configuration window.

fix: handler_ips300+ had a data_type type.
