+++ liguard/liguard_cmd.py
@@ -1,4 +1,5 @@
 import os
+import sys
 from liguard.gui.config_gui import resolve_for_application_root, resolve_for_default_workspace
 import yaml
 
@@ -9,44 +10,83 @@ from liguard.lbl.file_io import FileIO as LBL_File_IO
 
 from liguard.gui.logger_gui import Logger
 
-import threading
+import time
+import signal
+from threading import Thread, Event
 from queue import Queue
 
+from tqdm import tqdm
+
+stop_event = Event()
+
 def reader2queue(reader, size, target_queue):
-    for idx in range(size): target_queue.put(reader[idx])
-    target_queue.put(None)
+    for idx in range(size):
+        if stop_event.is_set():
+            target_queue.put(None)
+            break
+        target_queue.put(reader[idx])
 
 def queue2dict2queue(source_queue, key1, key2, target_queue):
     while True:
         data = source_queue.get()
-        if data is None: break
+        if data is None:
+            target_queue.put(None)
+            source_queue.task_done()
+            break
         data_dict = {key1: data[0], key2: data[1]}
         target_queue.put(data_dict)
         source_queue.task_done()
-    target_queue.put(None)
-
-def dicts2singledict(source_queues, target_queue):
+    
+def dicts2singledict(source_queues, target_queue, p_bar):
     while True:
         data = [source_queues[i].get() for i in range(len(source_queues))]
-        if None in data: break
+        if None in data:
+            target_queue.put(None)
+            for i in range(len(source_queues)): source_queues[i].task_done()
+            break
         data_dict = dict()
         for d in data: data_dict.update(d)
         target_queue.put(data_dict)
         for i in range(len(source_queues)): source_queues[i].task_done()
-    target_queue.put(None)
+        p_bar.update(1)
 
-def dict2proc2dict(source_queue, cfg, logger, processes, target_queue):
+def dict2proc2dict(source_queue, cfg, logger, processes, target_queue, p_bar):
     while True:
         data = source_queue.get()
-        if data is None: break
+        if data is None:
+            if target_queue: target_queue.put(None)
+            source_queue.task_done()
+            break
         for process in processes: process(data, cfg, logger)
         if target_queue: target_queue.put(data)
         source_queue.task_done()
-    if target_queue: target_queue.put(None)
+        p_bar.update(1)
+
+def signal_handler(sig, frame):
+    print("\nCtrl + C detected! Stopping ...")
+    stop_event.set()  # Signal all threads to stop
 
 def bulk_process(args):
-    import yaml
-    with open(args.config) as f:cfg = yaml.safe_load(f)
+    pipeline_dir = args.pipeline_dir
+    base_cfg_path = os.path.join(pipeline_dir, 'base_config.yml')
+    with open(base_cfg_path) as f:cfg = yaml.safe_load(f)
+    cfg['data']['pipeline_dir'] = pipeline_dir
+    
+    custom_algo_dir = os.path.join(pipeline_dir, 'algo')
+    custom_algos_cfg = dict()
+    if os.path.exists(custom_algo_dir):
+        for algo_type_path in os.listdir(custom_algo_dir):
+            algo_type_path = os.path.join(custom_algo_dir, algo_type_path)
+            if not os.path.isdir(algo_type_path): continue
+            if algo_type_path not in sys.path: sys.path.append(algo_type_path)
+            algo_type = os.path.basename(algo_type_path)
+            if algo_type not in custom_algos_cfg: custom_algos_cfg[algo_type] = dict()
[…]
