### 2024-11-15 708dd2e
docs: added detailed documentation for liguard and its submodules.
