+++ docs/supported_sensors.md
@@ -1,12 +1,12 @@
-# Supported Sensors
-`LiGuard` supports various LIDAR and RGB sensors. To use these sensors, you need to install official their official SDKs and dependencies. The following table lists the supported sensors and their corresponding extra dependencies:
+# Supported Sensors (Experimental)
+`LiGuard` also supports various LIDAR and RGB sensors as data sources for streaming data experiments. To use these sensors, you need to install their official SDKs and dependencies. The following table lists the supported sensors and their corresponding extra dependencies:
 
 ## RGB Sensors
 | Detector | Extra Dependencies | How to install |
 |----------|--------------------| ---------------|
 | Spinnaker Cameras   | `spinnaker-python` | go to [this link](https://www.teledynevisionsolutions.com/support/support-center/software-firmware-downloads/iis/spinnaker-sdk-download/spinnaker-sdk--download-files/?pn=Spinnaker+SDK&vn=Spinnaker+SDK), and download and install the appropriate version of the Spinnaker SDK for your environment. |
 
-To use the RGB sensor in your data processing pipeline, add the following entry under `sensors` in your pipeline configuration.
+To use the RGB sensor in your data processing pipeline, add and entry `sensors` in your pipeline's `base_config.yml` file and then add the following entry under that `sensors` entry.
 ```yaml
 camera: # camera sensor configurations, at this point only Flir cameras are supported, support for other cameras is coming soon
     enabled: True # set True to stream point clouds from sensor, please set False `data/camera` is enabled.
@@ -24,7 +24,7 @@ camera: # camera sensor configurations, at this point only Flir cameras are supp
 |----------|--------------------| ---------------|
 | Ouster LIDARs | `ouster-sdk` | `pip install ouster-sdk` |
 
-To use the LIDAR sensor in your data processing pipeline, add the following entry under `sensors` in your pipeline configuration.
+To use the LIDAR sensor in your data processing pipeline, add and entry `sensors` in your pipeline's `base_config.yml` file and then add the following entry under that `sensors` entry.
 ```yaml
 lidar: # lidar sensor configurations, at this point only Ouster lidars are supported, support for other lidars is coming soon
     enabled: True # set True to stream point clouds from sensor, please set False `data/lidar` is enabled.

[…]
