### 2024-11-15 1f9488a
docs: updated README.md with logo and navigation links; added algo_auxiliary.md for utility functions.
