### 2024-11-15 9bfcd51
fix: bugs in doc compilation, still some warning but overall the docs are compiles fine now.
