+++ liguard/gui/__init__.py
@@ -2,6 +2,8 @@
 gui Package
 ===========
 The gui package contains the modules responsible for creating the graphical user interface for the application. The modules under this package should not be modified unless for contribution to framework application logic. The package comprises of the following modules:
-- 1. `**config_gui.py**`: This module contains the class `BaseConfiguration` that is responsible for creating the configuration window for the application. The configuration window allows users to set the parameters for the algorithms and the pipeline execution order.
-- 2. `**logger_gui.py**`: This module contains the class `Logger` that is responsible for creating the logger window for the application. The logger window displays the logs generated during the execution of the algorithms in the pipeline. It also writes the logs to a log file(s) in logs directory specified in the configuration.
+
+1. **config_gui.py**: This module contains the class `BaseConfiguration` that is responsible for creating the configuration window for the application. The configuration window allows users to set the parameters for the algorithms and the pipeline execution order.
+2. **logger_gui.py**: This module contains the class `Logger` that is responsible for creating the logger window for the application. The logger window displays the logs generated during the execution of the algorithms in the pipeline. It also writes the logs to a log file(s) in logs directory specified in the configuration.
+
 """
