+++ liguard/algo/nn/__init__.py
@@ -2,6 +2,8 @@
 algo.nn package
 ================
 This package is experimental and contains Github repositories of algorithms based on neural networks.
+
 .. include:: ../../../docs/supported_deep_detectors.md
    :literal:
+   
 """
