### 2024-11-15 051eaf4
Create build_and_deploy_sphinx_docs.yml
