### 2024-11-15 e122308
a bug fix in build_and_deploy_sphinx_docs.yml
