### 2024-11-15 ccff22c
docs: added sphinx docs for github pages.
