+++ docs/sphinx/liguard_pipelines.md
@@ -1,0 +1,282 @@
+# An In-Depth Look into LiGuard Pipelines
+In `LiGuard`, a `pipeline` is a sequence of components that process LIDAR and/or image data. In other words, it defines an algorithm that performs a sequence of steps on given LIDAR and/or image data. Each component in the pipeline performs a specific task such as reading data, preprocessing, data analysis, or postprocessing. The components are connected in a sequence, and the data flows from one component to the next based on the defined priority of the components. The pipeline is executed in real-time, and the results are visualized in the GUI and/or stored.
+
+## Pipeline Directory Structure
+A `LiGuard` pipeline is organized in a structured directory format. The pipeline directory structure is automatically managed by `LiGuard`, minimizing the need for manual modifications. However, understanding this structure is beneficial, particularly when sharing pipelines with others. Below, we provide a detailed overview to enhance your comprehension of the directory layout and its components.
+
+The pipeline directory contains the following files and directories:
+```
+Pipeline Directory
+--------------------
+|-- base_config.yml
+|-- data_handler
+    |-- calib
+        |-- handler_custom_dataset1.py
+        |-- handler_custom_dataset2.py
+        |-- ...
+    |-- label
+        |-- handler_custom_dataset1.py
+        |-- handler_custom_dataset2.py
+        |-- ...
+|-- algo
+    |-- pre
+        |-- component1.yml
+        |-- component1.py
+        |-- component2.yml
+        |-- component2.py
+        |-- ...
+    |-- lidar
+        |-- component3.yml
+        |-- component3.py
+    |-- camera
+        ...
+    |-- calib
+        ...
+    |-- label
+        ...
+    |-- post
+        ...
+|-- outputs (default, can be changed from GUI)
+|-- logs (default, can be changed from GUI)
+```
+
+- `base_config.yaml`: A YAML file that contains the basic configuration of the pipeline. It includes the input data sources and their types, output directories, built-in data processing functions and their parameters, visualizer settings, log directory, and other general settings.
+- `data_handler`: A directory that contains subdirectories for different types of data handlers. Each subdirectory contains Python files for handling specific types of data. The subdirectories are:
+    - `calib`: Calibration data handlers for different datasets.
+    - `label`: Label data handlers for different datasets.
+- `algo`: A directory that contains subdirectories for different types of components. Each subdirectory contains YAML and Python files for the components. The subdirectories are:
+    - `pre`: Preprocessing components that are executed before the main processing.
+    - `lidar`: LIDAR data processing components.
+    - `camera`: Camera data processing components.
+    - `calib`: Calibration data processing components.
+    - `label`: Label data processing components.
+    - `post`: Postprocessing components that are executed after the main processing.
+- `outputs`: A directory where the output data generated by the pipeline is stored. The default output directory can be changed from the GUI.
+- `logs`: A directory where the log files generated during the execution of the pipeline are stored. The default log directory can be changed from the GUI.
+
+## Standard Functions Structure
+### Data Handlers
+Data handlers are responsible for reading calibration and label data. Each data handler is defined by a Python script that implements the data reading and processing functions for a specific dataset.
+#### Calibration Data Handler
+The calibration data handler script should have a variable `calib_file_extension` that specifies the extension of the calibration files in the `calib_subdir` directory. The `Handler` function in the script reads the calibration file and generates a dictionary containing the calibration data in KITTI format. The function signature should be as follows:
+```python
+import os
+import numpy as np
+
+calib_file_extension = '.txt' # this tells the extension of the calibration files in directory given by base_config['data']['calib_subdir']
+
+
+def Handler(calib_path: str): # don't change the function signature
+    """
[…]
