+++ docs/sphinx/liguard.pcd.rst
@@ -1,0 +1,53 @@
+liguard.pcd package
+===================
+
+Submodules
+----------
+
+liguard.pcd.file\_io module
+---------------------------
+
+.. automodule:: liguard.pcd.file_io
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+liguard.pcd.handler\_ouster\_os164 module
+-----------------------------------------
+
+.. automodule:: liguard.pcd.handler_ouster_os164
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+liguard.pcd.sensor\_io module
+-----------------------------
+
+.. automodule:: liguard.pcd.sensor_io
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+liguard.pcd.utils module
+------------------------
+
+.. automodule:: liguard.pcd.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+liguard.pcd.viz module
+----------------------
+
+.. automodule:: liguard.pcd.viz
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: liguard.pcd
+   :members:
+   :undoc-members:
+   :show-inheritance:
