+++ docs/sphinx/liguard.img.rst
@@ -1,0 +1,45 @@
+liguard.img package
+===================
+
+Submodules
+----------
+
+liguard.img.file\_io module
+---------------------------
+
+.. automodule:: liguard.img.file_io
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+liguard.img.handler\_flir\_bfspge16s2ccs module
+-----------------------------------------------
+
+.. automodule:: liguard.img.handler_flir_bfspge16s2ccs
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+liguard.img.sensor\_io module
+-----------------------------
+
+.. automodule:: liguard.img.sensor_io
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+liguard.img.viz module
+----------------------
+
+.. automodule:: liguard.img.viz
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: liguard.img
+   :members:
+   :undoc-members:
+   :show-inheritance:
