+++ docs/sphinx/modules.rst
@@ -1,0 +1,7 @@
+liguard
+=======
+
+.. toctree::
+   :maxdepth: 4
+
+   liguard

[…]
