### 2024-11-15 e1263c2
created build_and_deploy_sphinx_docs.yml in add_easy_algo.
