+++ .github/workflows/run_tests.yml
@@ -1,0 +1,36 @@
+name: Cross-Platform Pytest on Multiple Python Versions
+
+on:
+  push:
+    branches:
+      - dev_2.x
+  pull_request:
+    branches:
+      - dev_2.x
+
+jobs:
+  test:
+    # Use a matrix strategy to run tests on multiple OS and Python versions
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, windows-latest]  # Define the OS platforms to test
+        python-version: [3.8, 3.9, 3.10]    # Define the Python versions to test
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v3
+
+    - name: Set up Python
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{ matrix.python-version }}
+
+    - name: Upgrade pip
+      run: python -m pip install --upgrade pip
+
+    - name: Install dependencies
+      run: pip install -r tests/requirements.txt  # Adjust if you have a requirements file
+
+    - name: Run Pytest
+      run: pytest tests  # Adjust to the location of your tests
