### 2024-11-18 44d5785
chore: fixed typos in liguard_pipelines.md. chore: added badges for release, docs, test, and license.
