+++ docs/liguard_pipelines.md
@@ -247,36 +247,34 @@ Here, `FUNCTION_NAME` is the name of the component, `AGLO_TYPE` is the type of t
 
 ## Create Custom Data Handlers and Algorithm Components
 - ### Data Handlers
-    - Calibration Data Handler
-The easiest way to add custom data handlers to `LiGuard` is as follows:
-- Under `data->calib` in `Configuration` window, click `Create Custom Calib Data Handler` button.
-- Enter the name of the handler and click `OK`. This will create the necessary directory stucture and files for the custom data handler. It'll also open the Python script in the default editor in your OS.
-- Implement the data reading in the Python script by following the structure described above or in the comments in the script.
-- Save the file and click `Reload` button in the `Configuration` window to load the new data handler.
-- Under `data->calib->clb_type` in `Configuration` window, type the name of the custom data handler. Click `Apply` to use the custom data handler.
+    - Calibration Data Handler: The easiest way to add custom data handlers to `LiGuard` is as follows:
+        - Under `data->calib` in `Configuration` window, click `Create Custom Calib Data Handler` button.
+        - Enter the name of the handler and click `OK`. This will create the necessary directory stucture and files for the custom data handler. It'll also open the Python script in the default editor in your OS.
+        - Implement the data reading in the Python script by following the structure described above or in the comments in the script.
+        - Save the file and click `Reload` button in the `Configuration` window to load the new data handler.
+        - Under `data->calib->clb_type` in `Configuration` window, type the name of the custom data handler. Click `Apply` to use the custom data handler.
 
-You can also do this manually by creating the Python script according to the structure described above. Make sure to follow the naming convention and the structure of the file to ensure that `LiGuard` can load and execute the data handler correctly.
+        - Note: You can also do this manually by creating the Python script according to the structure described above. Make sure to follow the naming convention and the structure of the file to ensure that `LiGuard` can load and execute the data handler correctly.
 
-    - Label Data Handler
-The easiest way to add custom data handlers to `LiGuard` is as follows:
-- Under `data->label` in `Configuration` window, click `Create Custom Label Data Handler` button.
-- Enter the name of the handler and click `OK`. This will create the necessary directory stucture and files for the custom data handler. It'll also open the Python script in the default editor in your OS.
-- Implement the data reading in the Python script by following the structure described above or in the comments in the script.
-- Save the file and click `Reload` button in the `Configuration` window to load the new data handler.
-- Under `data->label->lbl_type` in `Configuration` window, type the name of the custom data handler. Click `Apply` to use the custom data handler.
+    - Label Data Handler: The easiest way to add custom data handlers to `LiGuard` is as follows:
+        - Under `data->label` in `Configuration` window, click `Create Custom Label Data Handler` button.
+        - Enter the name of the handler and click `OK`. This will create the necessary directory stucture and files for the custom data handler. It'll also open the Python script in the default editor in your OS.
+        - Implement the data reading in the Python script by following the structure described above or in the comments in the script.
+        - Save the file and click `Reload` button in the `Configuration` window to load the new data handler.
+        - Under `data->label->lbl_type` in `Configuration` window, type the name of the custom data handler. Click `Apply` to use the custom data handler.
 
-You can also do this manually by creating the Python script according to the structure described above. Make sure to follow the naming convention and the structure of the file to ensure that `LiGuard` can load and execute the data handler correctly.
+        - Note: You can also do this manually by creating the Python script according to the structure described above. Make sure to follow the naming convention and the structure of the file to ensure that `LiGuard` can load and execute the data handler correctly.
 
 - ### Algorithm Components
-The easiest way to add custom components to `LiGuard` is as follows:
-- Under `proc->lidar` in `Configuration` window, click `Create Custom Function` button.
[…]
