### 2024-11-19 61499d1
paper: added joss paper.md and related files.
