+++ paper/paper.md
@@ -1,0 +1,55 @@
+---
+title: 'LiGuard: Interactively and Rapidly Create Point-Cloud and Image Processing Pipelines'
+
+tags:
+  - Python
+  - GUI
+  - Open3D
+  - OpenCV
+  - lidar
+  - Point-Cloud Processing
+  - Image Processing
+  - Visualization
+authors:
+  - name: Muhammad Shahbaz
+    orcid: 0009-0003-6377-8274
+    affiliation: 1
+  - name: Shaurya Agarwal
+    orcid: 0000-0001-7754-6341
+    affiliation: 1
+affiliations:
+  - name: University of Central Florida, USA
+    index: 1
+date: 16 April 2024
+bibliography: paper.bib
+
+---
+# Summary
+There is a growing interest in the development of lidar-based applications in domains like robotics, autonomous driving, traffic monitoring, infrastructure inspection, and many areas of envrionmental surveying and mapping. In many cases, these applications rely on solutions that are not end-to-end but rather a sequential combination (a pipeline) of individual data processing functions including data readers, pre-processors, application specific algorithms/models, post-processors, and visualizers, etc. Creating such multi-step solutions often require extensive revisiting of the proposed processing pipeline, including enabling/disabling certain functions (for ablation studies), fine-tunning parameters, and even completely replacing parts of the solution. Such revisions pose challenges on the speed of research and even on the reuse of it due to interdependencies among functions. To address this issue, `LiGuard` is presented. `LiGuard` is a software framework that provides an easy-to-use graphical user interface (GUI) that helps researchers rapidly build and test their point cloud (and corresponding image) processing pipelines. It allows them to dynamically add/remove/reorder functions, adjust parameters of those functions, and visualize results, in a live and interactive manner compared to other packages where such modifications often require re-executing the program script. Moreover, because it creates all the meta files in strucutured directories, it allows easy sharing of the created pipelines or the individual functions used in those pipelines.
+
+`LiGuard` features, out of the box, data reading for many common dataset formats including support for reading calibration and label data. Moreover, it provides (an increasing list of) commonly used algorithm components ranging from basic data preprocessors to advanced object detection and tracking algorithms. Additionally, it establishes a straightforward standard for adding custom functions/algorithms, allowing users to integrate unique components into their pipelines. For latest updates on supported features, please see [LiGuard’s documentation](https://m-shahbaz-kharal.github.io/LiGuard-2.x/).
+
+Note: `LiGuard` is built on many standard python libraries and packages for data processing and visualization including Open3D [@Zhou2018], OpenCV [@opencv_library], Numpy [@harris2020array], and SciPy [@2020SciPy]. `LiGuard` is designed to be used interactively and is therefore not ideal for processing very large datasets. We provide a utility [`liguard_cmd`](https://github.com/m-shahbaz-kharal/LiGuard-2.x/blob/dev_2.x/liguard/liguard_cmd.py) that processes data slightly faster by removing GUI and implementing multithreading but since the data is still processed sequentially the output generation is as fast as the slowest step in the pipeline.
+
+# Statement of Need
[…]
