+++ paper/paper.bib
@@ -1,0 +1,175 @@
+@article{Goelles2021,
+  doi = {10.21105/joss.03471},
+  url = {https://doi.org/10.21105/joss.03471},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6}, number = {65}, pages = {3471},
+  author = {Thomas Goelles and Birgit Schlager and Stefan Muckenhuber and Sarah Haas and Tobias Hammer},
+  title = {`pointcloudset`: Efficient Analysis of Large Datasets of Point Clouds Recorded Over Time},
+  journal = {Journal of Open Source Software}
+}
+
+@article{Lamprecht2019,
+  doi = {10.21105/joss.00990},
+  url = {https://doi.org/10.21105/joss.00990},
+  year = {2019},
+  publisher = {The Open Journal},
+  volume = {4}, number = {36}, pages = {990},
+  author = {Sebastian Lamprecht},
+  title = {Pyoints: A Python package for point cloud, voxel and raster processing.},
+  journal = {Journal of Open Source Software}
+}
+
+@InProceedings{Rusu_ICRA2011_PCL,
+  author    = {Radu Bogdan Rusu and Steve Cousins},
+  title     = {{3D is here: Point Cloud Library (PCL)}},
+  booktitle = {{IEEE International Conference on Robotics and Automation (ICRA)}},
+  month     = {May 9-13},
+  year      = {2011},
+  address   = {Shanghai, China},
+  publisher = {IEEE}
+}
+
+@article{Zhou2018,
+    author    = {Qian-Yi Zhou and Jaesik Park and Vladlen Koltun},
+    title     = {{Open3D}: {A} Modern Library for {3D} Data Processing},
+    journal   = {arXiv:1801.09847},
+    year      = {2018},
+}
+
+@article{brown2012laspy,
+  title={Laspy},
+  author={Brown, Grant and Montaigu, T},
+  journal={URL: https://laspy.readthedocs.io/en/latest},
+  year={2012}
+}
+
+@Manual{pyntcloud,
+  title = {pyntcloud},
+  author = {{Pyntcloud Development Team}},
+  year = {2021},
+  url = {https://pyntcloud.readthedocs.io/en/latest/introduction.html},
+}
+
+@Article{Hunter:2007,
+  Author    = {Hunter, J. D.},
+  Title     = {Matplotlib: A 2D graphics environment},
+  Journal   = {Computing in Science \& Engineering},
+  Volume    = {9},
+  Number    = {3},
+  Pages     = {90--95},
+  abstract  = {Matplotlib is a 2D graphics package used for Python for
+  application development, interactive scripting, and publication-quality
+  image generation across user interfaces and operating systems.},
+  publisher = {IEEE COMPUTER SOC},
+  doi       = {10.1109/MCSE.2007.55},
+  year      = 2007
+}
+
+@article{opencv_library,
+    author = {Bradski, G.},
+    citeulike-article-id = {2236121},
+    journal = {Dr. Dobb's Journal of Software Tools},
+    keywords = {bibtex-import},
+    posted-at = {2008-01-15 19:21:54},
+    priority = {4},
+    title = {{The OpenCV Library}},
+    year = {2000}
+}
+
+@article{geiger2013vision,
+  title={Vision meets robotics: The kitti dataset},
+  author={Geiger, Andreas and Lenz, Philip and Stiller, Christoph and Urtasun, Raquel},
+  journal={The International Journal of Robotics Research},
+  volume={32},
+  number={11},
+  pages={1231--1237},
+  year={2013},
+  publisher={Sage Publications Sage UK: London, England}
+}
+
[…]
