### 2024-11-19 e877acf
chore: added some missing references.
