+++ paper/paper.bib
@@ -173,3 +173,22 @@ title      = {Array programming with {NumPy}},
     year = "2022"
 }
 
+@software{ros,
+  author = {{Stanford Artificial Intelligence Laboratory et al.}},
+  title = {Robotic Operating System},
+  url = {https://www.ros.org},
+  version = {ROS Melodic Morenia},
+  date = {2018-05-23},
+}
+
+@misc{pdal_contributors_2024_2616780,
+  author = {PDAL Contributors},
+  title = {PDAL Point Data Abstraction Library},
+  month = aug,
+  year = 2024,
+  doi = {10.5281/zenodo.10884408},
+  url = {
+    https://doi.org/10.5281/zenodo.10884408
+  }
+}
+
