### 2024-11-19 6c8f035
Create compile_paper.yml
