### 2024-11-19 577543b
paper: updated references and improved clarity in the paper.md.
