+++ paper/paper.md
@@ -32,19 +32,16 @@ There is a growing interest in the development of lidar-based applications in do
 Note: `LiGuard` is built on many standard python libraries and packages for data processing and visualization including Open3D [@Zhou2018], OpenCV [@opencv_library], Numpy [@harris2020array], and SciPy [@2020SciPy]. `LiGuard` is designed to be used interactively and is therefore not ideal for processing very large datasets. We provide a utility [`liguard_cmd`](https://github.com/m-shahbaz-kharal/LiGuard-2.x/blob/dev_2.x/liguard/liguard_cmd.py) that processes data slightly faster by removing GUI and implementing multithreading but since the data is still processed sequentially the output generation is as fast as the slowest step in the pipeline.
 
 # Statement of Need
-Recent advancements in LiDAR technology have significantly broadened its applicability across various domains, making it a potential sensor for precise measurement. Innovations such as solid-state LiDAR and nanophotonics-based devices have enhanced the performance, reliability, and cost-effectiveness of these sensors, enabling their use in diverse fields including autonomous driving, environmental monitoring, and industrial automation [@10.1002/lpor.202100511]. Additionally, the integration of advanced optics and beam steering technologies has improved the spatial resolution and operational efficiency of LiDAR systems, facilitating their deployment in complex environments for applications ranging from agriculture to urban planning ([@10.3390/s22239293]; [@10.3390/mi13060894]). As the lidar sensors are becoming more and more mainstream, there is an increasing need of software tools that help standardize rapid research and discovery, and allow easy reproducibility of the experiments.
+Recent advancements in LiDAR technology have significantly broadened its applicability across various domains, making it a potential sensor for precise measurement. Innovations such as solid-state LiDAR and nanophotonics-based devices have enhanced the performance, reliability, and cost-effectiveness of these sensors, enabling their use in diverse fields including autonomous driving, environmental monitoring, and industrial automation [@10.1002/lpor.202100511]. Additionally, the integration of advanced optics and beam steering technologies has improved the spatial resolution and operational efficiency of LiDAR systems, facilitating their deployment in complex environments for applications ranging from agriculture to urban planning [@10.3390/s22239293;@10.3390/mi13060894]. As the lidar sensors are becoming more and more mainstream, there is an increasing need of software tools that help standardize rapid research and discovery, and allow easy reproducibility of the experiments.
 
 `LiGuard` fulfills a critical need for interactive research in lidar data processing, where iterative adjustments and refinements to processing pipelines are essential and are otherwise time consuming to . While it is not optimized for large-scale datasets such as pointcloudset [@Goelles2021] or focus operating system wide utility and integeration such as ROS [@ros], its user-friendly interface allows researchers to dynamically modify and visualize their workflows, facilitating rapid experimentation and adaptation. PDAL [@pdal_contributors_2024_2616780], a notable library for creating point cloud processing pipelines, though offers much more versatile control over pipelines by providing abstract access, filtering, exploitation, and workflow management capabilities, it lacks interactivity and live parameter tuning which is often crucial in fine-tuning multi-step pipelines. PCL [@Rusu_ICRA2011_PCL], Open3D [@Zhou2018], and other [@brown2012laspy; @pyntcloud; @Hunter:2007; @Lamprecht2019], etc. are low-level libraries that offer fine access to underlying data and are designed to be used as base for the development of versatile 3D applications (`LiGuard` is primarily built on Open3D for its GUI and 3D visualizations needs).
 
[…]
