### 2024-11-19 779591a
delete: remove compile_paper.yml workflow
