### 2024-11-25 cf83fe9
feat: added support in sensor_io for CARLA lidar (lidar.ray_cast) and camera (camera.rgb) sensor.
fix: bug in ouster-sdk import.
