### 2025-01-14 #6751 @m-shahbaz-kharal
Hi @hugoledoux,

I am writing to kindly follow up regarding our recent update to the paper we submitted for review. We had mentioned the update in the this thread but noticed there hasn’t been a response yet. I wanted to confirm that the updated version of the paper has been successfully received and that the review process is ongoing.

We appreciate the time and effort the reviewers are dedicating to this process and understand the review process may take time. I just wanted to ensure everything is in order from our side.

Thanks
