### 2025-01-30 #6751 @hugoledoux
@m-shahbaz-kharal sorry for the long delay.

I see that the code is still not working on linux and/or macOS? Is there a (good) reason for this? Both reviewers, and myself, are macOS users... I see that the code is 100% Python, so I am wondering why that is not working.

Also, why is it only for Python 3.11, and for the lastest?
