### 2025-01-30 #6751 @m-shahbaz-kharal
Hi @hugoledoux,

There's no good reason that we could figure out, it is probably related to Open3D's GUI + Visualizer. We continue to try fixing the issue but for now we'd like to make it available for Windows 10 or later users only (stated in installation section in Readme).

Regarding Python versions, I believe it should work on later (than 3.11) versions too but stated only 3.8-3.11 because those versions are tested by us.
