### 2025-02-04 c92ef99
feat: add SimpleTrack algorithm for 3D multi-object tracking and updated config template.
