+++ liguard/lbl/lumpi_sdk/CameraViewer.py
@@ -1,0 +1,193 @@
+import numpy as np
+import cv2
+import os
+class CameraViewer:
+    def __init__(self, data_path, meta, session, mask_flag):
+        print("loading camera "+str(session))
+        try:
+            self.rvec=np.array(meta["session"][session]['rvec'])
+            self.intrinsic=np.array(meta["session"][session]['intrinsic'])
+            self.tvec=np.array(meta["session"][session]['tvec'])
+            self.inverse_extrinsic=np.array(meta["session"][session]['extrinsic'])
+            self.extrinsic=np.linalg.inv(self.inverse_extrinsic)
+            self.distortion=np.array(meta["session"][session]['distortion'])
+            self.fps=meta["session"][session]['fps']
+            self.decive_id=meta["session"][session]['deviceId']
+            self.img_path=os.path.join(data_path,"cam",str(meta["session"][session]["deviceId"]),"video.mp4")
+            self.mask_path=os.path.join(data_path,"cam",str(meta["session"][session]["deviceId"]),"mask.mp4")
+        except:
+            print("Camera meta "+str(session)+" not found")
+        self.frame_number=0
+        self.size= np.array([0, 0])
+        self.img= np.array([0, 0])
+        try:
+            self.img_capture=cv2.VideoCapture(self.img_path)
+            self.img_capture.set(cv2.CAP_PROP_POS_FRAMES, int(self.frame_number))
+            ret, self.img = self.img_capture.read()
+            self.size=self.img.shape
+            self.last_frame_number=self.img_capture.get(cv2.CAP_PROP_BUFFERSIZE)
+        except:
+            print("Camera "+str(session)+" video not found")
+        self.mask=[]
+        self.mask_capture=[]  
+        if mask_flag:
+            try:
+                self.mask_capture=cv2.VideoCapture(self.mask_path)
+                ret, self.mask = self.mask_capture.read()
+                self.mask_capture.set(cv2.CAP_PROP_POS_FRAMES, int(self.frame_number))
+                if self.size[0]!=self.mask.shape[0] or self.size[1]!=self.mask.shape[1]:
+                    self.size=self.mask.shape
+            except:
+                print("Camera "+str(session)+" mask not found")
+        self.lines= [[0, 1], [1, 2], [2, 3], [0, 3],
+        [4, 5], [5, 6], [6, 7], [4, 7],
+        [0, 4], [1, 5], [2, 6], [3, 7]]
+        self.fov_x =1
+        self.fov_y =1
+        self.planes={}
+        self.transformed_planes = {}
+        try:
+            self.calculate_fov()
+            self.get_frustum_planes(1,200)
+        except:
+            print("Could not calculate frustum")
+       
+    def calculate_fov(self):
+        self.fov_x = 2 * np.degrees(np.arctan2(self.size[1] , (2 * self.intrinsic[0, 0])))
+        self.fov_y = 2 * np.degrees(np.arctan2(self.size[0] , (2 * self.intrinsic[1, 1])))
+    def get_frustum_planes(self, near, far):
+        tan_fov_y = np.tan(np.radians(self.fov_y) / 2)
+        tan_fov_x = np.tan(np.radians(self.fov_x) / 2)
+        nh = tan_fov_y * near
+        nw = tan_fov_x*near
+        self.planes = {
+            "near": np.array([0, 0, 1, -near]),
+            "far": np.array([0, 0, -1, far]),
+            "left": np.array([1, 0, nw / near, 0]),
+            "right": np.array([-1, 0, nw / near, 0]),
+            "top": np.array([0, 1, nh / near, 0]),
[…]
