+++ liguard/lbl/lumpi_sdk/LumpiParser.py
@@ -1,0 +1,142 @@
+import pandas as pd
+import numpy as np
+from tqdm import tqdm
+import os
+import json
+import glob
+from plyfile import PlyData
+from liguard.lbl.lumpi_sdk.CameraViewer import CameraViewer
+class Pose:
+    def __init__(self,id,x,y,z,heading,l,w,h,shape,indices,time,classId=-1):
+        self.position = np.asarray([x, y, z])
+        self.heading = heading
+        self.shape = shape
+        self.dim = np.asarray([l, w, h]).astype(np.float64)
+        self.time = time
+        self.pc = []
+        self.id = int(id)
+        self.index = int(np.floor(float(time) * 10))
+        self.indices = np.asarray(indices).astype(np.int64)
+        self.classId = int(classId)
+        self.u = [np.cos(self.heading), np.sin(self.heading), 0]
+        self.corners = np.array([
+        [-l / 2, -l / 2, l / 2, l / 2, -l / 2, -l / 2, l / 2, l / 2],
+        [w / 2, -w / 2, -w / 2, w / 2, w / 2, -w / 2, -w / 2, w / 2],
+        [-h / 2, -h / 2, -h / 2, -h / 2, h / 2, h / 2, h / 2, h / 2]])
+        self.R = np.array([[self.u[0], -self.u[1], 0], [self.u[1], self.u[0], 0], [0, 0, 1]])
+        # Rotate the corners of the bounding box to align with the object's heading
+        self.corners = np.dot(self.R, self.corners)
+        eight_points = np.tile(self.position, (8, 1))
+        self.corners = self.corners + eight_points.transpose()
+    def interpolate(self, nextPose, time):
+        """
+        Interpolates the position, heading, and dimensions between the current pose and the next pose at a given time.
+
+        Parameters:
+        nextPose (Pose): The next pose to interpolate towards.
+        # Interpolate heading and normalize it to be within the range [-pi, pi]
+        heading = np.fmod(nextPose.heading * weight_next + self.heading * weight_current + np.pi, np.pi * 2) - np.pi
+
+        Returns:
+        tuple: A tuple containing the interpolated position, heading, and dimensions.
+        """
+        pose_time_difference = nextPose.time - self.time
+        time_difference = time - self.time
+        weight_next = time_difference / pose_time_difference
+        weight_current = 1 - weight_next
+        position = nextPose.position * weight_next + self.position * weight_current
+        heading = np.fmod(nextPose.heading * weight_next + self.heading * weight_current + np.pi, np.pi * 2) - np.pi
+        dim = self.dim * weight_current + nextPose.dim * weight_next
+        return position,heading,dim
+class LumpiParser:
+    """
+    LUMPI parser will parse the lumpi data
+    Track format is :frame_number, id, bbox_left, bbox_top, bbox_width, bbox_height, score, class, 1, loc_x, loc_y, loc_z, dim_1, dim_2, dim_3, rotation_y, [shape,3](if classId =1),[indices]
+
+    """
+    helptext="Track format is :frame_number, id, bbox_left, bbox_top, bbox_width, bbox_height, score, class, 1, loc_x, loc_y, loc_z, dim_1, dim_2, dim_3, rotation_y, [shape,3](if classId =1),[indices]"
+    def __init__(self, path):
+        self.path=path
+        self.tracks={}
+        self.indexOrdered={}
+        self.meta={}
+        self.cameras=[]
+        self.lidarPath=""
+        self.point_cloud_files=[]
+        self.pc=""
+        self.read_meta(self.path)
+        self.index=0
+        """
+        The Tracks will be stored in a dictonary first key is object id second key is point cloud index and the value is a Pose
+        """
+    def read_track(self, path):
+        """
[…]
