### 2025-02-07 63352cb
fix: remove unnecessary point cloud file list reading in LUMPI handler.
