### 2025-02-11 #6751 @hugoledoux
I asked and it is allowed to have a package not running on all 3 OSes... although I have to say that I personally do not really like it as it is not useful to many people.

@chenzhaiyu @tgoelles I gather from the comments that you both run macOS? Do you have access to a Windows machine for testing this? If not then I'll have to start a new review. Please let me know.
