### 2025-02-11 #6751 @m-shahbaz-kharal
> I asked and it is allowed to have a package not running on all 3 OSes... although I have to say that I personally do not really like it as it is not useful to many people.
>
> [@chenzhaiyu](https://github.com/chenzhaiyu) [@tgoelles](https://github.com/tgoelles) I gather from the comments that you both run macOS? Do you have access to a Windows machine for testing this? If not then I'll have to start a new review. Please let me know.

Thanks for the comment. We are working on changing the visualizer from open3d to vedo (https://github.com/marcomusy/vedo) that might fix cross-platform issues, however, it might take some time. So, really appreciate if for now it is reviewed based on Windows only. Thanks again for the time and consideration.
