### 2025-02-12 #6751 @tgoelles
Sorry I don't have access to windows. Maybe if they setup some kind of VM or something where I can remotely login it would work. But I agree the usefulness is greatly reduces if its windows only. So yes the vedo alternative looks like a possible alternative.
