### 2025-02-17 214cc8f
fix: STDF loading and parameters handling feat: implement BGFilterInvertCloth algorithm for background filtering
