+++ liguard/algo/lidar.py
@@ -4,7 +4,7 @@ from liguard.gui.logger_gui import Logger
 from liguard.algo.utils import AlgoType, algo_func, get_algo_params, make_key
 algo_type = AlgoType.lidar
 
-import os
+import os; from pathlib import Path
 import sys
 import numpy as np
 
@@ -280,25 +280,28 @@ def BGFilterSTDF(data_dict: dict, cfg_dict: dict, logger: Logger):
     all_skip_frames_keys = [f'{skip_frames_key}_{i}' for i in range(params['number_of_skip_frames_after_each_iter'])]
 
     # load filter if exists
-    if filter_loaded_key not in data_dict and params['load_filter']:
-        # add params to data_dict
-        data_dict[params_key] = params
-        
-        # load filter
-        data_outputs_dir = cfg_dict['data']['outputs_dir']
-        if not os.path.isabs(data_outputs_dir): data_outputs_dir = os.path.join(cfg_dict['data']['pipeline_dir'], data_outputs_dir)
-        filter_params = load_STDF_params(os.path.join(data_outputs_dir, params['filter_file']))
-        if filter_params:
-            data_dict[filter_key] = lambda pcd, threshold: make_STDF_filter(pcd, threshold, **filter_params)
-            data_dict[filter_loaded_key] = True
-            logger.log(f'Filter loaded from {params["filter_file"]}', Logger.INFO)
-        else:
-            data_dict[filter_loaded_key] = False
-            logger.log(f'Failed to load filter from {params["filter_file"]}. Calculating ...', Logger.WARNING)
+    if params['load_filter']:   
+        if filter_loaded_key not in data_dict:
+            # load filter
+            data_outputs_dir = cfg_dict['data']['outputs_dir']
+            if not os.path.isabs(data_outputs_dir): data_outputs_dir = Path(cfg_dict['data']['pipeline_dir'], data_outputs_dir)
+            filter_params = load_STDF_params(Path(data_outputs_dir, params['filter_file']))
+            if filter_params:
+                data_dict[filter_key] = lambda pcd, threshold: make_STDF_filter(pcd, threshold, **filter_params)
+                data_dict[filter_loaded_key] = True
+                data_dict[params_key] = params # to keep track of current params, so in case of change, filter is re-computed
+                logger.log(f'Filter loaded from {params["filter_file"]}', Logger.INFO)
+            else:
+                data_dict[filter_loaded_key] = False
+                logger.log(f'Failed to load filter from {params["filter_file"]}. Calculating ...', Logger.WARNING)
+
+    elif data_dict[filter_loaded_key]:
+        logger.log('A filter was loaded previously, unloading it to calculate new one.', Logger.INFO)
+        data_dict.pop(filter_key)
+        data_dict[filter_loaded_key] = False
     
     # generate filter if not exists
     if filter_key not in data_dict:
-        data_dict[params_key] = params
         # gather frames
         for i in range(params['number_of_frame_gather_iters']):
             gathering_done = gather_point_clouds(data_dict, cfg_dict, all_query_frames_keys[i], params['number_of_frames_in_each_gather_iter'])
@@ -315,15 +318,16 @@ def BGFilterSTDF(data_dict: dict, cfg_dict: dict, logger: Logger):
         data_dict[query_frames_key] = [get_fixed_sized_point_cloud(frame, params['number_of_points_per_frame']) for frame in data_dict[query_frames_key]]
         filter_params = calc_STDF_params(data_dict[query_frames_key], params['lidar_range_in_unit_length'], params['bins_per_unit_length'])
         data_dict[filter_key] = lambda pcd, threshold: make_STDF_filter(pcd, threshold, **filter_params)
[…]
