### 2025-02-28 m-shahbaz-kharal/LiGuard-2.x#1 @chenzhaiyu
Unexpected crashes and error handling
Hi there, thanks for polishing the software since your JOSS submission (https://github.com/openjournals/joss-reviews/issues/6751). While testing, I encountered multiple crashes.

1. Many times, clicking `Apply` caused the application to crash abruptly (the window disappears without an error). This happened several times during a 15-minute playing (see log below; the one-time “LiGuard exited successfully” was when I closed the app manually).
  ```
    (liguard) PS E:\Github> liguard-gui
  FEngine (64 bits) created at 0000026B8FE80050 (threading is enabled)
  FEngine resolved backend: OpenGL
  (liguard) PS E:\Github> liguard-gui
  FEngine (64 bits) created at 000001A3F7591050 (threading is enabled)
  FEngine resolved backend: OpenGL
  (liguard) PS E:\Github> liguard-gui
  FEngine (64 bits) created at 000001B9E38BE050 (threading is enabled)
  FEngine resolved backend: OpenGL
  (liguard) PS E:\Github> liguard-gui
  FEngine (64 bits) created at 00000199090EC050 (threading is enabled)
  FEngine resolved backend: OpenGL
  Quitting...
  LiGuard exited successfully.
  (liguard) PS E:\Github> liguard-gui
  FEngine (64 bits) created at 000002A874567050 (threading is enabled)
  FEngine resolved backend: OpenGL
  (liguard) PS E:\Github> liguard-gui
  FEngine (64 bits) created at 000001A9A963B050 (threading is enabled)
  FEngine resolved backend: OpenGL
  (liguard) PS E:\Github> liguard-gui
  FEngine (64 bits) created at 000002B4E90DA050 (threading is enabled)
  FEngine resolved backend: OpenGL
  (liguard) PS E:\Github> liguard-gui
  ```

2. I also saw some warnings and errors from the console (where I initially typed `liguard-gui`):
```
(liguard) PS E:\Github> liguard-gui
FEngine (64 bits) created at 0000023D6AFA4050 (threading is enabled)
FEngine resolved backend: OpenGL
D:\Softwares\Anaconda\envs\liguard\lib\site-packages\numpy\core\fromnumeric.py:3504: RuntimeWarning: Mean of empty slice.
  return _methods._mean(a, axis=axis, dtype=dtype,
D:\Softwares\Anaconda\envs\liguard\lib\site-packages\numpy\core\_methods.py:121: RuntimeWarning: invalid value encountered in divide
  ret = um.true_divide(
Quitting...
[Open3D WARNING] GLFW Error: The GLFW library is not initialized
[Open3D WARNING] GLFW Error: The GLFW library is not initialized
[Open3D WARNING] GLFW Error: The GLFW library is not initialized
```
However these were not shown in the applications `Log` window.

3. I even encountered a memory access error (see attached image; sorry for the language setting on this machine), which translates to:
```
python.exe – Application Error
The instruction at 0x00007FFA7C6BA7AE referenced memory at 0x0000000000000068. The memory could not be read.
To terminate the program, click “OK.”
```
![Image](https://github.com/user-attachments/assets/64a55ff8-357d-4b73-9bd0-62389a85db31)

My Env is Windows 11 and Anaconda with Python 3.10 and using your default data. Since these issues occur frequently and somewhat randomly, I perhaps can spare providing exact reproduction steps. It seems that certain configs trigger these errors (in that case it needs to be properly handled), and the current error handling is not good enough. Could you please investigate and improve the robustness of the software?
