### 2025-02-28 #6751 @chenzhaiyu
Besides the issue I raised there: https://github.com/m-shahbaz-kharal/LiGuard-2.x/issues/1, two small issues:
* Is `LiGuard-2.x` the best repo name with the probably verbose version suffix `2.x`?
* I cannot see the updated paper PDF, maybe regenerate one?
