### 2025-03-05 #6751 @hugoledoux
> * I cannot see the updated paper PDF, maybe regenerate one?

you can do yourself this with `@editorialbot generate pdf`
