### 2025-03-12 330cdbf
fix: safe open3d visualizer exit.
