### 2025-03-21 #6751 @m-shahbaz-kharal
Apologies for the delayed response. I wanted to provide an update that I’m actively working on a major overhaul to address the cross-platform support issues.

Thank you, @chenzhaiyu, for reporting the issue. As part of the update, I’m also improving exception handling within the configuration read/apply process.

I’ll share a detailed update with all the changes by this coming Friday (03/28).
