### 2025-03-21 #6751 @m-shahbaz-kharal
Sorry for delayed response. I am sending this to inform that I am actively working on an overhaul to fix cross-platform support issue.

Thank you, @chenzhaiyu, for reporting the issue. As part of the update, I’m also improving exception handling within the configuration read/apply process.

I’ll share a detailed update with all the changes by this coming Friday (03/28).
