### 2025-03-28 179c3c7
Major update, refactoring, and bug fixes.
- Moved resolve_for_application_root and resolve_for_default_workspace from config_gui.py to new gui_utils.py for modularity
- Updated all imports across liguard modules to reflect new location
- Refactored logger_gui.py: cleaned up logging logic, renamed methods, and improved UI updates
- Refactored liguard_gui.py:
  - Unified visualizer references (pcd_viz, img_viz) and removed redundant logic
  - Improved threading, input handling, and frame control
  - Simplified reset/start/quit workflows
- Enhanced error handling and file validation across file_i
[…]
