+++ liguard/resources/algo_func_template.py
@@ -1,7 +1,7 @@
 #########################################################################################################################
 # usually, the following imports are common for all the algorithms, so it is recommended to not remove them
 import inspect
-from liguard.gui.config_gui import resolve_for_application_root, resolve_for_default_workspace
+from liguard.gui.gui_utils import resolve_for_application_root, resolve_for_default_workspace
 from liguard.gui.logger_gui import Logger
 from liguard.algo.utils import AlgoType, algo_func, get_algo_params, make_key
 algo_type = AGLO_TYPE
