+++ liguard/img/viz.py
@@ -1,276 +1,292 @@
 import os
-from liguard.gui.config_gui import resolve_for_application_root, resolve_for_default_workspace
+import time
 
 import open3d as o3d
+import open3d.visualization.gui as gui
+import open3d.visualization.rendering as rendering
 import cv2
 import numpy as np
 
+from liguard.gui.gui_utils import resolve_for_application_root, resolve_for_default_workspace
 from liguard.gui.logger_gui import Logger
+
+import threading
 
 class ImageVisualizer:
     """
-    A class for visualizing images and bounding boxes using Open3D.
-
-    Args:
-        app: The application object.
-        cfg (dict): A dictionary containing configuration parameters.
-
-    Attributes:
-        app: The application object.
-        viz: The Open3D visualizer object.
-        img: The image to be visualized.
-        geometries (dict): A dictionary to store the geometries added to the visualizer.
-
-    Methods:
-        reset: Resets the visualizer and clears all geometries.
-        update: Updates the visualizer with new data.
-        redraw: Redraws the visualizer.
-        quit: Destroys the visualizer window.
+    Initializes the ImageVisualizer.
+        Args:
+            app (gui.Application): The application object.
+            logger (Logger): The logger object.
     """
 
-    def __init__(self, app, cfg: dict):
+    def __init__(self, app: gui.Application, logger: Logger=None):
         self.app = app
+        if logger: self.log = logger.log
+        else: self.log = lambda msg,lvl: print(f'[Image Visualizer][{Logger.__level_string__[lvl]}] {msg}')
+        
         # create visualizer
-        self.viz = o3d.visualization.Visualizer()
-        self.win_created = self.viz.create_window("Image Feed", width=int(1440/4), height=int(1080/4), left=480 - int(1440/4), top=30)
-        # init
-        # create necessary paths
-        if cfg['visualization']['camera']['save_images']:
-            # make sure the outputs_dir is created
-            data_outputs_dir = cfg['data']['outputs_dir']
-            if not os.path.isabs(data_outputs_dir): data_outputs_dir = os.path.join(cfg['data']['pipeline_dir'], data_outputs_dir)
-            self.image_save_path = os.path.join(data_outputs_dir, 'img_viz')
-            os.makedirs(self.image_save_path, exist_ok=True)
+        self.viz = o3d.visualization.O3DVisualizer(title="Image Feed", width=int(1440/4), height=int(1080/4))
+        self.viz.set_on_close(lambda: False)
+        self.viz.show_menu(False)
+        self.viz.show_settings = False
+        self.viz.show_skybox(False)
+
+        #
+        self.lock = threading.Lock()
+
+        # add to window
+        self.app.add_window(self.viz)
+        
         # reset
-        self.reset(cfg, True)
+        self.reset()
         
-    def reset(self, cfg, reset_bounding_box=False):
+    def reset(self, cfg=None, reset_bounding_box=False):
         """
-        Resets the visualizer and clears all geometries.
+        Resets the visualizer.
 
         Args:
-            cfg: The configuration parameters.
-            reset_bounding_box (bool): Whether to reset the bounding box or not.
+            cfg: A dictionary containing configuration parameters.
+            reset_bounding_box: Whether to reset the bounding box or not.
         """
+        # acquire lock
+        self.lock.acquire()
+
+        # update cfg
         self.cfg = cfg
-        # clear geometries
-        self.geometries = dict()
-        self.viz.clear_geometries()
-        # add default geometries
[…]
