+++ liguard/gui/config_gui.py
@@ -1,20 +1,14 @@
 import open3d.visualization.gui as gui
 
+from liguard.gui.logger_gui import Logger
+from liguard.gui.gui_utils import resolve_for_application_root, resolve_for_default_workspace
+
 import os
 import sys
-import time
 import yaml
 import ast
 
-def resolve_for_application_root(path:str) -> str:
-    application_root_dir = os.path.abspath(os.path.dirname(os.path.dirname(__file__)))
-    if not os.path.isabs(path): path = os.path.join(application_root_dir, path)
-    return os.path.abspath(path)
-
-def resolve_for_default_workspace(path:str) -> str:
-    default_workspace_dir = os.path.join(os.path.expanduser("~"), 'liguard-default-workspace')
-    if not os.path.isabs(path): path = os.path.join(default_workspace_dir, path)
-    return os.path.abspath(path)
+import threading
 
 class BaseConfiguration:
     def get_callbacks_dict():
@@ -28,21 +22,25 @@ class BaseConfiguration:
             'new_config': [],       # List to store callback functions for 'new_config' action
             'open_config': [],      # List to store callback functions for 'open_config' action
             'save_config': [],      # List to store callback functions for 'save_config' action
-            'save_as_config': [],   # List to store callback functions for 'save_as_config' action
             'apply_config': [],     # List to store callback functions for 'apply_config' action
             'quit_config': []       # List to store callback functions for 'quit_config' action
         }
         
-    def __init__(self, app: gui.Application, callbacks = get_callbacks_dict()):
+    def __init__(self, app: gui.Application, logger: Logger=None):
         """
         Initializes the ConfigGUI class.
 
         Args:
             app (gui.Application): The application object.
-            callbacks (dict): A dictionary containing the callbacks for the GUI.
+            logger (Logger): The logger object.
         """
+        # Set the application and logger objects
         self.app = app
-        self.callbacks = callbacks
+        if logger: self.log = logger.log
+        else: self.log = lambda msg,lvl: print(f'[Configuration][{Logger.__level_string__[lvl]}] {msg}')
+
+        # default empty callbacks
+        self.callbacks = BaseConfiguration.get_callbacks_dict()
 
         # Create a window for the configuration GUI
         self.mwin = app.create_window("Configuration", 480, 1080, x=0, y=30)
@@ -129,8 +127,7 @@ class BaseConfiguration:
         Returns:
             dict: The loaded configuration as a dictionary.
         """
-        with open(cfg_path) as f:
-            cfg = yaml.safe_load(f)
+        with open(cfg_path) as f: cfg = yaml.safe_load(f)
         return cfg
     
     def load_pipeline_algos(self):
@@ -381,8 +378,12 @@ class BaseConfiguration:
         ok_button.horizontal_padding_em = 0.5
         ok_button.vertical_padding_em = 0
         def ok_button_callback():
-            self.cfg[self.input_dialog_key] = getattr(self.input_dialog_widget, self.input_dialog_widget_value_variable)
             self.__close_dialog__()
+            try: self.cfg[self.input_dialog_key] = getattr(self.input_dialog_widget, self.input_dialog_widget_value_variable)
+            except Exception as e:
+                self.log(f'Failed to set configuration value: {e}', Logger.ERROR)
+                self.__show_issue_dialog__(f'Failed to set {self.input_dialog_key} value. See logs for details.')
+            
         if custom_callback: ok_button.set_on_clicked(custom_callback)
         else: ok_button.set_on_clicked(ok_button_callback)
         layout.add_child(ok_button)
@@ -405,31 +406,54 @@ class BaseConfiguration:
         
     def __new_config__(self):
         def create_base_pipeline(pipeline_dir):
[…]
