+++ liguard/algo/pre.py
@@ -1,5 +1,5 @@
 import inspect
-from liguard.gui.config_gui import resolve_for_application_root, resolve_for_default_workspace
+from liguard.gui.gui_utils import resolve_for_application_root, resolve_for_default_workspace
 from liguard.gui.logger_gui import Logger
 from liguard.algo.utils import AlgoType, algo_func, get_algo_params, make_key
 algo_type = AlgoType.pre
