### 2025-03-28 ba2f47e
most crashes are fixed. still more to go.
