### 2025-03-28 cda3110
fix: improve GUI responsiveness and logging functionality.
