### 2025-03-28 282c732
fix: improve GUI responsiveness and logging functionality.
