### 2025-03-28 2cbe6d1
feat: add show and hide methods for visualizers and update visibility logic
