### 2025-03-28 87c6e78
feat: add show and hide methods for visualizers and update visibility logic
