+++ liguard/img/viz.py
@@ -40,6 +40,18 @@ class ImageVisualizer:
         
         # reset
         self.reset()
+
+    def show(self):
+        """
+        Shows the visualizer window.
+        """
+        self.app.post_to_main_thread(self.viz, lambda: self.viz.show(True))
+
+    def hide(self):
+        """
+        Hides the visualizer window.
+        """
+        self.app.post_to_main_thread(self.viz, lambda: self.viz.show(False))
         
     def reset(self, cfg=None):
         """
