### 2025-03-28 e8c54bd
fix: enhance error logging in configuration update process with debug key tracking
