### 2025-03-28 ee23b39
fix: improve handling of unavailable data sources in the GUI
