### 2025-03-29 2a84e37
fix: bbox and trajectory ids are updated to let visualizer know that new geometries are being rendered
