+++ liguard/liguard_gui.py
@@ -251,7 +251,7 @@ class LiGuard:
                     else: process = __import__(proc, fromlist=['*']).__dict__[proc]
                     self.pre_processes[priority] = process
                 except Exception:
-                    self.logger.log(f'pre_processes creation failed for {proc.__name__}:\n{traceback.format_exc()}', Logger.CRITICAL)
+                    self.logger.log(f'pre_processes creation failed for {proc}:\n{traceback.format_exc()}', Logger.CRITICAL)
         self.pre_processes = [self.pre_processes[priority] for priority in sorted(self.pre_processes.keys())]
         self.logger.log(f'enabled pre_processes: {[f.__name__ for f in self.pre_processes]}', Logger.DEBUG)
         
@@ -267,7 +267,7 @@ class LiGuard:
                     else: process = __import__(proc, fromlist=['*']).__dict__[proc]
                     self.lidar_processes[priority] = process
                 except Exception:
-                    self.logger.log(f'lidar_processes creation failed for {proc.__name__}:\n{traceback.format_exc()}', Logger.CRITICAL)
+                    self.logger.log(f'lidar_processes creation failed for {proc}:\n{traceback.format_exc()}', Logger.CRITICAL)
         self.lidar_processes = [self.lidar_processes[priority] for priority in sorted(self.lidar_processes.keys())]
         self.logger.log(f'enabled lidar_processes: {[f.__name__ for f in self.lidar_processes]}', Logger.DEBUG)
         
@@ -283,7 +283,7 @@ class LiGuard:
                     else: process = __import__(proc, fromlist=['*']).__dict__[proc]
                     self.camera_processes[priority] = process
                 except Exception:
-                    self.logger.log(f'camera_processes creation failed for {proc.__name__}:\n{traceback.format_exc()}', Logger.CRITICAL)
+                    self.logger.log(f'camera_processes creation failed for {proc}:\n{traceback.format_exc()}', Logger.CRITICAL)
         self.camera_processes = [self.camera_processes[priority] for priority in sorted(self.camera_processes.keys())]
         self.logger.log(f'enabled camera_processes: {[f.__name__ for f in self.camera_processes]}', Logger.DEBUG)
 
@@ -299,7 +299,7 @@ class LiGuard:
                     else: process = __import__(proc, fromlist=['*']).__dict__[proc]
                     self.calib_processes[priority] = process
                 except Exception:
-                    self.logger.log(f'calib_processes creation failed for {proc.__name__}:\n{traceback.format_exc()}', Logger.CRITICAL)
+                    self.logger.log(f'calib_processes creation failed for {proc}:\n{traceback.format_exc()}', Logger.CRITICAL)
         self.calib_processes = [self.calib_processes[priority] for priority in sorted(self.calib_processes.keys())]
         self.logger.log(f'enabled calib_processes: {[f.__name__ for f in self.calib_processes]}', Logger.DEBUG)
 
@@ -315,7 +315,7 @@ class LiGuard:
                     else: process = __import__(proc, fromlist=['*']).__dict__[proc]
                     self.label_processes[priority] = process
                 except Exception:
-                    self.logger.log(f'label_processes creation failed for {proc.__name__}:\n{traceback.format_exc()}', Logger.CRITICAL)
+                    self.logger.log(f'label_processes creation failed for {proc}:\n{traceback.format_exc()}', Logger.CRITICAL)
         self.label_processes = [self.label_processes[priority] for priority in sorted(self.label_processes.keys())]
         self.logger.log(f'enabled label_processes: {[f.__name__ for f in self.label_processes]}', Logger.DEBUG)
         
@@ -331,7 +331,7 @@ class LiGuard:
[…]
