### 2025-03-29 7948697
fix: bump version to 2.1.4 and update open3d dependency to 0.19.0
