### 2025-03-29 m-shahbaz-kharal/LiGuard-2.x#1 @m-shahbaz-kharal
Hi @chenzhaiyu, thank you again for your testing and feedback.

I've just released [v2.1.4](https://github.com/m-shahbaz-kharal/LiGuard-2.x/releases/tag/v2.1.4) which addresses the crash and memory management issues you encountered, it is also available on PyPI so you can just use `pip install --upgrade LiGuard` .

Following is a quick summary of updates:
- Crash Fixes: Improved error handling and added validation checks for missing/invalid data to prevent random crashes.
- Refactoring: Reorganized GUI logic and module imports for better stability and maintainability.
- Visualizer Update: Migrated from `open3d.visualization.Visualizer` to `open3d.visualization.O3DVisualizer`, found it to be much more robust and has better memory management. There are still some warnings, but I found them to be not problematic.
- Logging Improvements: Enhanced logging to include more debug info and reflect accurate process states.

I have tested it on Windows 11 with Python 3.8 so far, but I am confident that it'll work on all the next versions of python. As far as macOS is concerned, I am looking for a macOS-based PC/laptop and will reply in the main review thread once I'll test it out.

These changes should significantly improve robustness and eliminate the crashes you reported. Please feel free to test the new release and let me know if any issues persist. Appreciate your time and help improving our submission to JOSS.
