### 2025-03-29 #6751 @m-shahbaz-kharal
**Update:** Yesterday, in our release [v2.1.4](https://github.com/m-shahbaz-kharal/LiGuard-2.x/releases/tag/v2.1.4), crash issues reported by @chenzhaiyu in [m-shahbaz-kharal/LiGuard-2.x#1](https://github.com/m-shahbaz-kharal/LiGuard-2.x/issues/1#issuecomment-2763027098) are fixed. The release includes many improvements in modularity, error handling, GUI responsiveness, and visualization—please check the release notes for more details.

**Note:** The release doesn't include support for macOS yet. It is requested that, if possible, **please consider our paper submission for a Windows only software**, and proceed for publication, if it meets the required criteria. We understand that cross-platform support is important, and we are changing the base libraries (for GUI and 3D visualization) to make it work. But based on our recent efforts, we feel that it is much more involved process and might take a long time. The request is, therefore, being put so to complete a timely publication of the paper.
