### 2025-04-03 4c29c49
fix: enhance calibration handler to validate Tr_velo_to_cam shape and set default color for unknown object classes
