### 2025-04-13 m-shahbaz-kharal/LiGuard-2.x#2 @chenzhaiyu
Unresponsive selection action
Part of JOSS review: https://github.com/openjournals/joss-reviews/issues/6751

I noticed that selecting entities in the UI feels a bit unresponsive. It seems that the selection action only triggers *after* I move the cursor following the initial click. I've attached two short videos showing the issue.

It would greatly improve the user experience if the UI responded immediately to a selection, without requiring additional cursor movement. Perhaps this behavior could be made more responsive or intuitive.

https://github.com/user-attachments/assets/161e166d-7e88-4422-a118-43f5508b380b

https://github.com/user-attachments/assets/77270940-fdac-44fc-b105-a5e0721f53d2
