### 2025-04-13 m-shahbaz-kharal/LiGuard-2.x#3 @chenzhaiyu
Suggestion to add issue templates
Part of JOSS review: https://github.com/openjournals/joss-reviews/issues/6751

Just a small suggestion — it might be helpful to add [issue templates](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository) to the repository. This could encourage more structured and informative bug reports or feature requests from users. 🙂
