### 2025-04-13 m-shahbaz-kharal/LiGuard-2.x#4 @chenzhaiyu
Broken assets on PyPI page
Part of JOSS review: https://github.com/openjournals/joss-reviews/issues/6751

The images (and possibly other assets) appear to be broken on the [PyPI page](https://pypi.org/project/LiGuard/), likely due to the use of relative paths. You might want to consider using absolute URLs that point directly to the assets in this repository :)
