### 2025-04-15 m-shahbaz-kharal/LiGuard-2.x#2 @m-shahbaz-kharal
@chenzhaiyu thanks for reporting issue, it seems like a bug in [open3d](https://github.com/isl-org/Open3D)'s GUI as I tested their [all_widgets.py](https://www.open3d.org/docs/release/python_example/visualization/index.html#all-widgets-py) example script and that also produces the same issue. However, the clicking is actually working, for example if you click on dropdown and move the mouse slightly it'll draw the dropdown menu. Also, this part of GUI is managed by `o3d.visualization.O3DVisualizer` class in original `open3d` code so not much can be done on our side to improve the responsiveness. I'll post the issue in the `open3d`'s repo and update our code to include the fix once that issue is resolved.
