### 2025-04-15 cf5547b
fix: add default color for unknown object classes in bounding box handler.
