### 2025-04-15 ee81406
fix: update README links and image sources to use absolute URLs.
