+++ README.md
@@ -1,5 +1,5 @@
 <p align="center">
-    <img src="docs/figs/liguard-logo.png" alt="Description" width="400" />
+    <img src="https://github.com/m-shahbaz-kharal/LiGuard-2.x/raw/dev_2.x/docs/figs/liguard-logo.png" alt="Description" width="400" />
 </p>
 
 <p align="center">
@@ -7,7 +7,7 @@
 </p>
 
 <p align="center">
-    <a href="#installation">Installation</a> | <a href="#usage">Usage</a> | <a href="#documentation">Documentation</a> | <a href="#contributing">Contributing</a> | <a href="#license">License</a>
+    <a href="https://github.com/m-shahbaz-kharal/LiGuard-2.x/?tab=readme-ov-file#installation">Installation</a> | <a href="https://github.com/m-shahbaz-kharal/LiGuard-2.x/?tab=readme-ov-file#usage">Usage</a> | <a href="https://m-shahbaz-kharal.github.io/LiGuard-2.x/">Documentation</a> | <a href="https://github.com/m-shahbaz-kharal/LiGuard-2.x/?tab=readme-ov-file#contributing">Contributing</a> | <a href="https://github.com/m-shahbaz-kharal/LiGuard-2.x/blob/dev_2.x/LICENSE.txt">License</a>
 </p>
 
 ![PyPI Release Version](https://img.shields.io/pypi/v/liguard?label=release)
@@ -20,7 +20,7 @@
 
 `LiGuard` features, out of the box, data reading for many common dataset formats including support for reading calibration and label data. Moreover, it provides (an increasing list of) commonly used algorithm components ranging from basic data preprocessors to advanced object detection and tracking algorithms. Additionally, it establishes a straightforward standard for adding custom functions/algorithms, allowing users to integrate unique components into their pipelines. Pipelines created in `LiGuard` are saved in structured directories, making it easy to share and reproduce results.
 
-![LiGuard Main Interface](docs/figs/liguard-main.png)
+![LiGuard Main Interface](https://github.com/m-shahbaz-kharal/LiGuard-2.x/blob/dev_2.x/docs/figs/liguard-main.png)
 *LiGuard's GUI Layout (from left to right): Configuration Window, Visualization Windows (Point Cloud Feed and Image Feed), and Log Window.*
 
 # Installation
@@ -43,10 +43,10 @@ Test an example pipeline:
 1. In the `Configuration` windows, click the `Open` button.
 2. Navigate to `examples/simple_pipeline`, click open, and then click `Apply`.
 3. Explore various functions under `proc` dropdown in the `Configuration` window. For example, under `proc/lidar/crop`, check the `enabled` checkbox, and click `Apply` to see the cropped LIDAR data.
-4. Press `left arrow` or `right arrow` to navigate through the frames. A list of all key bindings is available [here](docs/visualizer_key_bindings.md).
+4. Press `left arrow` or `right arrow` to navigate through the frames. A list of all key bindings is available [here](https://github.com/m-shahbaz-kharal/LiGuard-2.x/blob/dev_2.x/docs/visualizer_key_bindings.md).
 5. To save the pipeline, click the `Save` button in the `Configuration` window.
 
-For more details on pipelines, see [LiGuard Pipelines](docs/liguard_pipelines.md).
+For more details on pipelines, see [LiGuard Pipelines](https://github.com/m-shahbaz-kharal/LiGuard-2.x/blob/dev_2.x/docs/liguard_pipelines.md).
 
 # Documentation
[…]
