### 2025-04-15 c46ac23
fix: adjusted image source for main interface fix: bump version to 2.1.4.post2 in setup.py
