### 2025-04-15 m-shahbaz-kharal/LiGuard-2.x#4 @m-shahbaz-kharal
@chenzhaiyu thanks for noting the problem, I have just pushed commits https://github.com/m-shahbaz-kharal/LiGuard-2.x/commit/ee8140688443496772ab7180d50e76fbe68e6b60 and https://github.com/m-shahbaz-kharal/LiGuard-2.x/commit/c46ac23b5a7cf9492d972d55dab6d323368d3c42 to include absolute URLs. I have also uploaded a post release to [PyPI page](https://pypi.org/project/LiGuard/) fixing the links. Thanks.
