+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,18 @@
+---
+name: Feature request
+about: Suggest a feature (dataset support, add commonly used algorithm, researchers
+  often need ..., etc.)
+title: "[FEATURE REQUEST]"
+labels: enhancement
+assignees: m-shahbaz-kharal
+
+---
+
+**Is your feature request generally useful and is related to Lidar data processing?**
+A clear and concise description of the feature and how do you think it is generally useful for Lidar data processing.
+
+**Describe the solution you'd like**
+Please add details about the solution you need. Add references to sources that might be useful (Github repositories, articles explaining an algorithm, etc.)
+
+**Additional context**
+Add any other context or screenshots about the feature request here.
