+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,35 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: "[BUG] "
+labels: bug
+assignees: m-shahbaz-kharal
+
+---
+
+**Describe the bug**
+A clear and concise description of the bug.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+1. Go to '...'
+2. Click on '....'
+3. Scroll down to '....'
+4. See error
+
+**Expected behavior**
+A clear and concise description of your expectation.
+
+**Screenshot(s)**
+If needed, please post screenshots or short clip of the observed problem.
+
+**Log(s)**
+`LiGuard` generates `logs` folder in the project folder. Please post the relevant log file. Also, add the terminal output if an error is printed. Ignore warnings.
+
+**Environment:**
+ - OS: [e.g. Windows]
+ - LiGuard Version [e.g. 2.1.4.post1]
+ - Python Version [e.g. 3.8]
+
+**Additional context**
+Add any other context about the problem here.
