### 2025-04-15 m-shahbaz-kharal/LiGuard-2.x#3 @m-shahbaz-kharal
@chenzhaiyu thanks for suggestion. I have added the issue templates driven by our current method of fixing bugs. Also added one for feature request (primarily for researchers to requiring support for a dataset, sensor, build-in algorithm, etc.).
