### 2025-04-16 90eb394
chore: corrected key bindings in documentation.
