+++ docs/visualizer_key_bindings.md
@@ -4,12 +4,8 @@
 | `left arrow` | Previous frame |
 | `right arrow` | Next frame |
 | `space` | Play/Pause |
-| `+` | Increase Point Size |
-| `-` | Decrease Point Size |
 | `delete` | Reset Playback |
 | `[` | Jump to frame idx |
-| `mouse wheel` | Zoom in/out |
-| `left mouse button` | Scene Rotate |
-| `alt + left mouse button` | Camera Rotate |
-| `ctrl + left mouse button` | Camera Pan |
-| `shift + left mouse button` | Camera Roll |
+| `left mouse button` | Rotate (Arcball/Fly/Model/Environment) |
+| `ctrl + left mouse button` | Pan |
+| `shift + left mouse button` or `mouse wheel` | Zoom |
