### 2025-04-16 7694b70
chore: changed default point size to 1, set default ground plane to be along XY, set default rendering mode to raw to clearly see points.
fix: reading point size from configuration.
