### 2025-04-17 ad0d5c1
fix: update key event handling to check visibility of visualizers before processing input.
