### 2025-04-17 98b3540
chore: corrected key bindings information in documentation.
