### 2025-04-17 eb230e8
fix: window positioning bug in macOS
