### 2025-04-17 0677128
feat: implement macOS Quartz keyboard listener fallback and improve key event handling
