### 2025-04-17 6a84037
docs: update installation instructions to include macOS support and accessibility permissions note.
