### 2025-04-17 edea4af
chore: update Python version requirements to 3.10 or later to support Open3D=0.19.0 on both Windows and macOS.
