+++ docs/sphinx/README.md
@@ -15,7 +15,7 @@ Requirements:
 - OS
   - Windows 10 or later
   - macOS 10.14 Mojave or later
-- Python 3.8, 3.9, 3.10, or 3.11
+- Python 3.10 or later
 
 Install `LiGuard` with pip (from PyPI):
 ```bash
