### 2025-04-17 b4bfc31
chore: bump version to 2.1.5 as macOS is now supported.
