### 2025-04-17 #6751 @m-shahbaz-kharal
Thanks, @chenzhaiyu, for your valuable review. We appreciate your feedback and are committed to continuously improving our software.

@hugoledoux and @tgoelles: We have now released [v2.1.5](https://github.com/m-shahbaz-kharal/LiGuard-2.x/releases/tag/v2.1.5), which adds support for macOS 10.14 Mojave and later versions. Reviewers who couldn't review previously can now proceed. Please continue your review at your earliest convenience.
