### 2025-04-17 #6751 @tgoelles
Nice to see the macOS support. So have you tested also on Linux? Its not a requirement for me, but chances are that if it works for macOS it also works on linux.
I am looking forward to try this, but can not do it before May.
