### 2025-04-17 #6751 @m-shahbaz-kharal
> So have you tested also on Linux? Its not a requirement for me, but chances are that if it works for macOS it also works on linux.

@tgoelles: Thanks for your comment. I tested it today and it works perfectly on Linux as well.
