### 2025-04-18 f622b5d
docs: updated installation requirements to include Ubuntu support. chore: added missing Environment classifier for macOS.
