### 2025-04-18 92fd481
refactor: increase consistency of FileIO class structure for both calibration and label files.
