### 2025-04-21 fb878ea
feat: added support for visualizing 3D bounding box class labels.
