### 2025-05-07 m-shahbaz-kharal/LiGuard-2.x#6 @tgoelles
use pyproject.toml better
see for example: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#writing-pyproject-toml or https://docs.astral.sh/uv/concepts/projects/config/.

Take all information from setup.py and all requirements.txt into pyproject.toml. Optionally also use uv and the uv build backend. Use ependency-groups https://packaging.python.org/en/latest/specifications/dependency-groups/
