### 2025-05-07 #6751 @tgoelles
## General Comments

The GUI now runs on my Mac, and I was able to test it. This is a significant improvement—making the application cross-platform could substantially increase its user base. That said, my original major and specific comments have not been addressed, nor have I received any feedback on them ([see issue](https://github.com/m-shahbaz-kharal/LiGuard-JOSS/issues/1)). **These must be addressed before I can complete my review.**

### As I currently understand it, the tool aims to:

* Provide an opinionated but flexible framework for lidar and camera data processing
* Include some built-in processing steps
* Allow users to define new steps via provided templates
* Support batch processing via `liguard-cmd`

### Main drawbacks of the current approach:

* For each dataset, users must generate a folder structure or at least a YAML config file. There is no way to define a single pipeline and simply point it at multiple datasets.
* The function templates are somewhat convoluted. Some appear redundant—why not unify or simplify them?
* Custom postprocessing steps generate a YAML file rather than a Python script, which seems counterintuitive for users who expect to write and reuse code.
* There is no support for reusing algorithms from existing ROS packages or for integrating with community-maintained lidar processing libraries, such as [LiDAR-Processing-V2](https://github.com/YevgeniyEngineer/LiDAR-Processing-V2).

I would see a stronger use case for LiGuard **if**:

* It supported integration with ROS packages or at least accepted common ROS-related formats
* It provided a genuinely beginner-friendly way to interactively explore and experiment with algorithms

---

### Statement of Need

The most pressing issue remains the **Statement of Need**, which stil does not mention or cite important, widely used frameworks such as **ROS** or **PDAL**. These are foundational tools in lidar and robotic data processing, and omitting them gives the impression that the authors may not be aware of them.

Moreover, the paper includes this statement:

> “...the focus is still developing a python package that can be imported to custom, often hard-coded scripts, thus imposing limitations on reusability and scalability.”

This assertion misrepresents the capabilities of existing tools like `pointcloudset`. In practice, general-purpose Python libraries—particularly those without GUI dependencies—are **more** reusable and scalable. They can be imported into custom pipelines, integrated into batch processing systems, and deployed in production environments with minimal setup. For example, `pointcloudset` is currently used in a production environment to process data from four lidar units continuously for over six months.

LiGuard, by contrast, requires a tightly coupled setup of folder structures, YAML config files, and GUI or CLI components. This makes it harder to scale across diverse use cases or integrate into automated workflows.

The [pointcloudset documentation](https://virtual-vehicle.github.io/pointcloudset/tutorial_notebooks/usage.html#Applying-Functions-to-the-whole-Dataset) demonstrates how users can write custom Python functions and apply them across datasets - achieving a similar functional goals as LiGuard, but with more flexibility and fewer constraints.

Similarly, tools like **ROS**, combined with visualizers such as **RViz** or **Foxglove**, provide powerful interfaces for synchronizing and visualizing lidar and camera data. Of course, ROS introduces complexity, and a simpler tool like LiGuard could be helpful—**if** it were aware of this ecosystem and offered meaningful integration or compatibility.

Lastly, while `liguard-cmd` appears to use some form of multithreading, this is not mentioned or explained in the paper or documentation. If multithreading is a distinguishing feature, it should be clearly stated.

---

## Functionality

Once I click "Apply" in the GUI, I cannot undo or modify the step without restarting the program. This makes experimentation cumbersome.

Running `liguard-cmd` with the example pipeline initially showed progress (100%) on all steps, but then failed to exit cleanly and produced no output data.

---

## New Specific Comments

* Capitalize **Python** consistently.
* Clearly explain whether and how `liguard-cmd` uses multithreading, both in the paper and in the documentation.
