### 2025-05-07 21bc52c
fix: correct capitalization of "Python" in documentation.
