+++ paper/paper.md
@@ -29,7 +29,7 @@ There is a growing interest in the development of lidar-based applications in do
 
 `LiGuard` features, out of the box, data reading for many common dataset formats including support for reading calibration and label data. Moreover, it provides (an increasing list of) commonly used algorithm components ranging from basic data preprocessors to advanced object detection and tracking algorithms. Additionally, it establishes a straightforward standard for adding custom functions/algorithms, allowing users to integrate unique components into their pipelines. For latest updates on supported features, please see [LiGuard’s documentation](https://m-shahbaz-kharal.github.io/LiGuard-2.x/).
 
-Note: `LiGuard` is built on many standard python libraries and packages for data processing and visualization including Open3D [@Zhou2018], OpenCV [@opencv_library], Numpy [@harris2020array], and SciPy [@2020SciPy]. `LiGuard` is designed to be used interactively and is therefore not ideal for processing very large datasets. We provide a utility [`liguard_cmd`](https://github.com/m-shahbaz-kharal/LiGuard-2.x/blob/dev_2.x/liguard/liguard_cmd.py) that processes data slightly faster by removing GUI and implementing multithreading but since the data is still processed sequentially the output generation is as fast as the slowest step in the pipeline.
+Note: `LiGuard` is built on many standard Python libraries and packages for data processing and visualization including Open3D [@Zhou2018], OpenCV [@opencv_library], Numpy [@harris2020array], and SciPy [@2020SciPy]. `LiGuard` is designed to be used interactively and is therefore not ideal for processing very large datasets. We provide a utility [`liguard_cmd`](https://github.com/m-shahbaz-kharal/LiGuard-2.x/blob/dev_2.x/liguard/liguard_cmd.py) that processes data slightly faster by removing GUI and implementing multithreading but since the data is still processed sequentially the output generation is as fast as the slowest step in the pipeline.
 
 # Statement of Need
 Recent advancements in LiDAR technology have significantly broadened its applicability across various domains, making it a potential sensor for precise measurement. Innovations such as solid-state LiDAR and nanophotonics-based devices have enhanced the performance, reliability, and cost-effectiveness of these sensors, enabling their use in diverse fields including autonomous driving, environmental monitoring, and industrial automation [@10.1002/lpor.202100511]. Additionally, the integration of advanced optics and beam steering technologies has improved the spatial resolution and operational efficiency of LiDAR systems, facilitating their deployment in complex environments for applications ranging from agriculture to urban planning [@10.3390/s22239293;@10.3390/mi13060894]. As the lidar sensors are becoming more and more mainstream, there is an increasing need of software tools that help standardize rapid research and discovery, and allow easy reproducibility of the experiments.
