### 2025-05-07 #6751 @m-shahbaz-kharal
@tgoelles thanks for the review,
Please accept my sincere apology, **the editorial bot is pointing to the older version of `paper.md`**. I thought that it'll automatically find the paper in [LiGuard-2.x](https://github.com/m-shahbaz-kharal/LiGuard-2.x) repo under [joss-paper](https://github.com/m-shahbaz-kharal/LiGuard-2.x/tree/joss-paper) branch. But now I have pulled the changes from [joss-paper](https://github.com/m-shahbaz-kharal/LiGuard-2.x/tree/joss-paper) branch into the [dev-2.x](https://github.com/m-shahbaz-kharal/LiGuard-2.x/tree/dev_2.x) branch which is the default branch for current version. @hugoledoux can you please point the `editorialbot` to the updated `paper.md`.

The **updated paper acknowledges** existing libraries/frameworks like PDAL, pointcloudset, ROS, etc. It **reframes** `LiGuard` _not_ as a tool aiming to replace robust, scalable libraries like PDAL or pointcloudset, nor as a generalized end-to-end robotics framework like ROS. Instead, LiGuard is presented as a lightweight, GUI-based environment specifically designed for rapid prototyping and interactive experimentation with lidar and camera data processing pipelines.

## We try to answer the comments from your previous [review ](https://github.com/m-shahbaz-kharal/LiGuard-JOSS/issues/1) in the following:
> Why would researchers focus on feature-complete software development? I can understand that some prefer a GUI based workflow, but researchers are indeed focused on "researching novelties" which can be done with existing packages, tools and workflows. Maybe if the authors frame it differently it would find its audience, but I doubt that it will get traction in their indented audience of research in the field of "robotics, autonomous driving, traffic monitoring, infrastructure inspection, and many areas of environmental surveying and mapping".

The paper reframes `LiGuard` and emphasizes that its core strength lies in enabling live and interactive modification of pipeline steps and parameters directly through a GUI. This functionality is often manually implemented in notebook- or script-based developments.

> The authors also fail to mention PDAL which also has a pipeline functionality based on JSON configurations https://pdal.io/en/2.4.3/pipeline.html. What does LiGuard offer compared to PDAL?

PDAL and other related libraries/frameworks are acknowledged, and quick arguments are provided that place `LiGuard` as a useful tool for specific requirements rather than a replacement to the established libraries/frameworks.

> The sentence that the pointcloudset is a Python package that "...can be imported to custom, ofter hard-coded scripts thus imposing limitations on reusability and scalability" is wrong ... ... ... Specificly pointloudset uses dask and therefore can scale to huge clusters, wheras a GUI programm such as LiGuard can not.

We acknowledge that our previous description was poorly worded, we’ve removed such phrasings entirely. Moreover, in the paper we clearly mention that `LiGuard` is not optimized for large-scale datasets and is more focused for interactive fine-tuning and testing of processing pipeline components.

> Also the paper needs to cite the Robot Operating System as it is closely related to their main usecase. Both ROS and open3D offer also GUIs, which is not mentioned. A section on how to solve the same problem in ROS is needed, and how their approach offers some advantage.

While ROS is now cited in the updated version of the paper, we acknowledge its importance and relevance to the broader lidar processing ecosystem. However, providing a detailed section on how to solve the same problem using ROS would require an in-depth comparison and technical walkthrough that is beyond the scope of this submission.

> Lidar files are only supporded in the format of .bin and .npy which is for individual point clouds. ROS bag files are not mentioned and seems like they are not supported. Therefore, LiGuard only supports processing of individual point cloud files in bulk. How for example, would you find out the max value of x in all your point clouds?

`LiGuard` uses `Open3D` for I/O and should support most of lidar file types such as `pcd` and `ply` files. The `rosbag` file support is prospective. `LiGuard` provide utility functions such as [`gather_point_cloud`](https://m-shahbaz-kharal.github.io/LiGuard-2.x/algo_auxiliary.html#:~:text=utility%20functions%20are%3A-,gather_point_clouds,-%3A%20Gathers%20current_point_cloud_numpy%20in) to process multiple frames for aggregation or similar purposes.

## We try to give feedback to new comments in the following:
### Main drawbacks of the current approach:
> For each dataset, users must generate a folder structure or at least a YAML config file. There is no way to define a single pipeline and simply point it at multiple datasets.

A folder structure is intentional to share the pipeline more easily. The pipeline folder keeps the configuration, custom calibration and label _data_handlers_, and custom processing _algo_s intact. To change a pipeline to point to a different dataset, users can change _data_ section parameters to quickly test the same pipeline on other datasets.

> The function templates are somewhat convoluted. Some appear redundant—why not unify or simplify them?

We acknowledge that the templates contain a lot of comments making them convoluted but our intention behind such details is to allow new users to understand the structure. The redundant import was actually a typo, thanks for pointing out, I fixed that in a quick [commit](https://github.com/m-shahbaz-kharal/LiGuard-2.x/commit/9c2cd4bce6cbf19fb40a0f51219aaa12cc486f60).

> Custom postprocessing steps generate a YAML file rather than a Python script, which seems counterintuitive for users who expect to write and reuse code.

By using the button `Create Custom ... Function`, both a YAML (for exposing the parameters to GUI) and a Python script is created. These are stored in pipeline folder, users can share the pipeline folder to share their code to be reused.

> There is no support for reusing algorithms from existing ROS packages or for integrating with community-maintained lidar processing libraries.

We understand that such integration can add huge advantages and while ROS related files support, such as `rosbag`, might be added later, `LiGuard` is not intended as an extension to ROS. Therefore, it is unlikely that ROS packages will be supported. We argue that ROS is much bigger framework (a meta operating system rather than a pointcloud processing library) with a steep learning curve and should not be compared to this submission.

> while liguard-cmd appears to use some form of multithreading, this is not mentioned or explained in the paper or documentation. If multithreading is a distinguishing feature, it should be clearly stated.

The `liguard-cmd` utility uses configuration files created using the main `liguard-gui` to do the same operations but in bulk-processing mode. It uses multithreading to handle I/O separately. It also uses queues between each major processing stage including pre-processing, algorithms, and post-processing. Moreover, it removes the GUI and rendering completely to gain some processing speed improvements. However, please note that since the nature of pipelines is sequential, the performance gains might not be significant compared to using `liguard-gui` with visualization turned off.

## Functionality
> Once I click "Apply" in the GUI, I cannot undo or modify the step without restarting the program. This makes experimentation cumbersome.

Is it something like CTRL+Z that you are referring to? It should be possible to change parameters after clicking `Apply`. If something is not working as it should, please raise an issue.

>Running liguard-cmd with the example pipeline initially showed progress (100%) on all steps but then failed to exit cleanly and produced no output data.

This is possible a bug, can you please raise an issue to reproduce the problem.

> New Specific Comments
Capitalize Python consistently.
Clearly explain whether and how liguard-cmd uses multithreading, both in the paper and in the documentation.

"Python" is now capitalized consistently across paper and documentation.
`liguard-cmd` uses the same architecture as `liguard-gui`, but has multithreading at I/O, data queuing between major processing stages, and no GUI overhead. Since it is a small helper utility, rather than an entirely different module, it is stated (in the updated paper) that it processes the data slightly faster while also noting that the performance gain is affected by the slowest step in the pipeline.
