### 2025-05-08 m-shahbaz-kharal/LiGuard-2.x#8 @tgoelles
Add uvx to the documentation
You could add this to the documentation. This is a quick way to run the tool without and needs only uv to be installed.

uvx --from LiGuard liguard-gui
