### 2025-05-08 m-shahbaz-kharal/LiGuard-2.x#9 @tgoelles
[BUG] liguard-cmd does not finish running the example pipeline
**Describe the bug**
Running liguard-cmd stops without out finishing and control-c does not kill it.

**To Reproduce**
Steps to reproduce the behavior:
1.) In the terminal I run: uvx --from LiGuard liguard-cmd /Users/work/Documents/code/LiGuard-2.x/liguard/examples/simple_pipeline

2.) wait

3.) it stops after reading the data

**Expected behavior**
I expect it to run the whole pipeline and say Done or simething like that

**Log(s)**
`LiGuard` generates `logs` folder in the project folder. Please post the relevant log file. Also, add the terminal output if an error is printed. Ignore warnings.

uvx --from LiGuard liguard-cmd /Users/work/Documents/code/LiGuard-2.x/liguard/examples/simple_pipeline

    #########################################################
        _      _  _____                     _   ___    ___
        | |    (_)/ ____|                   | | |__ \  / _ \
        | |     _| |  __ _   _  __ _ _ __ __| |    ) || | | |
        | |    | | | |_ | | | |/ _` | '__/ _` |   / / | | | |
        | |____| | |__| | |_| | (_| | | | (_| |  / /_ | |_| |
        |______|_|\_____|\__,_|\__,_|_|  \__,_| |____(_)___/
                                       Headless Bulk Processor
    ##########################################################
    LiGuard's utility for no-GUI bulk data processing.

Logging level is too low. Setting to WARNING to prevent spam.
Reading data:   0%|                                       | 0/5 [00:00<?, ?it/s]Traceback (most recent call last):
  File "/Users/work/.cache/uv/archive-v0/mGtzvym5I38_btQCZjkBU/bin/liguard-cmd", line 12, in <module>
    sys.exit(main())
             ^^^^^^
  File "/Users/work/.cache/uv/archive-v0/mGtzvym5I38_btQCZjkBU/lib/python3.11/site-packages/liguard/liguard_cmd.py", line 329, in main
    bulk_process(args)
  File "/Users/work/.cache/uv/archive-v0/mGtzvym5I38_btQCZjkBU/lib/python3.11/site-packages/liguard/liguard_cmd.py", line 174, in bulk_process
    if not cfg['proc']['pre'][proc]['enabled']: continue
           ~~~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^
TypeError: 'bool' object is not subscriptable
Reading data: 100%|███████████████████████████████| 5/5 [00:20<00:00, 34.28it/s]

[log_20250508-073454.txt](https://github.com/user-attachments/files/20097434/log_20250508-073454.txt)

**Environment:**
 - OS: MAC 15.3.2
 - LiGuard Version 2.1.5.post1
 - Python Version 3.11.11
