### 2025-05-08 m-shahbaz-kharal/LiGuard-2.x#11 @tgoelles
Rethink the "priority"
Is priority the same as order? Should it be renamed to order?

Also the commend  "lower is higher" is ackward. Why not just make high is higher? Or if priority is in fact the order then I would expect it to be like 0,1,2,3 .....
