### 2025-05-08 #6751 @tgoelles
Ok, I understand I was still referring to the old version.

Once I have the new pdf version i will then adapt the review.

As for the functionality: Now the GUI works as expected and i can change settings. This and the new framing of the tool makes it more useful.

See also the new issues.  The Bug fix is important the other issues are just suggestions.

I thought a bit how I would use the tool, as I guess me and my colleagues are in the target user group. We work with jupyter notebooks for prototyping and interactive testing of parts of processing pipelines. This has the advantage that we can write code and test it out and then apply all the steps. Once this is done we make a package out of the processing steps and write tests. This are then reusable processing steps. I guess we will stick to this workflow. Nevertheless, it sure can be usefull for the community and it is great that the authors make it available. I would sure use it if it could read ROS files and it had many available processing steps and the templates are a bit more streamlined and easier to use.
