### 2025-05-09 #6751 @m-shahbaz-kharal
Thanks @tgoelles for the detailed review, I'll go over all the raised issues and fix those.

@hugoledoux, soft reminder that please set [LiGuard-2.x](https://github.com/m-shahbaz-kharal/LiGuard-2.x/tree/dev_2.x) as the repository for the submission.
