### 2025-05-12 1c88ade
chore: migrate setup configuration to pyproject.toml and remove setup.py
