+++ setup.py
@@ -1,60 +1,0 @@
-from setuptools import setup, find_packages
-import glob
-
-setup(
-    name="LiGuard",
-    version="2.1.5.post1",
-    author="Muhammad Shahbaz",
-    author_email="m.shahbaz.kharal@outlook.com",
-    description=(
-        "A research-purposed, GUI-powered, Python-based framework that allows easy "
-        "development of dynamic point-cloud (and accompanying image) data processing pipelines."
-    ),
-    long_description=open("README.md").read(),
-    long_description_content_type="text/markdown",
-    url="https://github.com/m-shahbaz-kharal/LiGuard-2.x",
-    classifiers=[
-        "Programming Language :: Python :: 3.10",
-        "Programming Language :: Python :: 3.11",
-        "Programming Language :: Python :: 3.12",
-        "Development Status :: 3 - Alpha",
-        "Environment :: Win32 (MS Windows)",
-        "Environment :: X11 Applications",
-        "Environment :: MacOS X",
-        "Intended Audience :: Developers",
-        "Intended Audience :: Education",
-        "Intended Audience :: Other Audience",
-        "Intended Audience :: Science/Research",
-        "License :: OSI Approved :: MIT License",
-        "Natural Language :: English",
-        "Operating System :: Microsoft :: Windows",
-        "Operating System :: POSIX :: Linux",
-        "Operating System :: MacOS :: MacOS X",
-        "Topic :: Scientific/Engineering :: Image Processing",
-        "Topic :: Scientific/Engineering :: Visualization",
-        "Topic :: Multimedia :: Graphics :: Viewers",
-        "Topic :: Utilities",
-    ],
-    keywords="point-cloud, image, processing, pipeline, dynamic, GUI, research, development, framework",
-    python_requires=">=3.10",
-    install_requires=[
-        "matplotlib",
-        "numpy<2.0.0",
-        "open3d==0.19.0",
-        "opencv-python",
-        "pynput",
-        "pillow",
-        "pyyaml",
-        "scipy",
-        "tqdm",
-    ],
-    packages=find_packages(),
-    include_package_data=True,
-    entry_points={
-        "console_scripts": [
-            "liguard-gui=liguard.liguard_gui:main",
-            "liguard-cmd=liguard.liguard_cmd:main",
-            "liguard-profiler=liguard.liguard_profiler:main",
-        ]
-    },
-)
