### 2025-05-12 m-shahbaz-kharal/LiGuard-2.x#6 @m-shahbaz-kharal
Thanks for the suggestion, `pyproject.toml` has been updated in commit https://github.com/m-shahbaz-kharal/LiGuard-2.x/commit/1c88ade4628bc5416cb6643c3a7f8bef00f60632.
