### 2025-05-12 6374cb4
docs: add alternative quick launch and developmental build instructions to README.
