### 2025-05-12 m-shahbaz-kharal/LiGuard-2.x#8 @m-shahbaz-kharal
Thanks for the suggestions, this quick option to run `LiGuard` is added to documentation in https://github.com/m-shahbaz-kharal/LiGuard-2.x/commit/6374cb41505e9eb6170790bdeabd03a188f1113e.
