### 2025-05-13 3850f21
feat: refactored bulk processing to use Dask for parallel execution and improved performance.
