+++ liguard/liguard_cmd.py
@@ -1,7 +1,12 @@
 import os
 import sys
+import time
+
 from liguard.gui.gui_utils import resolve_for_application_root, resolve_for_default_workspace
+
+import argparse
 import yaml
+from tqdm import tqdm
 
 from liguard.pcd.file_io import FileIO as PCD_File_IO
 from liguard.img.file_io import FileIO as IMG_File_IO
@@ -10,68 +15,19 @@ from liguard.lbl.file_io import FileIO as LBL_File_IO
 
 from liguard.gui.logger_gui import Logger
 
-import time
-import signal
-from threading import Thread, Event
-from queue import Queue
+import dask
+from dask import delayed, compute
+from dask.diagnostics import ProgressBar
 
-from tqdm import tqdm
-
-stop_event = Event()
-
-def reader2queue(reader, size, target_queue):
-    for idx in range(size):
-        if stop_event.is_set():
-            target_queue.put(None)
-            break
-        target_queue.put(reader[idx])
-
-def queue2dict2queue(source_queue, key1, key2, target_queue):
-    while True:
-        data = source_queue.get()
-        if data is None:
-            target_queue.put(None)
-            source_queue.task_done()
-            break
-        data_dict = {key1: data[0], key2: data[1]}
-        target_queue.put(data_dict)
-        source_queue.task_done()
-    
-def dicts2singledict(source_queues, target_queue, p_bar):
-    while True:
-        data = [source_queues[i].get() for i in range(len(source_queues))]
-        if None in data:
-            target_queue.put(None)
-            for i in range(len(source_queues)): source_queues[i].task_done()
-            break
-        data_dict = dict()
-        for d in data: data_dict.update(d)
-        target_queue.put(data_dict)
-        for i in range(len(source_queues)): source_queues[i].task_done()
-        p_bar.update(1)
-
-def dict2proc2dict(source_queue, cfg, logger, processes, target_queue, p_bar):
-    while True:
-        data = source_queue.get()
-        if data is None:
-            if target_queue: target_queue.put(None)
-            source_queue.task_done()
-            break
-        for process in processes: process(data, cfg, logger)
-        if target_queue: target_queue.put(data)
-        source_queue.task_done()
-        p_bar.update(1)
-
-def signal_handler(sig, frame):
-    print("\nCtrl + C detected! Stopping ...")
-    stop_event.set()  # Signal all threads to stop
 
 def bulk_process(args):
+    # load config
     pipeline_dir = args.pipeline_dir
     base_cfg_path = os.path.join(pipeline_dir, 'base_config.yml')
     with open(base_cfg_path) as f:cfg = yaml.safe_load(f)
     cfg['data']['pipeline_dir'] = pipeline_dir
-    
+
+    # load custom algorithms
     custom_algo_dir = os.path.join(pipeline_dir, 'algo')
     custom_algos_cfg = dict()
     if os.path.exists(custom_algo_dir):
@@ -86,88 +42,73 @@ def bulk_process(args):
                 with open(os.path.join(algo_type_path, algo_file_name)) as f: cust_algo_cfg = yaml.safe_load(f)
                 custom_algos_cfg[algo_type].update(cust_algo_cfg)
             cfg['proc'][algo_type].update(custom_algos_cfg[algo_type])
-    
+
+    # init logger
     logger = Logger()
     if cfg['logging']['level'] < Logger.WARNING:
         print('Logging level is too low. Setting to WARNING to prevent spam.')
         cfg['logging']['level'] = Logger.WARNING
     logger.reset(cfg)
 
-    # create dirs
+    # create output directory
[…]
