### 2025-05-13 83dba8f
chore: add 'dask' to project dependencies.
