### 2025-05-13 m-shahbaz-kharal/LiGuard-2.x#9 @m-shahbaz-kharal
@tgoelles thanks for reporting the bug, it has been fixed in this https://github.com/m-shahbaz-kharal/LiGuard-2.x/commit/3850f21ef1d7a99305224a9dd987197c2378ee9f commit. Additionally, [dask](https://github.com/dask/dask) is implemented for multithreaded operation in `liguard_cmd`. This update will be documented in the next release.

Please note that unless you set`--save_raw_results` flag, `liguard_cmd` does not output anything by default, unless of course a pipeline function such as `proc->post->create_per_object_pcdet_dataset` has been enabled.
