### 2025-05-13 m-shahbaz-kharal/LiGuard-2.x#11 @m-shahbaz-kharal
@tgoelles Thanks for the suggestion! This has been addressed in commit [87608f3](https://github.com/m-shahbaz-kharal/LiGuard-2.x/commit/87608f3e83dac0583b634e988aa14fd56946c03d).

Please update your `base_config.yml` by replacing all instances of `priority` with `order`.
This change is also documented in the [v2.1.5.post2](https://github.com/m-shahbaz-kharal/LiGuard-2.x/releases/tag/v2.1.5.post2) release notes.
