### 2025-05-13 #6751 @m-shahbaz-kharal
@tgoelles, I have responded to all of the issues you raised in:
[issue #6: use pyproject.toml better](https://github.com/m-shahbaz-kharal/LiGuard-2.x/issues/6)
[issue #8: Add uvx to the documentation](https://github.com/m-shahbaz-kharal/LiGuard-2.x/issues/8)
[issue #9: [BUG] liguard-cmd does not finish running the example pipeline](https://github.com/m-shahbaz-kharal/LiGuard-2.x/issues/9).
[issue #11: Rethink the "priority"](https://github.com/m-shahbaz-kharal/LiGuard-2.x/issues/11).

Thanks.
