### 2025-05-15 9fa186b
refactor: replace __import__ with our custom import_module utility function for dynamic imports.
