+++ liguard/liguard_gui.py
@@ -5,7 +5,7 @@ import traceback
 import open3d.visualization.gui as gui
 
 from liguard.gui.config_gui import BaseConfiguration as BaseConfigurationGUI
-from liguard.gui.gui_utils import resolve_for_application_root, resolve_for_default_workspace
+from liguard.gui.gui_utils import resolve_for_application_root, resolve_for_default_workspace, import_module
 from liguard.gui.logger_gui import Logger
 from liguard.liguard_profiler import Profiler
 
@@ -328,14 +328,14 @@ class LiGuard:
         ##########################################
         # pre processes
         self.pre_processes = dict()
-        built_in_pre_modules = __import__('liguard.algo.pre', fromlist=['*']).__dict__
+        built_in_pre_modules = import_module('liguard.algo.pre', fromlist=['*']).__dict__
         for proc in cfg['proc']['pre']:
             enabled = cfg['proc']['pre'][proc]['enabled']
             if enabled:
                 try:
                     order = cfg['proc']['pre'][proc]['order']
                     if proc in built_in_pre_modules: process = built_in_pre_modules[proc]
-                    else: process = __import__(proc, fromlist=['*']).__dict__[proc]
+                    else: process = import_module(proc, fromlist=['*']).__dict__[proc]
                     self.pre_processes[order] = process
                 except Exception:
                     self.logger.log(f'pre_processes creation failed for {proc}:\n{traceback.format_exc()}', Logger.CRITICAL)
@@ -344,14 +344,14 @@ class LiGuard:
         
         # lidar processes
         self.lidar_processes = dict()
-        built_in_lidar_modules = __import__('liguard.algo.lidar', fromlist=['*']).__dict__
+        built_in_lidar_modules = import_module('liguard.algo.lidar', fromlist=['*']).__dict__
         for proc in cfg['proc']['lidar']:
             enabled = cfg['proc']['lidar'][proc]['enabled']
             if enabled:
                 try:
                     order = cfg['proc']['lidar'][proc]['order']
                     if proc in built_in_lidar_modules: process = built_in_lidar_modules[proc]
-                    else: process = __import__(proc, fromlist=['*']).__dict__[proc]
+                    else: process = import_module(proc, fromlist=['*']).__dict__[proc]
                     self.lidar_processes[order] = process
                 except Exception:
                     self.logger.log(f'lidar_processes creation failed for {proc}:\n{traceback.format_exc()}', Logger.CRITICAL)
@@ -360,14 +360,14 @@ class LiGuard:
         
         # camera processes
         self.camera_processes = dict()
-        built_in_camera_modules = __import__('liguard.algo.camera', fromlist=['*']).__dict__
+        built_in_camera_modules = import_module('liguard.algo.camera', fromlist=['*']).__dict__
         for proc in cfg['proc']['camera']:
             enabled = cfg['proc']['camera'][proc]['enabled']
             if enabled:
                 try:
                     order = cfg['proc']['camera'][proc]['order']
                     if proc in built_in_camera_modules: process = built_in_camera_modules[proc]
-                    else: process = __import__(proc, fromlist=['*']).__dict__[proc]
+                    else: process = import_module(proc, fromlist=['*']).__dict__[proc]
                     self.camera_processes[order] = process
                 except Exception:
                     self.logger.log(f'camera_processes creation failed for {proc}:\n{traceback.format_exc()}', Logger.CRITICAL)
@@ -376,14 +376,14 @@ class LiGuard:
 
         # calib processes
         self.calib_processes = dict()
-        built_in_calib_modules = __import__('liguard.algo.calib', fromlist=['*']).__dict__
+        built_in_calib_modules = import_module('liguard.algo.calib', fromlist=['*']).__dict__
         for proc in cfg['proc']['calib']:
[…]
