### 2025-06-02 7126ca0
updated official citation for ROS, added citation for dask.
