+++ paper/paper.bib
@@ -173,14 +173,6 @@ title      = {Array programming with {NumPy}},
     year = "2022"
 }
 
-@software{ros,
-  author = {{Stanford Artificial Intelligence Laboratory et al.}},
-  title = {Robotic Operating System},
-  url = {https://www.ros.org},
-  version = {ROS Melodic Morenia},
-  date = {2018-05-23},
-}
-
 @misc{pdal_contributors_2024_2616780,
   author = {PDAL Contributors},
   title = {PDAL Point Data Abstraction Library},
@@ -192,3 +184,21 @@ title      = {Array programming with {NumPy}},
   }
 }
 
+@Manual{dask,
+  title = {Dask: Library for dynamic task scheduling},
+  author = {{Dask Development Team}},
+  year = {2016},
+  url = {http://dask.pydata.org},
+}
+
+@article{doi:10.1126/scirobotics.abm6074,
+    author = {Steven Macenski  and Tully Foote  and Brian Gerkey  and Chris Lalancette  and William Woodall },
+    title = {Robot Operating System 2: Design, architecture, and uses in the wild},
+    journal = {Science Robotics},
+    volume = {7},
+    number = {66},
+    pages = {eabm6074},
+    year = {2022},
+    doi = {10.1126/scirobotics.abm6074},
+    URL = {https://www.science.org/doi/abs/10.1126/scirobotics.abm6074}
+}
