### 2025-06-02 49deba9
the word lidar is consistent, fixed typos, and improved sentence structures where braces were overused.
