+++ paper/paper.md
@@ -25,18 +25,18 @@ bibliography: paper.bib
 
 ---
 # Summary
-There is a growing interest in the development of lidar-based applications in domains like robotics, autonomous driving, traffic monitoring, infrastructure inspection, and many areas of environmental surveying and mapping. In many cases, these applications rely on solutions that are not end-to-end but rather a sequential combination (a pipeline) of individual data processing functions including data readers, pre-processors, application-specific algorithms or models, post-processors, visualizers, etc. Creating such multi-step solutions often requires extensive revisions of the proposed processing pipeline, including enabling or disabling certain functions, fine-tuning parameters, and even completely replacing parts of the solution. Such revisions pose challenges on the speed of research and even on the reuse of it, due to interdependencies among functions. To address this issue, `LiGuard` is presented. `LiGuard` is a software framework that provides an easy-to-use graphical user interface (GUI) that helps researchers rapidly build and test their point cloud (and corresponding image) processing pipelines. It allows them to dynamically add/remove/reorder functions, adjust the parameters of those functions, and visualize results in a live, interactive manner compared to other packages where such modifications often require re-executing the program script. Moreover, because it creates all the meta files in structured directories, it allows easy sharing of the created pipelines or the individual functions used in those pipelines.
+There is a growing interest in the development of lidar-based applications in domains like robotics, autonomous driving, traffic monitoring, infrastructure inspection, and many areas of environmental surveying and mapping. In many cases, these applications rely on solutions that are not end-to-end but rather a sequential combination (a pipeline) of individual data processing functions such as data readers, pre-processors, application-specific algorithms or models, post-processors, visualizers, etc. Creating such multi-step solutions often requires extensive revisions of the proposed processing pipeline, including enabling or disabling certain functions, fine-tuning parameters, and even completely replacing parts of the solution. Such revisions pose challenges on the speed of research and even on the reuse of it, due to interdependencies among functions. To address this issue, `LiGuard` is presented. It is a software framework that provides an easy-to-use graphical user interface (GUI) that helps researchers rapidly build and test their point cloud (and corresponding image) processing pipelines. It allows them to dynamically add/remove/reorder functions, adjust the parameters of those functions, and visualize results in a live, interactive manner compared to other packages where such modifications often require re-executing the program script. Moreover, because it creates all the meta files in structured directories, it allows easy sharing of the created pipelines or the individual functions used in those pipelines.
 
 `LiGuard` features, out of the box, data reading for many common dataset formats including support for reading calibration and label data. Moreover, it provides an increasing list of commonly used algorithm components ranging from basic data preprocessors to advanced object detection and tracking algorithms. Additionally, it establishes a straightforward standard for adding custom functions/algorithms, allowing users to integrate unique components into their pipelines. For the latest updates on supported features, please see [LiGuard’s documentation](https://m-shahbaz-kharal.github.io/LiGuard-2.x/).
 
[…]
