### 2025-06-02 #6751 @m-shahbaz-kharal
Dear @tgoelles,

Thanks for noting the issues. I have updated the paper to fix all the typos and have further made small changes to make some sentences clearer, hence, avoiding overuse of braces. I also added the official ROS citation that includes doi. I am attaching the PDF for convenience [here](https://github.com/m-shahbaz-kharal/LiGuard-2.x/actions/runs/15398185864/artifacts/3243151807).

Best regards,
Muhammad Shahbaz
