### 2025-06-03 #6751 @tgoelles
Dear @m-shahbaz-kharal,

Thanks for the quick corrections. There are still some DOI missing but the bot will check that anyway.
I now marked everything as complete in the checklist. All the best for finishing your PhD.

BR,
Thomas Gölles
