### 2025-06-03 #6751 @m-shahbaz-kharal
Dear @hugoledoux,

I hope you are doing well.

Could you kindly point the bot to the following repository and proceed?
https://github.com/m-shahbaz-kharal/LiGuard-2.x

Additionally, could you please let me know the next steps in the process? Kindly inform me if there is anything further required from our side.

Thank you very much for your time and support.

Best regards,
Muhammad Shahbaz
