### 2025-06-20 a922b4f
A few fixes while proof-reading the paper
