+++ paper/paper.bib
@@ -5,7 +5,7 @@
   publisher = {The Open Journal},
   volume = {6}, number = {65}, pages = {3471},
   author = {Thomas Goelles and Birgit Schlager and Stefan Muckenhuber and Sarah Haas and Tobias Hammer},
-  title = {`pointcloudset`: Efficient Analysis of Large Datasets of Point Clouds Recorded Over Time},
+  title = {{pointcloudset}: Efficient Analysis of Large Datasets of Point Clouds Recorded Over Time},
   journal = {Journal of Open Source Software}
 }
 
@@ -16,7 +16,7 @@
   publisher = {The Open Journal},
   volume = {4}, number = {36}, pages = {990},
   author = {Sebastian Lamprecht},
-  title = {Pyoints: A Python package for point cloud, voxel and raster processing.},
+  title = {Pyoints: A {Python} package for point cloud, voxel and raster processing.},
   journal = {Journal of Open Source Software}
 }
 
@@ -27,6 +27,7 @@
   month     = {May 9-13},
   year      = {2011},
   address   = {Shanghai, China},
+  doi       = {10.1109/icra.2011.5980567},
   publisher = {IEEE}
 }
 
