+++ paper/paper.bib
@@ -20,15 +20,16 @@
   journal = {Journal of Open Source Software}
 }
 
-@InProceedings{Rusu_ICRA2011_PCL,
-  author    = {Radu Bogdan Rusu and Steve Cousins},
-  title     = {{3D is here: Point Cloud Library (PCL)}},
-  booktitle = {{IEEE International Conference on Robotics and Automation (ICRA)}},
-  month     = {May 9-13},
-  year      = {2011},
-  address   = {Shanghai, China},
-  publisher = {IEEE}
-}
+@INPROCEEDINGS{Rusu_ICRA2011_PCL,
+  author={Rusu, Radu Bogdan and Cousins, Steve},
+  booktitle={2011 IEEE International Conference on Robotics and Automation}, 
+  title={3D is here: Point Cloud Library (PCL)}, 
+  year={2011},
+  volume={},
+  number={},
+  pages={1-4},
+  keywords={Point cloud compression;Three-dimensional displays;Libraries;Codes;Cloud computing;Data visualization;Robots},
+  doi={10.1109/ICRA.2011.5980567}}
 
 @article{Zhou2018,
     author    = {Qian-Yi Zhou and Jaesik Park and Vladlen Koltun},
