### 2025-06-20 5bf8880
a few more fixes
