+++ paper/paper.bib
@@ -78,17 +78,6 @@
     year = {2000}
 }
 
-@article{geiger2013vision,
-  title={Vision meets robotics: The kitti dataset},
-  author={Geiger, Andreas and Lenz, Philip and Stiller, Christoph and Urtasun, Raquel},
-  journal={The International Journal of Robotics Research},
-  volume={32},
-  number={11},
-  pages={1231--1237},
-  year={2013},
-  publisher={Sage Publications Sage UK: London, England}
-}
-
 @misc{openpcdet2020,
     title={OpenPCDet: An Open-source Toolbox for 3D Object Detection from Point Clouds},
     author={OpenPCDet Development Team},
