### 2025-06-20 #6751 @m-shahbaz-kharal
Thanks @hugoledoux,

- I pulled the PR you made
- Release tag: [v2.1.5.post3](https://github.com/m-shahbaz-kharal/LiGuard-2.x/releases/tag/v2.1.5.post3).
- Zenodo DOI: https://doi.org/10.5281/zenodo.15705277
