### 2025-06-23 #6751 @tgoelles
@crvernon. Please don't. The authors had LiDAR before and I said they should write lidar instead. The lidar community wants this, which is then also more consistent with sonar and radar which are similar. see https://lidarmag.com/wp-content/uploads/PDF/LiDARNewsMagazine_DeeringStoker-CasingOfLiDAR_Vol4No6.pdf
