### 2025-06-23 #6751 @hugoledoux
> [@editorialbot](https://github.com/editorialbot) accept
>
> Interesting [@tgoelles](https://github.com/tgoelles) and thank you for the reference. I will accept this as justification and move this on to acceptance.

I personally changed the way I write "lidar" since @tgoelles gave this as comments, and as a teacher, I now send this document to students!
