### 2025-06-23 #6751 @m-shahbaz-kharal
Thank you so much! we are really honored to have our work published with JOSS. Huge thanks to @hugoledoux for the entire editing process and to @chenzhaiyu and @tgoelles for the time and all the insightful reviews; your contributions have greatly improved LiGuard. I'm truly grateful 🙏.
